### 2020-06-18 #2354 @whedon
[PRE REVIEW]: DearScholar: A mobile application to conduct qualitative and quantitative diary research
**Submitting author:** @pmkruyen (<a href="http://orcid.org/0000-0003-0109-1744">Peter Kruyen</a>)
**Repository:** <a href="https://github.com/pmkruyen/dearscholar" target ="_blank">https://github.com/pmkruyen/dearscholar</a>
**Version:** 1.48
