### 2020-06-18 #2354 @arfon
> Thanks @arfon for your message. Completely understand. Your bot Whedon is asking me to suggest reviews in the first message above, should I do this or wait till further notice?

Feel free to suggest some reviewers now (please avoid mentioning them directly to avoid unnecessary GitHub notifications for folks) which will help when an editor is assigned.
