### 2020-06-24 #2354 @pmkruyen
@arfon @kthyng thanks for checking... I have build DearScholar using Framework7; dearscholar.js--the heart of the project and a lot of work--; plus all the html-files and dearscholar.css are mine (code solely written by me); also see the table at the bottom of the README file. Best Peter
