### 2020-06-24 #2354 @pmkruyen
@arfon and @kthyng To follow up, I just realized, I included the complete codebase of Framework7 in my repro as matter of convenience, but (obviously) not everything is needed for DearScholar to work.
