### 2020-06-30 #2354 @cMadan
@pmkruyen, I've looked over your submission and have three initial comments:

1. Your license has some additional text at the top, in addition to a standard OSI license. Publishing in JOSS requires that you *only* use a standard OSI-compliant license with no modifications.

1. I would appreciate if you could add a few screenshots (just one picture concatenating a screenshots from a few different parts of the app). Since it's very front-end oriented in that research participants would be using it, it would be helpful to see how it looks 'out of the box', though I do understand that it's just a bit of HTML code for another researcher to change this.

1. It would be helpful to provide the MySQL database initialisation code for setting up the databases for the first time. I have a feeling that with 10 tables, someone could easily make a typo, or set a column to the wrong data format, so it would be great to provide the syntax for automating that step.
