+++ readme.md
@@ -75,7 +75,7 @@ DearScholar has the capacity to send short in-app messages to specific responden
 DearScholar has been built using [Apache Cordova / Phonegap](https://phonegap.com/), and [Framework7 (v5.4.1)] (https://framework7.io) by Vladimir Kharlampidi and his team. Next to several general Cordova plugins, DearScholar uses [fingerprint-aio](https://github.com/NiklasMerz/cordova-plugin-fingerprint-aio) to enable Touch ID and Face Recognition, [sqlite-evcore-extbuild-free](https://www.npmjs.com/package/cordova-sqlite-evcore-extbuild-free) to use sqlite tables, and [cordova-plugin-ionic-webview](https://github.com/ionic-team/cordova-plugin-ionic-webview) to replace UIWebView with WKWebView. 
 
 ## Configuration
-To configure Dearscholar, ten tables need to be set up in a secured database (such as mysql) on a secured server. The tables include variables (in the columns) and specific settings or respondents' data (in the rows). In this section, these ten tables are described. 
+To configure Dearscholar, ten tables need to be set up in a secured database (such as mysql) on a secured server and the proper user rights needs to be given to individual researchers (administrators) and respondents (users of DearScholar). The tables include variables (in the columns) and specific settings or respondents' data (in the rows). The MySQL database initialisation code for the database, ten tables and user rights can be found [here](https://github.com/pmkruyen/dearscholar/blob/master/mysql). In this section, these ten tables are described.
 
 ### 1) authentication table
 A table with the following columns, settings are stored for a single respondent on each row:
