+++ readme.md
@@ -52,6 +52,8 @@ In this second part of this text, DearScholar's current features and technical d
 2) When opening DearScholar for the first time, a respondent is asked to: * allow push notifications; * fill out the username and password that (s)he received from the researcher; * choose a 4-digit PIN if the device does not support Touch ID or Face Recognition; and * agree to the informed consent form.
 3) When everthing goes well, DearScholar pulls the required survey tables and settings from the server, and the respondent is directed to the measurement schedule (homepage).
 
+![alt text](https://github.com/pmkruyen/dearscholar/blob/master/screenshots/4.png)
+
 ### Let respondents answer questions
 When logging in to DearScholar, a respondent is directed to the the measurement schedule (homepage) with all measurement occasions (dates). Future measurement occasions—measurement occasions beyond the current date—are locked and marked with a 'closed lock' icon. 
 
