+++ readme.md
@@ -53,7 +53,7 @@ In this second part of this text, DearScholar's current features and technical d
 3) When everthing goes well, DearScholar pulls the required survey tables and settings from the server, and the respondent is directed to the measurement schedule (homepage).
 
 <p align="center">
-  <img src=https://github.com/pmkruyen/dearscholar/blob/master/screenshots/4.png width="400"/>
+  <kbd><img src=https://github.com/pmkruyen/dearscholar/blob/master/screenshots/4.png width="350"/></kbd>
 </p>
 
 ### Let respondents answer questions
