### 2020-06-30 #2354 @pmkruyen
@cMadan thanks for being the editor of my submission and your initial comments:

1. I deleted the modifications. My university wants me to include those lines; I now included them in the README file (under copyright). Hope that this is acceptable. Or should these lines be included in the separate files?

2. Screenshots were included in the App Store and Google Play but not on Github. With some trial and error, I managed to include a few screenshots in the README file too.

3. Aha. I added a file containing the MySQL code for the database and tables, plus the user rights assignment settings (see the mysql text file).

Best,
Peter
