+++ iOS/platforms/browser/www/js/dearscholar.js
@@ -290,6 +290,68 @@ function onBackKeyDown() {
     }
 }
 
+
+// Style the login screen
+document.addEventListener("deviceready", function(){
+            $$('#showPasswordIcon').html('')
+            $$("#password").removeClass('passwordisempty');
+            document.getElementById('password').type = 'password';
+    
+            // Enable the show password button when the password field is empty
+            if ($('#password.input-with-value').length==0){
+                $$("#password").addClass('passwordisempty');
+                $$('#showPasswordIcon').html('eye')
+            }
+    
+            $$('#password').on('keyup', function (e) {
+                if(password.value.length==0)
+                {
+                    $$("#password").addClass('passwordisempty');
+                    $$('#showPasswordIcon').html('eye')
+                }
+            })
+    
+            $$('#showPasswordButton').on('click', function (e) {
+                var el = document.getElementsByClassName("passwordisempty");
+                
+                if (el.length==1){
+                    if(document.getElementById('password').type == 'password') {
+                        document.getElementById('password').type = 'text';
+                        $$('#showPasswordIcon').html('eye_slash')
+                    }      
+                    else {
+                        document.getElementById('password').type = 'password';
+                        $$('#showPasswordIcon').html('eye')
+                    }
+                }
+            })
+    
+            //Style the login page dependend on whether TouchID is (not) available    
+            Fingerprint.isAvailable(isAvailableSuccess, isAvailableError);
+
+            function isAvailableSuccess(result) {
+                document.getElementById("loginButton").innerHTML = "Log in met Touch ID of gezichtsherkenning";
+                //$("#fourthBlock" ).append("<id='TouchIDAvailableButUsePIN'><li><a href='#' class='item-link list-button' id='loginButton' >Log in met PIN</a></li>")
+                
+                //Add event handler to the login button
+                $$('#loginButton').on('click', function (e) {
+                    loginTOUCH()
+                })
+            }
+
+            function isAvailableError(error) {
+                document.getElementById("loginButton").innerHTML = "Log in met PIN"; 
+                
+                //Add event handler to the login button
+                $$('#loginButton').on('click', function (e) {
+                    loginPIN()
+                })
+            }
+            
+}, false);
+
+
+
 //Attach functions when the keyboard is shown and closed to hide and show the 'bottomButtons' in the modules, respectively.
 window.addEventListener("keyboardWillShow", function(e) {
     $(".bottomButtons").hide();
@@ -308,7 +370,7 @@ window.addEventListener("keyboardDidHide", function(e) {
 var timeoutID;
 
 function goInactive() {
-    app.views.main.router.back('/',{ignoreCache: true, reload: true});
+    app.views.main.router.back('/',{force: true, ignoreCache: true, reload: true});
     app.panel.close("left")
 }
 
@@ -487,6 +549,7 @@ function authentication(){
                             window.localStorage.setItem('project', settings[0].data.project);
                             window.localStorage.setItem('q0_startdate', settings[0].data.q0_startdate);
                             window.localStorage.setItem('q0_occasions', settings[0].data.q0_occasions);
+                            window.localStorage.setItem('q0_intervaltype', settings[0].data.q0_intervaltype);
                             window.localStorage.setItem('q0_interval', settings[0].data.q0_interval);
                      
                             // register the device to receive notifications
@@ -582,8 +645,7 @@ $$('#acceptConsent').on('click', function (e) {
                  {  
                     surveyStructure = JSON.parse(data).data0;
                     pageStructure = JSON.parse(data).data1;
-                    questionTable = JSON.parse(data).data2;
[…]
