+++ Android26/www/js/dearscholar.js
@@ -487,6 +487,7 @@ function authentication(){
                             window.localStorage.setItem('project', settings[0].data.project);
                             window.localStorage.setItem('q0_startdate', settings[0].data.q0_startdate);
                             window.localStorage.setItem('q0_occasions', settings[0].data.q0_occasions);
+                            window.localStorage.setItem('q0_intervaltype', settings[0].data.q0_intervaltype);
                             window.localStorage.setItem('q0_interval', settings[0].data.q0_interval);
                      
                             // register the device to receive notifications
@@ -582,8 +583,7 @@ $$('#acceptConsent').on('click', function (e) {
                  {  
                     surveyStructure = JSON.parse(data).data0;
                     pageStructure = JSON.parse(data).data1;
-                    questionTable = JSON.parse(data).data2;
-                                         
+                    questionTable = JSON.parse(data).data2;                                         
                     DiaryDatabase.transaction(function (tx){
                         tx.executeSql('DROP TABLE IF EXISTS surveyStructure');
                         tx.executeSql('DROP TABLE IF EXISTS pageStructure');
@@ -592,7 +592,7 @@ $$('#acceptConsent').on('click', function (e) {
                         tx.executeSql('CREATE TABLE IF NOT EXISTS surveyInformation (surveydate,A,B,C,D)');
                         tx.executeSql('CREATE TABLE IF NOT EXISTS surveyStructure (consent, manual, moduleAname, moduleBname, moduleCname, moduleDname)');
                         tx.executeSql('CREATE TABLE IF NOT EXISTS pageStructure (module1, classp, idp, header, backbuttontid, backbuttontext, backbuttononclick, backbuttonhref, nextbuttonid, nextbuttontext, nextbuttononclick,nextbuttonhref)');
-                        tx.executeSql('CREATE TABLE IF NOT EXISTS questiontable (module2, tab, idq, type, question, footer)');
+                        tx.executeSql('CREATE TABLE IF NOT EXISTS questiontable (module2, tab, idq, type, question, categories, footer)');
                         tx.executeSql('CREATE TABLE IF NOT EXISTS responseTable (id, surveydate, timestamp)');
                         
                         DiaryDatabase.transaction(function(tx) {
@@ -623,7 +623,8 @@ $$('#acceptConsent').on('click', function (e) {
                                                  questionTable[i].data.tab, 
                                                  questionTable[i].data.idq, 
                                                  questionTable[i].data.type, 
-                                                 questionTable[i].data.question, 
+                                                 questionTable[i].data.question,
+                                                 questionTable[i].data.categories,
                                                  questionTable[i].data.footer
                                                  );
                             listQuestionIds.push(questionTable[i].data.idq);
@@ -685,27 +686,35 @@ $$('#acceptConsent').on('click', function (e) {
     }
 
     // function to populate the questiontable database
-    function addItemQuestiontable(DiaryDatabase, module2, tab, idq, type, question, footer) {
+    function addItemQuestiontable(DiaryDatabase, module2, tab, idq, type, question, categories, footer) {
 
         DiaryDatabase.transaction(function (tx) {
 
-        var query = "INSERT INTO questiontable (module2, tab, idq, type, question, footer) VALUES (?,?,?,?,?,?)";
+        var query = "INSERT INTO questiontable (module2, tab, idq, type, question, categories, footer) VALUES (?,?,?,?,?,?,?)";
 
-        tx.executeSql(query, [module2, tab, idq, type, question, footer], function(tx) {
+        tx.executeSql(query, [module2, tab, idq, type, question, categories, footer], function(tx) {
         });
     });
     }
 
 // function to create the survey schedule
 
[…]
