+++ Android28/platforms/browser/www/js/dearscholar.js
@@ -255,7 +255,14 @@ routes: [
     {
         name: 'about',
         path: '/about/',
-        url: './about.html'
+        url: './about.html',
+        on: {
+        pageInit: function (e, page) {
+                $$('#consentAboutButton').on('click', function (e) {
+                    startConsentAbout()
+                });
+        }
+        }
     },
 ],
 // ... other parameters  
@@ -301,8 +308,8 @@ window.addEventListener("keyboardDidHide", function(e) {
     cordova.plugins.Keyboard.hideKeyboardAccessoryBar(true);
 });
 
-/////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Login functions
+///////////////////////////////////////////////////////////////////////////////
+///// Specific ANDROID functions
 
 // Timer functions to logout (return to index) after # time of inactivity.
 var timeoutID;
@@ -322,6 +329,8 @@ $$(window).on('click keyup', function (e) {
     window.clearTimeout(timeoutID);
     startTimer();
 });
+
+///////////////////////////////////////////////////////////////////////////////
 
 //Login procedure if touchID is enabled
 function loginTOUCH(){
@@ -487,6 +496,7 @@ function authentication(){
                             window.localStorage.setItem('project', settings[0].data.project);
                             window.localStorage.setItem('q0_startdate', settings[0].data.q0_startdate);
                             window.localStorage.setItem('q0_occasions', settings[0].data.q0_occasions);
+                            window.localStorage.setItem('q0_intervaltype', settings[0].data.q0_intervaltype);
                             window.localStorage.setItem('q0_interval', settings[0].data.q0_interval);
                      
                             // register the device to receive notifications
@@ -549,6 +559,11 @@ $$('#acceptConsent').on('click', function (e) {
         })   
 })
 
+//Add event handler to the close consent link in the About button.
+$$('#closeConsent').on('click', function (e) {
+    app.popup.close(".consentAbout")
+})
+
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////
 // Database functions needed to setup the diary database and all tables
     // Wait for Phonegap to load
@@ -582,8 +597,7 @@ $$('#acceptConsent').on('click', function (e) {
                  {  
                     surveyStructure = JSON.parse(data).data0;
                     pageStructure = JSON.parse(data).data1;
-                    questionTable = JSON.parse(data).data2;
-                                         
+                    questionTable = JSON.parse(data).data2;                                         
                     DiaryDatabase.transaction(function (tx){
                         tx.executeSql('DROP TABLE IF EXISTS surveyStructure');
                         tx.executeSql('DROP TABLE IF EXISTS pageStructure');
@@ -592,7 +606,7 @@ $$('#acceptConsent').on('click', function (e) {
                         tx.executeSql('CREATE TABLE IF NOT EXISTS surveyInformation (surveydate,A,B,C,D)');
                         tx.executeSql('CREATE TABLE IF NOT EXISTS surveyStructure (consent, manual, moduleAname, moduleBname, moduleCname, moduleDname)');
                         tx.executeSql('CREATE TABLE IF NOT EXISTS pageStructure (module1, classp, idp, header, backbuttontid, backbuttontext, backbuttononclick, backbuttonhref, nextbuttonid, nextbuttontext, nextbuttononclick,nextbuttonhref)');
-                        tx.executeSql('CREATE TABLE IF NOT EXISTS questiontable (module2, tab, idq, type, question, footer)');
+                        tx.executeSql('CREATE TABLE IF NOT EXISTS questiontable (module2, tab, idq, type, question, categories, footer)');
                         tx.executeSql('CREATE TABLE IF NOT EXISTS responseTable (id, surveydate, timestamp)');
                         
                         DiaryDatabase.transaction(function(tx) {
[…]
