+++ iOS/plugins/cordova-plugin-ionic-webview/src/ios/IONAssetHandler.m
@@ -1,77 +1,0 @@
-#import "IONAssetHandler.h"
-#import <MobileCoreServices/MobileCoreServices.h>
-#import "CDVWKWebViewEngine.h"
-
-@implementation IONAssetHandler
-
--(void)setAssetPath:(NSString *)assetPath {
-    self.basePath = assetPath;
-}
-
-- (void)webView:(WKWebView *)webView startURLSchemeTask:(id <WKURLSchemeTask>)urlSchemeTask
-{
-    NSString * startPath = @"";
-    NSURL * url = urlSchemeTask.request.URL;
-    NSString * stringToLoad = url.path;
-    NSString * scheme = url.scheme;
-
-    if ([scheme isEqualToString:IONIC_SCHEME]) {
-        startPath = self.basePath;
-        if ([stringToLoad isEqualToString:@""] || [url.pathExtension isEqualToString:@""]) {
-            startPath = [startPath stringByAppendingString:@"/index.html"];
-        } else {
-            startPath = [startPath stringByAppendingString:stringToLoad];
-        }
-    } else if ([scheme isEqualToString:IONIC_FILE_SCHEME]) {
-        if (![stringToLoad isEqualToString:@""]) {
-            startPath = stringToLoad;
-        }
-    }
-
-    NSData * data = [[NSData alloc] initWithContentsOfFile:startPath];
-    NSInteger statusCode = 200;
-    if (!data) {
-        statusCode = 404;
-    }
-    NSURL * localUrl = [NSURL URLWithString:url.absoluteString];
-    NSString * mimeType = [self getMimeType:url.pathExtension];
-    id response = nil;
-    if (data && [self isMediaExtension:url.pathExtension]) {
-        response = [[NSURLResponse alloc] initWithURL:localUrl MIMEType:mimeType expectedContentLength:data.length textEncodingName:nil];
-    } else {
-        NSDictionary * headers = @{ @"Content-Type" : mimeType, @"Cache-Control": @"no-cache"};
-        response = [[NSHTTPURLResponse alloc] initWithURL:localUrl statusCode:statusCode HTTPVersion:nil headerFields:headers];
-    }
-    
-    [urlSchemeTask didReceiveResponse:response];
-    [urlSchemeTask didReceiveData:data];
-    [urlSchemeTask didFinish];
-
-}
-
-- (void)webView:(nonnull WKWebView *)webView stopURLSchemeTask:(nonnull id<WKURLSchemeTask>)urlSchemeTask
-{
-    NSLog(@"stop");
-}
-
--(NSString *) getMimeType:(NSString *)fileExtension {
-    if (fileExtension && ![fileExtension isEqualToString:@""]) {
-        NSString *UTI = (__bridge_transfer NSString *)UTTypeCreatePreferredIdentifierForTag(kUTTagClassFilenameExtension, (__bridge CFStringRef)fileExtension, NULL);
-        NSString *contentType = (__bridge_transfer NSString *)UTTypeCopyPreferredTagWithClass((__bridge CFStringRef)UTI, kUTTagClassMIMEType);
-        return contentType ? contentType : @"application/octet-stream";
-    } else {
-        return @"text/html";
-    }
-}
-
--(BOOL) isMediaExtension:(NSString *) pathExtension {
-    NSArray * mediaExtensions = @[@"m4v", @"mov", @"mp4",
-                           @"aac", @"ac3", @"aiff", @"au", @"flac", @"m4a", @"mp3", @"wav"];
-    if ([mediaExtensions containsObject:pathExtension]) {
-        return YES;
-    }
-    return NO;
-}
-
-
-@end
