+++ iOS/plugins/cordova-sqlite-evcore-extbuild-free/spec/www/spec/db-open-close-delete-test.js
@@ -1,2131 +1,0 @@
-/* 'use strict'; */
-
-var MYTIMEOUT = 12000;
-
-// NOTE: DEFAULT_SIZE wanted depends on type of browser
-
-// FUTURE TODO replace in test(s):
-function ok(test, desc) { expect(test).toBe(true); }
-function equal(a, b, desc) { expect(a).toEqual(b); } // '=='
-function strictEqual(a, b, desc) { expect(a).toBe(b); } // '==='
-
-// XXX TODO REFACTOR OUT OF OLD TESTS:
-var wait = 0;
-var test_it_done = null;
-function xtest_it(desc, fun) { xit(desc, fun); }
-function test_it(desc, fun) {
-  wait = 0;
-  it(desc, function(done) {
-    test_it_done = done;
-    fun();
-  }, MYTIMEOUT);
-}
-function stop(n) {
-  if (!!n) wait += n
-  else ++wait;
-}
-function start(n) {
-  if (!!n) wait -= n;
-  else --wait;
-  if (wait == 0) test_it_done();
-}
-
-var isWindows = /MSAppHost/.test(navigator.userAgent);
-var isAndroid = !isWindows && /Android/.test(navigator.userAgent);
-var isFirefox = /Firefox/.test(navigator.userAgent);
-var isWebKitBrowser = !isWindows && !isAndroid && /Safari/.test(navigator.userAgent);
-var isBrowser = isWebKitBrowser || isFirefox;
-var isEdgeBrowser = isBrowser && (/Edge/.test(navigator.userAgent));
-var isChromeBrowser = isBrowser && !isEdgeBrowser && (/Chrome/.test(navigator.userAgent));
-var isSafariBrowser = isWebKitBrowser && !isEdgeBrowser && !isChromeBrowser;
-
-// should avoid popups (Safari seems to count 2x)
-var DEFAULT_SIZE = isSafariBrowser ? 2000000 : 5000000;
-// FUTURE TBD: 50MB should be OK on Chrome and some other test browsers.
-
-// NOTE: While in certain version branches there is no difference between
-// the default Android implementation and implementation #2,
-// this test script will also apply the androidLockWorkaround: 1 option
-// in case of implementation #2.
-var pluginScenarioList = [
-  isAndroid ? 'Plugin-implementation-default' : 'Plugin',
-  'Plugin-implementation-2'
-];
-
-var pluginScenarioCount = isAndroid ? 2 : 1;
-
-var mytests = function() {
-
-  describe('Open database parameter test(s)', function() {
-    // TBD skip plugin test on browser platform (not yet supported):
-    if (isBrowser) return;
-
-    for (var i=0; i<pluginScenarioCount; ++i) {
-
-      describe(pluginScenarioList[i] + ': open database file name test(s)', function() {
-        var scenarioName = pluginScenarioList[i];
-        var suiteName = scenarioName + ': ';
-        var isImpl2 = (i === 1);
-
-        // NOTE: MUST be defined in function scope, NOT outer scope:
-        var openDatabase = function(first, second, third, fourth, fifth, sixth) {
-          var dbname, okcb, errorcb;
-
-          if (first.constructor === String ) {
-            dbname = first;
-            okcb = fifth;
-            errorcb = sixth;
-          } else {
-            dbname = first.name;
-            okcb = second;
-            errorcb = third;
-          }
-
-          if (!isImpl2) {
-            return window.sqlitePlugin.openDatabase({name: dbname, location: 'default'}, okcb, errorcb);
-          }
-
-          var dbopts = {
-            name: 'i2-'+dbname,
-            androidDatabaseImplementation: 2,
-            androidLockWorkaround: 1,
-            location: 1
-          };
-
[…]
