+++ iOS/plugins/cordova-sqlite-evcore-extbuild-free/spec/www/spec/db-tx-error-handling-test.js
@@ -1,4076 +1,0 @@
-/* 'use strict'; */
-
-var MYTIMEOUT = 20000;
-
-// NOTE: DEFAULT_SIZE wanted depends on type of browser
-
-var isWindows = /MSAppHost/.test(navigator.userAgent);
-var isAndroid = !isWindows && /Android/.test(navigator.userAgent);
-var isFirefox = /Firefox/.test(navigator.userAgent);
-var isWebKitBrowser = !isWindows && !isAndroid && /Safari/.test(navigator.userAgent);
-var isBrowser = isWebKitBrowser || isFirefox;
-var isEdgeBrowser = isBrowser && (/Edge/.test(navigator.userAgent));
-var isChromeBrowser = isBrowser && !isEdgeBrowser && (/Chrome/.test(navigator.userAgent));
-var isSafariBrowser = isWebKitBrowser && !isEdgeBrowser && !isChromeBrowser;
-
-// should avoid popups (Safari seems to count 2x)
-var DEFAULT_SIZE = isSafariBrowser ? 2000000 : 5000000;
-// FUTURE TBD: 50MB should be OK on Chrome and some other test browsers.
-
-// NOTE: While in certain version branches there is no difference between
-// the default Android implementation and implementation #2,
-// this test script will also apply the androidLockWorkaround: 1 option
-// in case of implementation #2.
-var scenarioList = [
-  isAndroid ? 'Plugin-implementation-default' : 'Plugin',
-  'HTML5',
-  'Plugin-implementation-2'
-];
-
-var scenarioCount = (!!window.hasWebKitWebSQL) ? (isAndroid ? 3 : 2) : 1;
-
-var mytests = function() {
-
-  for (var i=0; i<scenarioCount; ++i) {
-    // TBD skip plugin test on browser platform (not yet supported):
-    if (isBrowser && (i === 0)) continue;
-
-    describe(scenarioList[i] + ': tx error handling (detailed tx error handling) test(s)', function() {
-      var scenarioName = scenarioList[i];
-      var suiteName = scenarioName + ': ';
-      var isWebSql = (i === 1);
-      var isImpl2 = (i === 2);
-      // XXX TBD WORKAROUND SOLUTION for (WebKit) Web SQL on Safari browser (TEST DB NAME IGNORED):
-      var recycleWebDatabase = null;
-
-      // NOTE: MUST be defined in function scope, NOT outer scope:
-      var openDatabase = function(name, ignored1, ignored2, ignored3) {
-        if (isWebSql && isSafariBrowser && !!recycleWebDatabase)
-          return recycleWebDatabase;
-        if (isImpl2) {
-          return window.sqlitePlugin.openDatabase({
-            // prevent reuse of database from default db implementation:
-            name: 'i2-'+name,
-            androidDatabaseImplementation: 2,
-            androidLockWorkaround: 1,
-            location: 1
-          });
-        }
-        if (isWebSql) {
-          return recycleWebDatabase =
-            window.openDatabase(name, '1.0', 'Test', DEFAULT_SIZE);
-        } else {
-          return window.sqlitePlugin.openDatabase({name: name, location: 0});
-        }
-      }
-
-      describe(scenarioList[i] + ': tx error semantics test(s)', function() {
-
-        /* found due to investigation of litehelpers/Cordova-sqlite-storage#226: */
-        it(suiteName + 'SKIP SQL CALLBACKS after syntax error with no handler', function(done) {
-          var db = openDatabase('first-syntax-error-with-no-handler.db', '1.0', 'Test', DEFAULT_SIZE);
-          expect(db).toBeDefined();
-
-          db.transaction(function(tx) {
-            expect(tx).toBeDefined();
[…]
