### 2020-07-05 331fb7f
Update dearscholar.js
