### 2020-07-06 #2354 @cMadan
@pmkruyen, thanks for making those changes.

1. I think that should be fine now. JOSS submissions need to be released under an (unmodified) OSI license, but it's fine to have an additional copyright statement. (@openjournals/joss-eics, can you confirm this text is suitable?)

2. Thanks for including the screenshots in the README, I think it makes the app functions much clearer for new users.

3. I also appreciate you adding th MySQL initialisation code.

Do you have any additional suggestions for reviewers (see the initial comment in this issue for some guidance) ?
