### 2020-07-06 #2354 @pmkruyen
Hi @cMadan, thanks for your reply.

1. To be sure: I searched for [at]openjournals/joss-eics, because I thought I had to check something. On rereading your response, I believe you would like the joss team to confirm your response? The link to [at]openjournals/joss-eics throws a 400 error however.

2. :)

3. Triggered by your MySQL suggestion, I'm now working on a example php script that can researchers use as template to let DearScholar communicate with the server (but there are some delays here because Phonegap is no longer working for iOS production and, consequently, I need invest some time to set up the building tools locally).

With regard to suggestions for reviewers, do you would me to suggest additional reviewers besides the potential reviewers I suggested above?

> gcdeshpande (mobile apps)
> olade014 (android)
> alletsee (survey research and longitudinal data collection)
> darthpathos (survey research and longitudinal data collection)
> andytwoods (javascript, psychology, online research)

Best,
Peter
