### 2020-07-06 #2354 @danielskatz
> I think that should be fine now. JOSS submissions need to be released under an (unmodified) OSI license, but it's fine to have an additional copyright statement. (@openjournals/joss-eics, can you confirm this text is suitable?)

It seems fine
