### 2020-07-06 #2354 @gcdeshpande
Yes

Thanks and Regards
Gajendra C. Deshpande
Assistant Professor, CSE@KLS.GIT
Belgaum - 590008
Karnataka, INDIA

Blog: http://gajendramoksha.wordpress.com
Website: http://sites.google.com/a/git.edu/gcd
Facebook: https://www.facebook.com/gcdeshpande

[http://csta.acm.org/Membership/sub/badge.jpg]
