### 2020-07-06 #2354 @alletsee
Thanks for reaching out. Before deciding if I accept to review, I think it makes sense to discuss expectations regarding the programming languages. I have minimal experience with JavaScript and no experience with mobile app development. However, I think that my knowledge about survey research would be good enough to review the paper itself.

In essence, I'm not sure if it makes sense to review in a situation where my ability to review the software contribution is very limited. Is there a guideline / recommendation from the JOSS team on this, @cMadan ?
