### 2020-07-08 #2354 @gcdeshpande
@cMadan request you to kindly share the evaluation criteria and process.
