### 2020-07-08 #2354 @kyleniemeyer
Hello @gcdeshpande, here is a link to our review criteria: https://joss.readthedocs.io/en/latest/review_criteria.html

@alletsee Thanks for responding—if you are not really able to review the software, then we might prefer to find alternate reviewers (JOSS papers are fairly short). However, if you feel comfortable at least reviewing the functionality and documentation of the software, bringing your domain expertise into account, then that would still be helpful for us.
