+++ Android26/plugins/cordova-plugin-ionic-webview/src/android/com/ionicframework/cordova/webview/AndroidProtocolHandler.java
@@ -68,14 +68,22 @@ public class AndroidProtocolHandler {
   }
 
   public InputStream openFile(String filePath) throws IOException  {
-    File localFile = new File(filePath);
+    String realPath = filePath.replace(WebViewLocalServer.fileStart, "");
+    File localFile = new File(realPath);
     return new FileInputStream(localFile);
   }
 
   public InputStream openContentUrl(Uri uri)  throws IOException {
+    Integer port = uri.getPort();
+    String realPath;
+    if (port == -1) {
+      realPath = uri.toString().replace(uri.getScheme() + "://" + uri.getHost() + WebViewLocalServer.contentStart, "content:/");
+    } else {
+      realPath = uri.toString().replace(uri.getScheme() + "://" + uri.getHost() + ":" + port + WebViewLocalServer.contentStart, "content:/");
+    }
     InputStream stream = null;
     try {
-      stream = context.getContentResolver().openInputStream(Uri.parse(uri.toString().replace(WebViewLocalServer.ionicContentScheme + ":///", "content://")));
+      stream = context.getContentResolver().openInputStream(Uri.parse(realPath));
     } catch (SecurityException e) {
       Log.e(TAG, "Unable to open content URL: " + uri, e);
     }
