### 2020-07-12 512e632
add device.platform == browser
