+++ App/www/js/dearscholar.js
@@ -383,7 +383,7 @@ function goInactive() {
      if(device.platform == "iOS"){
         app.views.main.router.back('/',{force: true, ignoreCache: true, reload: true});
      };
-     if(device.platform == "Android"|| device.platform == "android"){
+     if(device.platform == "Android"|| device.platform == "android" || device.platform == "browser"){
         app.views.main.router.back('/',{ignoreCache: true, reload: true});
      };
     app.panel.close("left")
@@ -865,7 +865,7 @@ function measurementDates (startdate=window.localStorage.getItem("q0_startdate")
          document.getElementById("emptyconsentform").innerHTML += consentText;
           
          //restyle the Android toolbar 
-         if(device.platform == "Android"|| device.platform == "android"){
+         if(device.platform == "Android"|| device.platform == "android" || device.platform == "browser"){
             $(".toolbar-bottom").css("height", "120px");
             $(".toolbar-inner").css("height", "40px");
          }  
@@ -891,7 +891,7 @@ function measurementDates (startdate=window.localStorage.getItem("q0_startdate")
          document.getElementById("emptyconsentformAbout").innerHTML += consentText;
           
          //restyle the Android toolbar 
-         if(device.platform == "Android"|| device.platform == "android"){
+         if(device.platform == "Android"|| device.platform == "android" || device.platform == "browser"){
             $(".toolbar-bottom").css("height", "120px");
             $(".toolbar-inner").css("height", "40px");
          }
@@ -1015,7 +1015,7 @@ function startModule(module, adhoc) {
                 </div>
             </div>`
     }
-    if (device.platform=="Android"|| device.platform == "android"){
+    if (device.platform=="Android"|| device.platform == "android" || device.platform == "browser"){
     return `
         <div id="${data.id}" class=${data.classp}>
             <div class="block-header">${data.header}</div>
