### 2020-07-13 0b1779a
fixed device.platform browser error
