+++ readme.md
@@ -230,7 +230,7 @@ To log in, populate the survey modules with question pages and questions, and st
 A sample php script that can be used to let DearScholar communicate with the server can be found [here](https://github.com/pmkruyen/dearscholar/blob/master/dearscholar.php).
 
 #### Sample study
-To set up a sample study, [this](https://github.com/pmkruyen/dearscholar/blob/master/dearscholar.php) sample php file can be used. When generating random usernames and passwords, avoid using question marks (?) and ampersands (&).
+To set up a sample study, [this](https://github.com/pmkruyen/dearscholar/blob/master/samplestudy.php) sample php file can be used. When generating random usernames and passwords, avoid using question marks (?) and ampersands (&).
 
 ### Step 3: Experiment with DearScholar on your mobile device
 In order experiment with DearScholar on iOS and Android devices, the app should be build (compiled):
