### 2020-07-14 #2354 @pmkruyen
Hi @gcdeshpande, thank you for your willingness to review my submission.

For your information, inspired by @cMadan 's initial suggestions, I took the opportunity-while my submission still being in 'pre-review'-to write a sample php script to let DearScholar communicate with the server (`dearscholar.php`). Also, I wrote a php file to set up a sample study (`samplestudy.php`). Both files are in the root of the repro. Last, I fixed some minor bugs caused by changes in the compilation process for iOS. If you need more information to get DearScholar running, let me know.

Best,
Peter
