### 2020-07-15 4480b6a
complete samplestudy settings
