### 2020-07-15 5b1fc44
complete samplestudy settings
