### 2020-07-15 c0aeacd
complete samplestudy settings
