### 2020-07-15 #2354 @kinow
@cMadan hello. Finished reading the comments in this issue, and also the link to the review criteria above, as well as having a quick look at the paper PDF.

I'd be happy to review it. It would be my first review for JOSS, so let me know if there's anything else I need to read besides the review criteria (even though that document seems to cover most of what I could imagine for reviewing a paper about a software).

I used Cordova before too (~4 years ago I think?), and at work I have a co-worker actively maintaining a phonegap/ionic app. So I should be able to review the project code, confirm it works, and finally review the paper as well.

Thanks!
Bruno
