+++ Android26/www/js/dearscholar.js
@@ -13,6 +13,9 @@ id: 'net.peterkruyen.diary',
 dialog: {
     title: 'DearScholar',
   },
+navbar: {
+    mdCenterTitle: true
+  },
 routes: [
     {
         path: '/',
@@ -307,14 +310,12 @@ function onBackKeyDown() {
 window.addEventListener("keyboardWillShow", function(e) {
     //if(device.platform == "Android"|| device.platform == "android"){
         $(".bottomButtons").hide();
-        cordova.plugins.Keyboard.hideKeyboardAccessoryBar(false);
     //}
 });
 
 window.addEventListener("keyboardDidHide", function(e) {
     //if(device.platform == "Android"|| device.platform == "android"){
         $(".bottomButtons").show();
-        cordova.plugins.Keyboard.hideKeyboardAccessoryBar(true);
     //}
 });
 
