### 2020-07-18 1b3a24c
fix 'done' button on keyboard
