### 2020-07-22 8cd2e8a
update loginscreen Android
