+++ Android26/www/img/LICENSE.md
@@ -1,1 +1,2 @@
-DearScholar icons provided by Icons8 (https://icons8.com/)
+DearScholar icons provided by Icons8 (https://icons8.com/) * icon size = 70%
+
