### 2020-07-22 #2354 @cMadan
@pmkruyen @kinow @gcdeshpande, thank you for your patience. I think we're all set to start the review, so I will set that up and we can begin the actual review process.

@kinow, thanks for mentioning this is your first JOSS review, feel free to ask me questions in the review issue if you're not sure about something!
