### 2020-07-22 #2506 @whedon
[REVIEW]: DearScholar: A mobile application to conduct qualitative and quantitative diary research
**Submitting author:** @pmkruyen (<a href="http://orcid.org/0000-0003-0109-1744">Peter Kruyen</a>)
**Repository:** <a href="https://github.com/pmkruyen/dearscholar.git" target ="_blank">https://github.com/pmkruyen/dearscholar.git</a>
**Version:** v1.0.49
