+++ readme.md
@@ -1,3 +1,5 @@
+[![status](https://joss.theoj.org/papers/1896b88f26b987b9c7a07035751afd7b/status.svg)](https://joss.theoj.org/papers/1896b88f26b987b9c7a07035751afd7b)
+
 # Overview
 Because collecting longitudinal data becomes more important in academic research nowadays and because limitations with available tools, DearScholar has been developed.
 
