### 2020-07-23 pmkruyen/dearscholar#18 @kinow
Build instructions
Hi @pmkruyen

Could you point to the building instructions, please? There are two folders with `package.json`, and running `yarn install` should probably download the dependencies. But given these are cordova/phonegap projects, I believe I would need some dependencies installed (node, npm/yarn, cordova, etc).

Would you have it available anywhere, please?

Related to JOSS paper review: https://github.com/openjournals/joss-reviews/issues/2506

Thanks
Bruno
