### 2020-07-23 #2506 @kinow
Looking at this item now, under documentation

> Installation: Does installation proceed as outlined in the documentation?

https://github.com/pmkruyen/dearscholar/issues/18
