### 2020-07-23 pmkruyen/dearscholar#19 @kinow
Documentation issue rendering lists
I think the lists in the docs are not being rendered correctly. Maybe due to the lack of space of newline break. The syntax used by GitHub sometimes is a bit different, though they also offer [a guide](https://guides.github.com/features/mastering-markdown/) on how to style with Markdown.

![image](https://user-images.githubusercontent.com/304786/88250202-f60a9380-ccfa-11ea-804f-da42743c7d84.png)

Found while reviewing https://github.com/openjournals/joss-reviews/issues/2506
