### 2020-07-23 pmkruyen/dearscholar#20 @kinow
Provide a test reporting, or document how the project is tested
Hi,

It looks like BrowserStack is used to test the mobile app. Is there any dashboard or report that is available to users? They may be interested to know if there are any failures for a build/version, or if there is any information on what was tested.

Do you have test coverage anywhere?

What about the backend app? I think that's PHP, from what I understand so far (waiting for #18 to build the environment and test locally). Are there any tests for the backend?

From: https://github.com/openjournals/joss-reviews/issues/2506

Cheers
Bruno
