### 2020-07-23 pmkruyen/dearscholar#19 @pmkruyen
Thanks for pointing out. I fixed this in the readme.md files (master and developer).
