### 2020-07-23 pmkruyen/dearscholar#19 @kinow
Thanks for fixing it so quickly @pmkruyen !
