### 2020-07-23 pmkruyen/dearscholar#20 @pmkruyen
Hi Bruno,

Yes, I use BrowserStack. When you send me your email address, I will grant you access to the test environment, and send you the login credentials for both the developer iOS and Android version. Ok?

Meanwhile, I think you also suggest to set up an automatic testing procedure that generates reports automatically? Until now, we (: me, my PhD students who use the app in their research, and end users) have tested the app only manually (: e.g., whether the measurement schedule is rendered properly, whether the data is correctly stored on the server, whether push notifications are received). I think I need some time to set up automatic testing procedures as this is very new to me :).

With regard to the backend app. It is not so fancy yet. I set up phpmyadmin panel for the researchers (which is shield by an additional password and can only be accessed through an eduroam vpn connection [used by higher education institutes]). I don't think I need any tests here, but if you have suggestions how to achieve this, I would be happy to hear (and implement).

Best,
Peter
