### 2020-07-23 pmkruyen/dearscholar#20 @kinow
Hi Peter,

I think it would be useful if developers/users could run the tests. Are you using Selenium, Nightwatch, Cypress, or any other test system for testing with the browser at BrowserStack? Or is it fully manual?

The backend doesn't need a lot of tests, but I think it would be important to offer a minimum set of tests. The [project homepage](https://peterkruyen.net/dearscholar.html) has a card/panel about security, for example. There could be a test to verify that a password is salted, or an error message is given when the user doesn't provide user and password, etc.

Or even simpler unit tests.

This is related to this item from my checklist, under Documentation:

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I think we can leave it unchecked if there are no tests for now. Not sure if that's an issue or not, I think after I finish the review the editor will take a look at the checklist and comment.
