### 2020-07-23 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

So far, I have used the PhoneGap Desktop App to build the app locally (so downloading DearScholar, and opening this project using in the PhoneGap Desktop App should work). [Also see, "Guide for potential contributors and academic researchers" in the readme file].

When this doesn't work or you want more specific building instructions, please let me know.

An option is to write a small program that installs the Desktop App, php, MySQL, and myPhPadmin locally?

Best, Peter
