### 2020-07-23 pmkruyen/dearscholar#20 @pmkruyen
Hi Bruno,

The app = It is fully manual now, but will set up such an automatic testing procedure. Let you know when I have it running.

The backend = I will have another look at this, will let you know when I have implemented some tests here too :)

Best,
Peter
