### 2020-07-23 pmkruyen/dearscholar#18 @kinow
Hi Peter,

I will have to check how to use phonegap, to remind me how to install and run it. But it would be simpler if there were build instructions.

Here's an example of [Annif](https://github.com/NatLibFi/Annif/#basic-install), a Python backend that includes a small JS frontend. There is also a Dockerfile that can be used for both building instructions, or to run the application locally.

Or this [other project](https://github.com/52North/SOS#quick-start) that contains both detailed installation instructions (points to another page) and how to run the application with docker.

Perhaps you could just say something like:

- Download phonegap (does it work with any version?)
- Run the following commands for the Android folder
- Run the following commands for the App folder (not sure if there's any order)
- To run locally, then run the Android emulator pointing to a certain file/apk
- Access the application in the URL `http://...`, etc

Having a Docker container is convenient, but I think the installation instructions are more important. Especially if there are any limitations (OS version, system library, etc).

For example, there is at least one PHP script using `mysql_error` function. I think it was [removed in PHP 7](https://www.php.net/mysql_error/)? I am planning to confirm that later, but perhaps if you could take a look at the version of the software required, that'd be easier.

Cheers
Bruno
