### 2020-07-23 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

Thanks for these detailed suggestions. Very much appreciated. I will give priority to this issue--that is, provide more detailed installation instructions and check docker, which I have never used--and solve issue #20 next.

Best,
Peter

PS. Thanks for point out this inconsistency with the PHP script, will check all this. For your information I use `PHP 7.2.24 ` and `mysql  Ver 14.14` on Ubuntu. I will include this in the instructions too.
