### 2020-07-23 pmkruyen/dearscholar#18 @kinow
Perfect Peter! I could be wrong on the PHP/MySQL issue.

Very interested to see what DearScholar looks like running it locally while reading the code to understand how it works :-)

Cheers
Bruno
