### 2020-07-23 pmkruyen/dearscholar#18 @pmkruyen
Thanks :) ... as I am no software developer--just a researcher with programming skills in R--making this was (is) fun *and* challenging, gradually getting better in it. All feedback is very welcome.
