### 2020-07-23 pmkruyen/dearscholar#18 @pmkruyen
Thanks :) ... All feedback is very welcome.
