### 2020-07-23 #2506 @cMadan
@pmkruyen, thanks for pointing out the version number issue. I'll update it to 1.048 for now, but usually the accepted version of the software is a later version number than the submitted one (e.g., changes in response to the reviewers' comments), so I will update it again when we're closer to acceptance.
