### 2020-07-23 #2506 @kinow
Hi @cMadan , thanks for assigning me as reviewer here. I will take up your offer, when you said I could ask if I had questions. I hope these are not too silly.

Q1/ The check list and the linked documentation are great! I can see the performance check box can be left unchecked if there are no claims of performance. What if the project doesn't have installation instructions, or tests? I asked the project maintainer to add those and he is working on that now. But what if the author decides those are not important? Then I just leave the checkbox unchecked and say that I have finished my review?

Q2/ When I was reviewing the `readme.md` file of the project, I found some markdown syntax issues. Normally I use GitHub UI to quickly send a PR whenever if find these in other projects. But since I am a reviewer of the project's paper, I wasn't sure if that would be alright? Would that be OK?

Q3/ After the review is done, should I avoid contributing to projects of papers I reviewed?

Thanks!
Bruno
