### 2020-07-24 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

To give you an update on the installation instructions.

I am able to build a Docker image of the server (using this repro: https://github.com/jcavat/docker-lamp) and storing it on Docker Hub. However, when I download the image and try to restart it, it does not work (the admin panel does not show up at any port on my Mac and neither can I reach the php file). Once I have solved this issue, I will write clear installation instructions.

Best,
Peter
