### 2020-07-24 pmkruyen/dearscholar#18 @kinow
Hi Peter!

Really impressed! You mentioned your daily job was not writing code. I was going to comment here saying that preparing an image for Docker could be difficult. But looks like you are already doing great progress!

If you'd like to get a small container size, take a look at the Docker best practices [doc](https://docs.docker.com/develop/develop-images/dockerfile_best-practices/). It's probably a ~20-40 minutes article, but may be helpful in case you ever need to use other images that appear to be too large.

Cheers
Bruno
