### 2020-07-27 #2506 @cMadan
Hi @kinow,

Q1. For publication at JOSS, we would require install instructions and some form of testing. (E.g., automated tests are not required, but some minimal example of input/output would be.) As is the case here, then we would then wait for the author to work on those aspects of the project, and the review is 'on-going' until the project is suitable for publication. So, it's not just a 'one-pass and done' situation, but more of periodic back-and-forth as the authors attempt to address your comments.

Q2. Go ahead, that's perfectly fine! (And probably would help speed the process along.)

Q3. It's generally viewed that contributions you make as a reviewer do not count towards co-authorship on the project. But, should you continue to contribute to the project after the JOSS review process, then it may make sense for you to be a co-author on subsequent publications that come after the JOSS paper, and would involve a discussion with the project's authors accordingly.

I hope that helps clarify things!
