+++ paper.md
@@ -29,7 +29,7 @@ A limited number of (commercial) diary research apps have been developed, includ
 DearScholar is a hybrid, open-source smartphone application (app) that can be used on iOS devices (iPhones and Ipads) and Android devices (basically all other smartphones and tablets) to conduct diary studies (and, obviously, can also be used for other types of longitudinal research such as repeated-survey designs and log studies). DearScholar's aim is to facilitate the research process for both respondents and researchers taken the limitations of the alternatives for DearScholar into consideration. Researchers only have to specify measurement occassions (measurement schedule),  survey layout, and question format. Respondents have to download the app on the App Store<sup>1</sup> or Google Play<sup>2</sup>, fill out their credentials, and start participating.
 
 # Use cases
-Currently, DearScholar is used in a study by Glenn Houtgraaf MSc, dr. Peter M. Kruyen and prof. dr. Sandra van Thiel to investigate work-related creativity in government organizations. We use DearScholar to follow about 100 participants over a period of six months, asking them closed- and open-ended questions at bi-weekly measurement occassions to investigate creative processes. In September and October, Liesbeth Faas MSc, dr. Peter M. Kruyen, adn prof. dr. Sandra van Thiel will replicate this study in local care teams.
+Currently, DearScholar is used in a study by Glenn Houtgraaf MSc, dr. Peter M. Kruyen and prof. dr. Sandra van Thiel to investigate work-related creativity in government organizations. We use DearScholar to follow about 100 participants over a period of six months, asking them closed- and open-ended questions at bi-weekly measurement occassions to investigate creative processes. In September and October, Liesbeth Faas MSc, dr. Peter M. Kruyen, and prof. dr. Sandra van Thiel will replicate this study in local care teams.
 
 # Acknowledgement
 DearScholar is developed within the context of the research programme "The creative public servant: Observations, explanations and consequences" with project number 406.18.R8.028, financed by the Dutch Research Council (NWO). The author wants to express his graditute to Glenn Houtgraaf MSc, Liesbeth Faas MSc, and the ICT Services (Radboud University, Nijmegen, the Netherlands) for their advice, testing, and feedback; prof. dr. Sandra van Thiel for her encouragements; and last, but not least, all (pilot) respondents for their effort and feedback during the developmental process.
