### 2020-07-27 6e7fb94
update installation instructions
