+++ App/www/js/dearscholar.js
@@ -1,3 +1,5 @@
+// Copyright (c) 2020 P.M. Kruyen, Institute for Management Research, Radboud University, the Netherlands. 
+
 var app = new Framework7({
 // App root element
 root: '#app',
@@ -183,7 +185,7 @@ routes: [
                     
                     $.ajax({
                         type:"POST",  
-                        url:"https://peterkruyen.net/diary/dearscholar.php", data: dataString,
+                        url: locationPhP, data: dataString,
                         crossDomain: true,
                         cache: false, 
                         success:function(data){  
@@ -256,6 +258,11 @@ routes: [
 var $$ = Dom7;
 
 var mainView = app.views.create('.view-main')
+
+///////////////////////////////////////////////////////////////////////////////
+// General variables
+var locationPhP = "http://localhost/dearscholar.php";
+//var locationPhP = "https://peterkruyen.net/diary/dearscholar.php";
 
 ///////////////////////////////////////////////////////////////////////////////
 // Style the login screen
@@ -484,7 +491,7 @@ function storePIN(PIN){
     
         $.ajax({
                  type:"POST",  
-                 url:"https://peterkruyen.net/diary/dearscholar.php", data: dataString,
+                 url: locationPhP, data: dataString,
                  crossDomain: true,
                  cache: false, 
                  success:function(data)  
@@ -517,7 +524,7 @@ function checkPIN(PIN){
     
         $.ajax({
                  type:"POST",  
-                 url:"https://peterkruyen.net/diary/dearscholar.php", data: dataString,
+                 url: locationPhP, data: dataString,
                  crossDomain: true,
                  cache: false, 
                  success:function(data)  
@@ -552,7 +559,7 @@ function authentication(){
             
                 $.ajax({
                  type:"POST",  
-                 url:"https://peterkruyen.net/diary/dearscholar.php", data: dataString,
+                 url: locationPhP, data: dataString,
                  crossDomain: true,
                  cache: false, 
                  success:function(data)  
@@ -574,7 +581,7 @@ function authentication(){
             
                                 $.ajax({
                                     type:"POST",  
-                                    url:"https://peterkruyen.net/diary/dearscholar.php", data: dataString,
+                                    url: locationPhP, data: dataString,
                                     crossDomain: true,
                                     cache: false, 
                                     success:function(data)  
@@ -616,7 +623,7 @@ $$('#acceptConsent').on('click', function (e) {
                  
         $.ajax({
             type:"POST",  
-            url:"https://peterkruyen.net/diary/dearscholar.php", data: dataString,
+            url: locationPhP, data: dataString,
             crossDomain: true,
             cache: false, 
             success:function(data){  
@@ -653,7 +660,7 @@ $$('#acceptConsent').on('click', function (e) {
          var dataString="uname="+uname+"&pwd="+pwd+"&project="+project+"&findvalues=true";
                  $.ajax({
                  type:"POST",  
-                 url:"https://peterkruyen.net/diary/dearscholar.php", data: dataString,
+                 url: locationPhP, data: dataString,
                  crossDomain: true,
                  cache: false, 
                  success:function(data)  
@@ -1432,7 +1439,7 @@ function uploadToServer(module){
 
         $.ajax({
                  type:"POST",  
-                 url:"https://peterkruyen.net/diary/dearscholar.php", data: dataString,
+                 url: locationPhP, data: dataString,
                  crossDomain: true,
                  cache: false, 
                  success:function(data)  
@@ -1506,7 +1513,7 @@ function checkNewMessages(){
          var dataString="uname="+uname+"&pwd="+pwd+"&messagesCheck=true";
                  $.ajax({
                  type:"POST",  
-                 url:"https://peterkruyen.net/diary/dearscholar.php", data: dataString,
[…]
