+++ readme.md
@@ -234,7 +234,7 @@ To get the developer version of DearScholar running on your computer, the follow
 1) download the app files and folders as specified below (under file structure);
 2) set up an Apache webserver (for testing you may want to use a local Docker container); 
 3) on this webserver, set up all the required tables and files.
-3) install either the desktop app or command line interface of [Phonegap](https://phonegap.com/getstarted/);
+3) install the desktop app of [Phonegap](https://phonegap.com/getstarted/);
 4) compile DearScholar on your computer using Phonegap ... the app appears in your web browser; and 
 5) you can start improving DearScholar.
 
@@ -268,9 +268,9 @@ Obviously, after Step 3 and 4, you may want to submit your modified  version of 
 When you submit an app to the App Store or Google Play Store for publication, your app will be reviewed. The review procedure can be a black box, hence, I share some experiences with my submissions of DearScholar to the stores. While submissions to Google Play Store have been flawless so far, over time the App Store rejected several builds, probability because DearScholar has been categorized as a *Health and Health Research* app. To get the app published on the App Store, I have to show evidence of ethical approvement by a formal ethical committee for the running research projects; information about my relationship with the university were I work and the running research projects; and detailed specifications of the Informed Consent Form that needed to be included in the app. While information about the reasons for the study, information about confidentiality and handling of data, a point of contact for participant questions, and withdrawal procedures were specified already, the App Store reviewers also required information about the benefits and risks for research participants as well as a note about consent procedures for minors.
 
 ## Installation
-Researchers and contributors who want to test, contribute, or want to build their own version of DearScholar can follow the following installation steps (after reading the **Guide for potential contributors and academic researchers** above). It is my ambition to automate these steps. Meanwhile, performing the steps manually gives a good first impression about how DearScholar functions.
+Researchers and contributors who want to test, contribute, or want to build their own version of DearScholar can follow the  installation instructions in this section (after reading the **Guide for potential contributors and academic researchers** above). 
 
-*This section is in progress*. 
+It is my ambition to automate these steps. Meanwhile, performing the steps manually gives a good first impression about how DearScholar functions.
 
 ### Step 1: Server setup
 To use Dearscholar, an Apache webserver (version 2.4.29) needs to be set up that runs MySQL (5.7.30), PHP (version 7.4.6), and phpMyAdmin (version 5.0.2) as front panel for researchers. For production, you would run this webserver on a real webserver. For testing, a local [Docker](https://www.docker.com/) container can be used as testing environment on your own device (laptop, pc). 
@@ -299,7 +299,7 @@ You are now in the (virtual) terminal of the Apache webserver. In this terminal,
 Refresh phpMyAdmin in your browser to check if the two databases and the (empty) tables are created. Exit MySQL by running ```exit;``` in the terminal. 
 
 ### Step 3: MySQL setup - part 2
-Now it is time to populate both databases for a sample study. To do so, we first insert usernames, hashed passwords, and the project settings for five mock respondents in the authentication table (in the ```respondents``` database).
+Populate both databases for a sample study. To do so, we first insert usernames, hashed passwords, and the project settings for five mock respondents in the authentication table (in the ```respondents``` database).
 
 Open the file [samplestudy-part1.php](https://github.com/pmkruyen/dearscholar/blob/master/samplestudy-part1.php) in a text editor and ammend the ```$MyPassword``` variable using the MySQL admin password that was generated at first run of the Docker container.
 
@@ -316,10 +316,19 @@ Next, create the response tables in the ```project``` database and set up the ap
 
[…]
