### 2020-07-27 cb99a7a
update installation instructions
