### 2020-07-27 1739b33
update installation instructions
