### 2020-07-28 34658eb
update installation instructions
