### 2020-07-28 32e96ed
update installation instructions
