### 2020-07-28 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

Thanks for your help.

Over the weekend I wrote the installation instructions in the README.md file. I learned a lot about Docker, really helpful. Putting everything in the Docker seems ideal to me, but I thought it is best to start with writing out manual installation instructions (but using a Docker image for the server make things a lot easier already).

Could you please have a look at these instructions, and tell me if the instructions are clear _and_ work to launch DearScholar locally?

Thanks,
Peter

(Note. I also corrected some minor errors and updated some procedures in the sample-study set up files.)
