### 2020-07-28 1184359
update installation instructions
