### 2020-07-28 #2506 @pmkruyen
Hi @gcdeshpande and @kinow,

To give a short update:

- Because the App Store approved version 1.0.48 of the app, I merged the developer and master branch in the GitHup repro to keep things manageable. The GitHub version is 1.0.49. Both versions are similar except that version 1.0.49 offers researchers the possibility to change the numbers of survey modules while version 1.0.48 does not offer this possibility. I am planning to upload version 1.0.49 soon to the stores because of requirements for my current research projects, taking your feedback into account.

- Over the weekend, I wrote down installation instructions (see 'Installation' section, in `Readme.md`. Hope these instructions are clear and work :)[also see issue 18 in the project repro].

- Will search for ways to include some (automatic) testing for the server and the app.

Best,
Peter
