### 2020-07-28 d0314d1
Update readme (typos/re-wording)
