### 2020-07-28 pmkruyen/dearscholar#18 @pmkruyen
Thanks for testing.

O... I understand, the `google-services.json` file is missing. I did not upload it because it contains sensitive data (needed for sending push notifications to Android devices).

I have uploaded a sample `google-services.json` file in the App and Android26 file. This fixed the error on my machine. Can you let me know if this works for you too?

Thanks,
Peter

PS. Also thanks for showing me how you build the app on the command line :)
