### 2020-07-28 pmkruyen/dearscholar#18 @kinow
Thanks for the `google-services.json` file. It appears to have fixed my error. We are doing good progress I think!

Next error:

```
Updating resource files at platforms/android
  copy  google-services.json platforms/android/app/google-services.json (new file)
Prepared android project successfully
Executing script found in plugin cordova-plugin-add-swift-support for hook "after_prepare": plugins/cordova-plugin-add-swift-support/src/add-swift-support.js
Cannot find module 'xcode'
Require stack:
- /home/kinow/Development/javascript/workspace/dearscholar/Android26/plugins/cordova-plugin-add-swift-support/src/add-swift-support.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/node_modules/cordova-lib/src/hooks/HooksRunner.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/node_modules/cordova-lib/src/plugman/install.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/node_modules/cordova-lib/src/plugman/plugman.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/node_modules/cordova-lib/cordova-lib.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/src/help.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/src/cli.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/bin/cordova
Error: Cannot find module 'xcode'
Require stack:
- /home/kinow/Development/javascript/workspace/dearscholar/Android26/plugins/cordova-plugin-add-swift-support/src/add-swift-support.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/node_modules/cordova-lib/src/hooks/HooksRunner.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/node_modules/cordova-lib/src/plugman/install.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/node_modules/cordova-lib/src/plugman/plugman.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/node_modules/cordova-lib/cordova-lib.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/src/help.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/src/cli.js
- /home/kinow/.npm/_npx/16912/lib/node_modules/cordova/bin/cordova
    at Function.Module._resolveFilename (internal/modules/cjs/loader.js:713:15)
    at Function.Module._load (internal/modules/cjs/loader.js:618:27)
    at Module.require (internal/modules/cjs/loader.js:771:19)
    at require (internal/modules/cjs/helpers.js:68:18)
    at Object.<anonymous> (/home/kinow/Development/javascript/workspace/dearscholar/Android26/plugins/cordova-plugin-add-swift-support/src/add-swift-support.js:20:15)
    at Module._compile (internal/modules/cjs/loader.js:868:30)
    at Object.Module._extensions..js (internal/modules/cjs/loader.js:879:10)
    at Module.load (internal/modules/cjs/loader.js:731:32)
    at Function.Module._load (internal/modules/cjs/loader.js:644:12)
    at Module.require (internal/modules/cjs/loader.js:771:19)
```

Would that `xcode` refer to Xcode as in the Apple Xcode? :confused:
