### 2020-07-28 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

Thanks for super fast testing!

Euhm.. I have not seen this error before (likely because I used the phonegap compiler online?). Yes, it has to do with Apple `Xcode` and `cordova-plugin-add-swift-support`.

I found some resources online. Will try to replicate your error tonight, and find out whether the solution proposed in Discussion 1 does the trick.

[Discussion 1](https://forum.ionicframework.com/t/error-cannot-find-module-xcode-but-im-building-for-andorid/161736/2
) --> `npm install --save xcode`

[Discussion 2](https://github.com/cyph/cordova-plugin-chooser/issues/13
)

[Discussion 3](https://github.com/akofman/cordova-plugin-add-swift-support/issues/61)

[Discussion 4](https://www.outsystems.com/forums/discussion/54428/error-cannot-find-module-xcode/)

Thanks for your patience and effort.
