### 2020-07-28 pmkruyen/dearscholar#18 @kinow
Thanks @pmkruyen , calling it a day now. But can test it again tomorrow. I forgot to say what's my environment, sorry.

I am on Ubuntu 18.04.4 LTS :+1: just in case that's helpful. Cheers
