### 2020-07-28 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno, good morning ... :), I received the error too. To solve this, I did:

`npm install --save xcode`

`sudo cordova platform rm android`

`cordova platform add android`

(Note, I used `cordova-android@^8.0.0`)
