### 2020-07-28 pmkruyen/dearscholar#18 @pmkruyen
Can you then test the app now with `cordova run android`? I cannot install `Android SDK` at the moment due to limited space on my MacBook, but if you receive further errors, I will make some room and go for it.

Best,
Peter
