### 2020-07-29 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

Thanks for trying and the docs. I am going to give it try as well tonight to fix the `Broken AVD system path` error. Almost there!

Peter

What I note is that you are in the `Android26` folder (for Android 8.0 and 8.5), not in the `App` folder (for higher Android versions). I do not think this causes the issue, but is something to take into consideration when testing.
