### 2020-07-29 pmkruyen/dearscholar#18 @kinow
>What I note is that you are in the Android26 folder (for Android 8.0 and 8.5), not in the App folder (for higher Android versions). I do not think this causes the issue, but is something to take into consideration when testing.

Same issue. Logged in again to try that as I thought that was a really good guess! But the problem must be somewhere else. Almost!
