### 2020-07-29 pmkruyen/dearscholar#18 @kinow
@pmkruyen found a post after I sent this message, saying for the SDK I needed now to have a different directory structure.

![image](https://user-images.githubusercontent.com/304786/88803647-e5798200-d200-11ea-9b8d-560ac994ff52.png)

So moving the tools under a newly created `cmdline-tools` directory, and re-doing the steps, everything appears to work.

![image](https://user-images.githubusercontent.com/304786/88803733-080b9b00-d201-11ea-90ed-0d9bf80351a5.png)

The UI is blank, but I assume that's because I don't have the backend running.

Next step will be to create a container with PHP + Apache to run these scripts, start them mapping a port like `8000` or `8080`, and then finally find out how to tell the UI to point to the local backend.

:crossed_fingers:

I hope you didn't start downloading the SDK (or freeing up space for that).

Thanks
Bruno
