### 2020-07-29 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

👍  super cool! Will try to replicate your steps.

I have downloaded the command line tools too; but cannot see the other files. I assume that you generated those other files using the steps in the tutorial you posted above?

I am glad the logo is there, but the empty screen worries me somewhat, seems that the app does not pass the splash screen?
Even without the backend running, the login screen should appear:

We will find out :)
