### 2020-07-30 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

**Android emulator**
Sorry, it did not succeed in getting the Android version running on a the emulator. I am halfway the tutorial.

**Browser**
Yes, I have noticed this behavior too (see the readme file) caused by `cordova-plugin-fingerprint-aio`. Can you check if this module has been added to the `plugins folder `(browser --> platform_www --> plugins).

Have you installed / use the PhoneGap desk emulator as specified in the installation instructions? When I pull the project from Github, and (re)open the app folder in this emulator, I did the following:

> Note. In your browser, the login screen should show the login button with the label 'Log in met PIN'. When the login button has the label 'Log in', the cordova-plugin-fingerprint-aio has not loaded correctly. Also, in this case, the browser console displays the error "Fingerprint is not defined". In this case, you cannot login to DearScholar. To solve this issue, grab a cup of coffee, wait a few minutes, and 'Force Reload' the browser window.

Does this work for you too?
