### 2020-07-30 pmkruyen/dearscholar#18 @kinow
I'm having problems now to start the Android emulator. Removed the `browser` platform, but no luck. But I did find some posts, that suggest this could be the problem??

https://github.com/pmkruyen/dearscholar/blob/a6d9efcf6fb5cc00754d75e7fa685ceb402f0dcd/App/config.xml#L31

It appears to define security permissions only for that domain @pmkruyen ? Could that be the problem?

Bruno
