### 2020-07-30 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

**Android**

I set the restrictions because of iOS restrictions, but I don't think that should be the problem because you would still see the login screen. For testing, you can remove it.

**Browser**
Let me know whether my suggestions above solved the loading issues.

Thanks for your help.
Peter
