### 2020-07-30 pmkruyen/dearscholar#18 @kinow
Used

```
<meta http-equiv="Content-Security-Policy" content="default-src data: *; img-src 'self' data:; style-src 'self' 'unsafe-inline'; script-src * 'unsafe-inline' 'unsafe-eval'">
```

And it loaded most resources. The fingerprint plug-in wouldn't load, but that's not a huge issue. The push plug-in also was giving 404, so used `<script type="text/javascript" src="plugins/phonegap-plugin-push/www/push.js"></script>` and it appears to have been loaded fine.

Alas the browser still display a blank page. Also could't load in the emulator, same blank page :disappointed:
