### 2020-07-30 pmkruyen/dearscholar#18 @pmkruyen
Sorry, we are going to solve this.

What I did:

1.  I cloned the project to my MacBook.
2. I opened the app with PhoneGap desktop --> I see the login page with 'Log in'.
`http://192.168.2.15:3000/
`3. I navigated to the App directory and ran:
`cordova platform add browser`

4. The terminal shows
```
Adding browser project...
Creating Cordova project for cordova-browser:
	Path: /Users/peterkruyen/surfdrive/Sandra/app/git/dearscholarx/App/platforms/browser
	Name: DearScholar
Installing "cordova-plugin-add-swift-support" for browser
Installing "cordova-plugin-device" for browser
Installing "cordova-plugin-dialogs" for browser
Installing "cordova-plugin-fingerprint-aio" for browser
Installing "cordova-plugin-ionic-keyboard" for browser
Installing "cordova-plugin-ionic-webview" for browser
Installing "cordova-plugin-splashscreen" for browser
Installing "cordova-plugin-statusbar" for browser
Installing "cordova-sqlite-evcore-extbuild-free" for browser
Installing "cordova-support-google-services" for browser
Installing "phonegap-plugin-multidex" for browser
Installing "phonegap-plugin-push" for browser
Cannot find module 'xcode'
```
5. I went back to `http://192.168.2.15:3000/ ` and I see: "log in met PIN", and using the earlier created Docker, I can log in. (So even without the 'xcode' module.)

===
Did you try to open the app at the PhoneGap port `http://192.168.2.15:3000/ ` (when you open the index page, mine is at `file:///Users/peterkruyen/surfdrive/Sandra/app/git/dearscholarx/App/platforms/browser/www/index.html` it does not work either)?
===

When this does not work either, I will install a docker ubuntu container and run above steps too, to find out how to solve this issue.
