### 2020-07-30 pmkruyen/dearscholar#18 @pmkruyen
My version of the PhoneGap emulator is v0.4.5 :)
