### 2020-07-30 pmkruyen/dearscholar#18 @pmkruyen
Optionally, when there is still the blank screen, you can remove all plugins and check what happens? Also, what happens when you try to run the [hello world app](http://docs.phonegap.com/getting-started/3-create-your-app/desktop/)?

By the way, I realise, sometimes a blank screen occurs when the javascript is broken. But that should not be the case now, because I would not be able to start the app too.

Peter
