### 2020-07-30 pmkruyen/dearscholar#18 @pmkruyen
In particular, can you find out what happens when you remove the `phonegap-plugin-push` plugin?
