### 2020-07-31 pmkruyen/dearscholar#18 @kinow
>I opened the app with PhoneGap desktop --> I see the login page with 'Log in' on port:

Was going to install PhoneGap Desktop, but I think that doesn't have a Linux version.

>PhoneGap Desktop is currently available for Mac and Windows. - http://docs.phonegap.com/references/desktop-app/

>My version of the PhoneGap emulator is v0.4.5 :)

```bash
kinow@ranma:~/Development/javascript/workspace/dearscholar/App$ npx cordova -v
npx: installed 432 in 7.525s
9.0.0 (cordova-lib@9.0.1)
kinow@ranma:~/Development/javascript/workspace/dearscholar/App$ npx phonegap -v
npx: installed 776 in 10.845s
9.0.0
```

>PS. Thanks for adding the push.js file. I excluded it because it includes information about the push notification settings (see readme), better indeed to include an empty mock file.

Sounds like a good idea :+1:

>When this does not work either, I will install a docker ubuntu container and run above steps too, to find out how to solve this issue.

That might require using an X server with Docker, which is not so straightforward. I've only used [one image](https://github.com/alanbchristie/cylc-docker#running-with-the-gui-osx) that worked with GUI.
