### 2020-07-31 pmkruyen/dearscholar#18 @kinow
>Did you try to open the app at the PhoneGap port http://192.168.2.15:3000/ (when you open the index page directly-- at .../git/dearscholar/App/platforms/browser/www/index.html--I see only the 'log in' button, and an error that the fingerprint is not defined)?

When I run `npx cordova run browser`, it launches the browser for me with the URL http://localhost:8000/index.html. It contains the android emulator favicon. The source code is also the same as `www/index.html`, but served from another directory with the files produced after `npx cordova build` (I checked the source code yesterday after I changed it a bit).

Nothing listening on port 3000 as far as I can tell.

```bash
netkinow@ranma:~/Development/javascript/workspace/dearscholar/App$ netstat -tlnpgrep 3000
(Not all processes could be identified, non-owned process info
 will not be shown, you would have to be root to see it all.)

```
