### 2020-07-31 pmkruyen/dearscholar#18 @kinow
Plugins installed for me:

```bash
kinow@ranma:~/Development/javascript/workspace/dearscholar/App$ npx cordova plugins ls
npx: installed 432 in 7.564s
cordova-plugin-add-swift-support 2.0.2 "AddSwiftSupport"
cordova-plugin-device 2.0.3 "Device"
cordova-plugin-dialogs 2.0.2 "Notification"
cordova-plugin-fingerprint-aio 3.0.1 "FingerprintAllInOne"
cordova-plugin-ionic-keyboard 2.2.0 "cordova-plugin-ionic-keyboard"
cordova-plugin-ionic-webview 5.0.0 "cordova-plugin-ionic-webview"
cordova-plugin-splashscreen 5.0.3 "Splashscreen"
cordova-plugin-statusbar 2.4.3 "StatusBar"
cordova-sqlite-evcore-extbuild-free 0.11.0 "Cordova sqlite storage - free enterprise version with Android performance/memory improvements and extra features for PhoneGap Build"
cordova-support-google-services 1.3.2 "cordova-support-google-services"
phonegap-plugin-multidex 1.0.0 "Multidex"
phonegap-plugin-push 2.3.0 "PushPlugin"
```

>In particular, can you find out what happens when you remove the phonegap-plugin-push plugin?

```bash
kinow@ranma:~/Development/javascript/workspace/dearscholar/App$ npx cordova plugins rm phonegap-plugin-push
npx: installed 432 in 8.919s
Uninstalling 2 dependent plugins.
Uninstalling cordova-support-google-services from android
Subproject Path: CordovaLib
Subproject Path: app
Uninstalling phonegap-plugin-multidex from android
Subproject Path: CordovaLib
Subproject Path: app
Uninstalling phonegap-plugin-push from android
Subproject Path: CordovaLib
Subproject Path: app
Uninstalling phonegap-plugin-push from browser
js-module uninstall called : plugins/phonegap-plugin-push/www/push.js
js-module uninstall called : plugins/phonegap-plugin-push/www/browser/push.js
Removing "phonegap-plugin-push"
Removing plugin phonegap-plugin-push from config.xml file...
Removing phonegap-plugin-push from package.json
kinow@ranma:~/Development/javascript/workspace/dearscholar/App$ npx cordova plugins ls
npx: installed 432 in 7.439s
cordova-plugin-add-swift-support 2.0.2 "AddSwiftSupport"
cordova-plugin-device 2.0.3 "Device"
cordova-plugin-dialogs 2.0.2 "Notification"
cordova-plugin-fingerprint-aio 3.0.1 "FingerprintAllInOne"
cordova-plugin-ionic-keyboard 2.2.0 "cordova-plugin-ionic-keyboard"
cordova-plugin-ionic-webview 5.0.0 "cordova-plugin-ionic-webview"
cordova-plugin-splashscreen 5.0.3 "Splashscreen"
cordova-plugin-statusbar 2.4.3 "StatusBar"
cordova-sqlite-evcore-extbuild-free 0.11.0 "Cordova sqlite storage - free enterprise version with Android performance/memory improvements and extra features for PhoneGap Build"
cordova-support-google-services 1.3.2 "cordova-support-google-services"
phonegap-plugin-multidex 1.0.0 "Multidex"
```

Followed by `npx cordova run build`:

```
kinow@ranma:~/Development/javascript/workspace/dearscholar/App$ npx cordova build
npx: installed 432 in 7.508s
Checking Java JDK and Android SDK versions
ANDROID_SDK_ROOT=/home/kinow/Android/Sdk (recommended setting)
ANDROID_HOME=/home/kinow/Android/Sdk (DEPRECATED)
Failed to instantiate ProjectBuilder builder: Error: Cannot find module 'compare-func'
Require stack:
- /home/kinow/Development/javascript/workspace/dearscholar/App/platforms/android/cordova/lib/builders/ProjectBuilder.js
- /home/kinow/Development/javascript/workspace/dearscholar/App/platforms/android/cordova/lib/builders/builders.js
- /home/kinow/Development/javascript/workspace/dearscholar/App/platforms/android/cordova/lib/build.js
- /home/kinow/Development/javascript/workspace/dearscholar/App/platforms/android/cordova/Api.js
- /home/kinow/.npm/_npx/11622/lib/node_modules/cordova/node_modules/cordova-lib/src/cordova/util.js
- /home/kinow/.npm/_npx/11622/lib/node_modules/cordova/node_modules/cordova-lib/src/platforms/platforms.js
- /home/kinow/.npm/_npx/11622/lib/node_modules/cordova/node_modules/cordova-lib/src/plugman/install.js
- /home/kinow/.npm/_npx/11622/lib/node_modules/cordova/node_modules/cordova-lib/src/plugman/plugman.js
- /home/kinow/.npm/_npx/11622/lib/node_modules/cordova/node_modules/cordova-lib/cordova-lib.js
- /home/kinow/.npm/_npx/11622/lib/node_modules/cordova/src/help.js
- /home/kinow/.npm/_npx/11622/lib/node_modules/cordova/src/cli.js
- /home/kinow/.npm/_npx/11622/lib/node_modules/cordova/bin/cordova
```
