### 2020-07-31 pmkruyen/dearscholar#18 @kinow
>Optionally, when there is still the blank screen, you can remove all plugins and check what happens? Also, what happens when you try to run the hello world app?

From the error above I think at least the push plugin needs to stay? I think there might be something wrong with my environment. Maybe I'm missing another step required to set up cordova or android sdk. Or maybe I need to modify `config.xml` to match something in my environment :thinking:
