### 2020-07-31 pmkruyen/dearscholar#18 @kinow
Peter,

I fixed a few browser console errors via #24. Then I fixed the last one that was the location of the `push.js` (but this is not in that PR, as you said there was another solution, an empty file I think?).

Anyhow, I found this error in the console then.

```
GEThttp://localhost:8000/plugins/cordova-plugin-fingerprint-aio/www/Fingerprint.js
Blocked By Extension
```

![image](https://user-images.githubusercontent.com/304786/88999971-3cd73980-d34a-11ea-8e2f-8c48479ad9a2.png)

The part about an extension was interesting, so I disaled uBlock origin ad blocker in Firefox.

![image](https://user-images.githubusercontent.com/304786/89000028-62644300-d34a-11ea-8040-4abb7d3ee73a.png)

So good progress this week :-) Friday afternoon now here in New Zealand. If I have some spare time over the weekend I will see if I can try to use the backend too.

At the moment there's another message in the browser console, regarding some openDatabase function? Filling in random user + pass, and clicking Log in, nothing happens. No network requests, and no browser console messages.

Thanks for the help!
Bruno
