### 2020-07-31 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

Thanks for all effort. I appreciate it! I hadn't expect that it would take so much time to set up DearScholar, I am sorry for this.

**Browser**
When you see the "Log in" button, this means that the `cordova-plugin-fingerprint-aio` plugin has not been loaded (initialized). Likely, I suspect the same problem occurs with the the `cordova-sqlite-evcore-extbuild-free` plugin because of the database error.

I suspect these problems are caused by not being able to use the PhoneGap emulator (unavailable for Ubuntu I understand from your comments), and hence--likely--there is nothing at `http://localhost:8000` but at`http://localhost:8000/index.html`. When I navigate to the later address I got the same incorrect "Log in" button, albeit without 'your' errors.  Also, the problem persists when using the rendered browser pages at `.../platforms/browser/www/index.html`. (The other pages won't show nothing useful at all (e.g, `/schedule.html`) because the way Phonegap and `Framework7` work).

Importantly, DearScholar was made for Android and iOS and the PhoneGap emulator makes it possible to test the app in a browser with the plugins working correctly. While it would be cool when all worked correctly when using the rendered browser version by Cordova--people can also use the app online (testing and production)--I am afraid that takes some time I (in particular because of `Framework7`) and prefer not to make it part of this publication (but will definitely have a look at it, maybe it easier to fix than I think).

Meanwhile, I found some resources specifying how to emulate a Cordova app without the emulator (in particular the [PhoneGap CLI](https://www.npmjs.com/package/phonegap) page :). Will give it a try over the weekend. We will succeed!

**Android**
When I succeed with the Browser emulator issue, I will continue with the Android emulator (also because it seems that both issues are connected from what I understand from your comments and the resources online).

**Furthermore**

- Thanks for your fixes, will review them and approve!

- Yesterday I also fixed a small bug in the iOS version caused by deviant behavior of websql in iOS.

- Yes, the 'Log in' button is not a smart choice for a button that does not work (it is path dependency here, the login functions cause a lot of troubles...). Will change the button label that will be shown when there are login issues.

- When the fingerprint plugin is correctly initialized you will get an error message when you insert an incorrect username a/o password.

- Thanks for continuing with the backend testing.

Thanks,
Peter
