### 2020-07-31 7c6748a
fixed bug in icon Module D, iOS
