### 2020-07-31 pmkruyen/dearscholar#18 @kinow
> And have a good weekend! 1st_place_medal

Thanks, have a good one too! :sun_with_face:
