### 2020-07-31 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

**Browser**

I have found an alternative way:

Cloned the latest version on my laptop, check in `dearscholar.js `if `var locationPhP = "http://localhost/dearscholar.php"`;

1. ` cordova platform add browser `
2. `cordova emulate `
3. `http://localhost:8000/index.html` --> static page with 'Log in' button, nothing works (but no errors thanks to you).
4. `http://localhost:8000/` --> the app is served with the 'Log in met PIN' button.

I found an additional tutorial [here](https://hostadvice.com/how-to/how-to-install-apache-cordova-on-ubuntu-18-04-lts/) for Ubuntu. Maybe there some helpful tips for you?

In an issue on StackOverflow [here](https://stackoverflow.com/questions/37101783/simulate-cordova-plugins-in-browser) it is suggested to use a `cordova serve` but that results in an error for the iOS build.

In another issue on StackOverflow [here](https://stackoverflow.com/questions/29730482/cordova-run-browser-not-working-ubuntu-14-04) some tips are given to run the browser in Chrome with some small adjustments.

In another issue [here](https://stackoverflow.com/questions/50781544/alternative-ways-to-test-simulate-android-cordova-app-in-browser) it is suggested to use [cordova-simulate](https://www.npmjs.com/package/cordova-simulate).

Hope that these resources will help.

Thanks and enjoy your weekend.
