### 2020-07-31 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

See my updated previous comment. I succeeded after first deleting the platform folders (after having run the `cordova serve` command first).

Best,
Peter
