### 2020-07-31 pmkruyen/dearscholar#18 @kinow
Hi Peter,

I can see the Log in met PIN. And entering a four digit random pin, it appears to POST to `peterkruyen.net`. So the application appears to work using that option?

Does that mean that now I can change the code to point to localhost, start the backend, and log in offline (i.e. without needing user/pass or pin from peterkruyen.net)?

If so I will leave the mobile app as-is, and will move on to the backend (where I may bother you more regarding the DB tables :-) )

Thanks a lot for the links, going through them now :coffee:
Bruno
