### 2020-07-31 pmkruyen/dearscholar#18 @kinow
>I found an additional tutorial here for Ubuntu. Maybe there some helpful tips for you?

Very similar to the way I did. But I used `yarn global add cordova`, which is equivalent to `npm install --global cordova` (in the past `yarn` shared similar cmdline, but later changed to this new parameters)

>In an issue on StackOverflow here it is suggested to use a cordova serve but that results in an error for the iOS build.

Error for me as well.

```
Using cordova-fetch for cordova-ios@^5.1.1
Adding ios project...
Creating Cordova project for the iOS platform:
	Path: platforms/ios
	Package: net.peterkruyen.diary
	Name: DearScholar
iOS project created with cordova-ios@5.1.1
"framework" tag with type "podspec" is deprecated and will be removed. Please use the "podspec" tag.
"framework" tag with type "podspec" is deprecated and will be removed. Please use the "podspec" tag.
doc.find is not a function
```

>In another issue on StackOverflow here some tips are given to run the browser in Chrome with some small adjustments.

The code snippet appears to be for an older version of `cordova`. I cannot see any `switch` statement, or even an `if` statement where I could use the case specified.

I can, however, launch Google Chrome with the same command arguments suggested `google-chrome --test-type --disable-web-security --user-data-dir=/tmp/dearscholar`. Navigating to `http://localhost:800` shows me the same Log in met PIN, as with Firefox. But good to learn this trick too. Might be helpful to launch the browser like this to disable extensions (it would have solved my issue with uBlock from yesterday).

>In another issue here it is suggested to use cordova-simulate.

Thought it would be a repository from some developer, and wasn't sure if this tool would be well maintained. Turns out it is under the GitHub Microsoft organization. Looks useful, but if we can use the backend with the current `cordova` tool, we won't need it for now :crossed_fingers: (but good to know!).

Thanks!
Bruno
