### 2020-07-31 pmkruyen/dearscholar#18 @kinow
Oh, looks like I can direct it to my `localhost` server. We could add the following to building instructions?

1. Edit `App/www/js/dearscholar.js` replacing `locationPhP` with the server URL
2. Create a MySQL database (with Docker, for example)
  2.1 Start server, `docker run --rm --name dearscholar-mysql -e MYSQL_USER=dearscholar -e MYSQL_PASSWORD=dearscholar -e MYSQL_ROOT_PASSWORD=my-secret -e MYSQL_DATABASE=dearscholar  -d mysql:8`
  2.2 Load the MySQL schema `docker exec -i dearscholar-mysql sh -c 'exec mysql -uroot -p"my-secret"' < mysql`

I was going to continue with the rest of commands to set up the PHP side too, but looks like the `mysql` script needs some tweaking to support MySQL 8? Is there a specific version of MySQL that is necessary?

```
kinow@ranma:~/Development/javascript/workspace/dearscholar$ docker exec -i dearscholar-mysql sh -c 'exec mysql -uroot -p"my-secret"' < mysql
mysql: [Warning] Using a password on the command line interface can be insecure.
ERROR 1064 (42000) at line 136: You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near 'NULL' at line 1

```
