### 2020-07-31 pmkruyen/dearscholar#18 @kinow
Running the instructions from readme, the MySQL container appears to get stuck and fail to start?

```bash
kinow@ranma:~/Development/javascript/workspace/dearscholar$ docker run --rm --name dearscholar -it -p "9000:9000" -v ${PWD}/App:/app -v ${PWD}/mysql:/var/lib/mysql mattrayner/lamp:latest-1804
Updating for PHP 7.4
Replacing CLI php.ini values
Editing APACHE_RUN_GROUP environment variable
Editing phpmyadmin config
Setting up MySQL directories
Allowing Apache/PHP to write to the app
Allowing Apache/PHP to write to MySQL
Editing MySQL config
=> An empty or uninitialized MySQL volume is detected in /var/lib/mysql
=> Installing MySQL ...
mysqld: Can't create directory '/var/lib/mysql/' (Errcode: 17 - File exists)
2020-07-31T22:14:57.043288Z 0 [Warning] TIMESTAMP with implicit DEFAULT value is deprecated. Please use --explicit_defaults_for_timestamp server option (see documentation for more details).
2020-07-31T22:14:57.045058Z 0 [ERROR] Aborting

=> Done!
=> Waiting for confirmation of MySQL service startup
=> Waiting for confirmation of MySQL service startup

```
