### 2020-08-01 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

Thanks. We are going in the right direction. 👍 A quick reply:

**Backend**
I have not encountered that mysql error before. Could it be caused by specifying the username and password in the command line? I did not specify one of them, the docker created a password for the admin (or root) user. I will replicate your steps tonight.

**Browser**
Will look at your posts later today, will edit / clarify the instructions where needed. I will to add some instructions for Linux users based on your feedback, which I hope you can verify.

**Android**
I made progress yesterday evening. I will add the iOS instructions later. I can then also add some Linux instructions based on your comments and feedback.

I am now at the point to install and load an emulator. I see that I have to specify the `API level` of the emulator. _Question_: when selecting the emulator, which api level have you specified? It should be >= 28. If you choose a lower API level, the app fails.

(note. the Android26 folder includes the settings that work for api 26 and 27 [it is a small adjustment in de `config.xlm` file).

Best,
Peter
