### 2020-08-01 pmkruyen/dearscholar#18 @pmkruyen
By the way, what happens when you delete the folder `mysqld: Can't create directory '/var/lib/mysql/' (Errcode: 17 - File exists)` and rerun the docker?

I now see that the other issue has been solved?

> EDIT: re-reading readme.md, will follow the Docker instructions that you wrote, that should fix this issue, please ignore last message @pmkruyen :) and thanks for the detailed instructions!

Peter
