### 2020-08-01 pmkruyen/dearscholar#18 @kinow
>Thanks. We are going in the right direction. +1 A quick reply:

Agree! Good progress, we might close this issue very soon :crossed_fingers:

>I am now at the point to install and load an emulator. I see that I have to specify the API level of the emulator. Question: when selecting the emulator, which api level have you specified? It should be >= 28. If you choose a lower API level, the app fails.

I used 28. Initially I was trying 26, in the Android26 folder. Then after you mentioned that was only for <=26, I switched to App folder with the 28 version.

I also chose 28, because the first time I ran the `npx cordova build` command in the App folder, it found that I had the Android 26 installed, but instead said it needed 28, and downloaded it.

>By the way, what happens when you delete the folder mysqld: Can't create directory '/var/lib/mysql/' (Errcode: 17 - File exists) and rerun the docker?

My understanding is that that is not a folder on my computer,

```
kinow@ranma:~/Development/java/apache/commons-imaging$ file /var/lib/mysql
/var/lib/mysql: cannot open `/var/lib/mysql' (No such file or directory)
```

, but instead a folder in the container. And after reading the container README file, I **think** that we are mapping the file `mysql` from this repository onto `/var/lib/mysql`... which I think was supposed to be a folder?

>I now see that the other issue has been solved?

Yup! I was trying to run a separate MySQL container. I didn't remember you had written new instructions with Docker :-)
