### 2020-08-01 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno

**Android**
While I cannot log in yet (need to set the correct path to the local `Dearscholar.php` file), ... I managed to run the Android emulator... --> pass the splash screen.

I manually deleted all build files, including those in `the node folder`. Next, I ran `cordova platform add android `. Got a warning about several missing modules and had to run `npm i cordova-common`. And on `cordova run android`...

![login screen](https://user-images.githubusercontent.com/20207147/89105183-3666c180-d41f-11ea-953e-b33ed31f1cbe.gif)

When I the app works as expected on Android, I will update the installation instructions.  Hope you can verify these instructions (for Linux).

PS. I got this warning now  `npm install cordova-android@9.0.0  to resolve 1 vulnerability`... will do so, hope that does not have an impact on the performance of the app. I will check whether it is also wise to upgrade ios to this version. Will let you know.

**Browser**
Great that it works! Will update the installation instructions. Hope you can verify these instructions these written (for Linux).

**Backend**
With regard to the folder, that is interesting, mine is not on the container but at `/users/peterkruyen/mysql/` (so not at `/var/lib/mysql`?

Best,
Peter
