### 2020-08-01 d767a57
fix Android emulator - step 1
