+++ Android26/plugins/cordova-sqlite-evcore-extbuild-free/spec/www/spec/db-tx-string-test.js
@@ -1,2815 +1,0 @@
-/* 'use strict'; */
-
-var MYTIMEOUT = 12000;
-
-// NOTE: DEFAULT_SIZE wanted depends on type of browser
-
-var isWindows = /MSAppHost/.test(navigator.userAgent);
-var isAndroid = !isWindows && /Android/.test(navigator.userAgent);
-var isFirefox = /Firefox/.test(navigator.userAgent);
-var isWebKitBrowser = !isWindows && !isAndroid && /Safari/.test(navigator.userAgent);
-var isBrowser = isWebKitBrowser || isFirefox;
-var isEdgeBrowser = isBrowser && (/Edge/.test(navigator.userAgent));
-var isChromeBrowser = isBrowser && !isEdgeBrowser && (/Chrome/.test(navigator.userAgent));
-var isSafariBrowser = isWebKitBrowser && !isEdgeBrowser && !isChromeBrowser;
-var isMac = !isBrowser && /Macintosh/.test(navigator.userAgent);
-var isAppleMobileOS = /iPhone/.test(navigator.userAgent) ||
-      /iPad/.test(navigator.userAgent) || /iPod/.test(navigator.userAgent);
-
-// should avoid popups (Safari seems to count 2x)
-var DEFAULT_SIZE = isSafariBrowser ? 2000000 : 5000000;
-// FUTURE TBD: 50MB should be OK on Chrome and some other test browsers.
-
-// NOTE: While in certain version branches there is no difference between
-// the default Android implementation and implementation #2,
-// this test script will also apply the androidLockWorkaround: 1 option
-// in case of implementation #2.
-// The following openDatabase settings are used for Plugin-implementation-2
-// on Android:
-// - androidDatabaseImplementation: 2
-// - androidLockWorkaround: 1
-var scenarioList = [
-  isAndroid ? 'Plugin-implementation-default' : 'Plugin',
-  'HTML5',
-  'Plugin-implementation-2'
-];
-
-var scenarioCount = (!!window.hasWebKitWebSQL) ? (isAndroid ? 3 : 2) : 1;
-
-var mytests = function() {
-
-  for (var i=0; i<scenarioCount; ++i) {
-    // TBD skip plugin test on browser platform (not yet supported):
-    if (isBrowser && (i === 0)) continue;
-
-    describe(scenarioList[i] + ': tx string test(s)', function() {
-      var scenarioName = scenarioList[i];
-      var suiteName = scenarioName + ': ';
-      var isWebSql = (i === 1);
-      var isImpl2 = (i === 2);
-      // TBD WORKAROUND SOLUTION for (WebKit) Web SQL on Safari browser:
-      var recycleWebDatabase = null;
-
-      // NOTE 1: MUST be defined in proper describe function scope, NOT outer scope.
-      // NOTE 2: Using same database name in this script to avoid issue with
-      //         "Too many open files" on iOS with WKWebView engine plugin.
-      //         (FUTURE TBD NEEDS INVESTIGATION)
-      var openDatabase = function(name_ignored, ignored1, ignored2, ignored3) {
-        var name = 'string-test.db';
-        if (isWebSql && isSafariBrowser && !!recycleWebDatabase)
-          return recycleWebDatabase;
-        if (isImpl2) {
-          return window.sqlitePlugin.openDatabase({
-            // prevent reuse of database from default db implementation:
-            name: 'i2-'+name,
-            // explicit database location:
-            location: 'default',
-            androidDatabaseImplementation: 2,
-            androidLockWorkaround: 1
-          });
-        }
-        if (isWebSql) {
-          return recycleWebDatabase =
-            window.openDatabase(name, '1.0', 'Test', DEFAULT_SIZE);
-        } else {
[…]
