+++ App/plugins/phonegap-plugin-push/src/android/com/adobe/phonegap/push/PushPlugin.java
@@ -1,657 +1,0 @@
-package com.adobe.phonegap.push;
-
-import android.annotation.TargetApi;
-import android.app.Activity;
-import android.app.NotificationChannel;
-import android.app.NotificationManager;
-import android.content.ContentResolver;
-import android.content.Context;
-import android.content.SharedPreferences;
-import android.content.res.Resources;
-import android.media.AudioAttributes;
-import android.net.Uri;
-import android.os.Build;
-import android.os.Bundle;
-import android.support.v4.app.NotificationCompat;
-import android.support.v4.app.NotificationManagerCompat;
-import android.util.Log;
-
-import com.google.firebase.iid.FirebaseInstanceId;
-import com.google.firebase.messaging.FirebaseMessaging;
-
-import org.apache.cordova.CallbackContext;
-import org.apache.cordova.CordovaInterface;
-import org.apache.cordova.CordovaPlugin;
-import org.apache.cordova.CordovaWebView;
-import org.apache.cordova.PluginResult;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.io.IOException;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.ArrayList;
-import java.util.List;
-
-import me.leolin.shortcutbadger.ShortcutBadger;
-
-public class PushPlugin extends CordovaPlugin implements PushConstants {
-
-  public static final String LOG_TAG = "Push_Plugin";
-
-  private static CallbackContext pushContext;
-  private static CordovaWebView gWebView;
-  private static List<Bundle> gCachedExtras = Collections.synchronizedList(new ArrayList<Bundle>());
-  private static boolean gForeground = false;
-
-  private static String registration_id = "";
-
-  /**
-   * Gets the application context from cordova's main activity.
-   *
-   * @return the application context
-   */
-  private Context getApplicationContext() {
-    return this.cordova.getActivity().getApplicationContext();
-  }
-
-  @TargetApi(26)
-  private JSONArray listChannels() throws JSONException {
-    JSONArray channels = new JSONArray();
-    // only call on Android O and above
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-      final NotificationManager notificationManager = (NotificationManager) cordova.getActivity()
-          .getSystemService(Context.NOTIFICATION_SERVICE);
-      List<NotificationChannel> notificationChannels = notificationManager.getNotificationChannels();
-      for (NotificationChannel notificationChannel : notificationChannels) {
-        JSONObject channel = new JSONObject();
-        channel.put(CHANNEL_ID, notificationChannel.getId());
-        channel.put(CHANNEL_DESCRIPTION, notificationChannel.getDescription());
-        channels.put(channel);
-      }
-    }
-    return channels;
-  }
-
-  @TargetApi(26)
-  private void deleteChannel(String channelId) {
-    // only call on Android O and above
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-      final NotificationManager notificationManager = (NotificationManager) cordova.getActivity()
-          .getSystemService(Context.NOTIFICATION_SERVICE);
-      notificationManager.deleteNotificationChannel(channelId);
-    }
-  }
-
-  @TargetApi(26)
-  private void createChannel(JSONObject channel) throws JSONException {
-    // only call on Android O and above
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-      final NotificationManager notificationManager = (NotificationManager) cordova.getActivity()
-          .getSystemService(Context.NOTIFICATION_SERVICE);
-
[…]

[…]
