+++ readme.md
@@ -50,7 +50,6 @@ In this second part of this text, DearScholar's current features and technical d
 * Current features;
 * Framework;
 * Options and settings;
-* Guide for potential contributors and academic researchers;
 * Installation;
 * File structure of the app;
 * Structure of dearscholar.js, the heart of the app.
@@ -99,7 +98,7 @@ DearScholar can be used with Google's Firebase to get the push notifications wor
 DearScholar has the capacity to send short in-app messages to specific respondents. For example, respondents can be sent thank-you messages to show engagement, small encouragements when respondents have missed a measurement occasions, ask follow-up questions, or invited respondents to elaborate on particular answers over the phone or through email. These in-app messages are displayed on a separate message screen. An envelope icon appears in the app’s status bar when new messages have been sent. When respondents have read the message, the researcher is noted.
 
 ## Framework
-DearScholar has been built using [Apache Cordova / Phonegap](https://phonegap.com/), and [Framework7 (v5.4.1)](https://framework7.io) by Vladimir Kharlampidi and his team. Next to several general Cordova plugins, DearScholar uses [fingerprint-aio](https://github.com/NiklasMerz/cordova-plugin-fingerprint-aio) to enable Touch ID and Face Recognition, [sqlite-evcore-extbuild-free](https://www.npmjs.com/package/cordova-sqlite-evcore-extbuild-free) to use sqlite tables, and [cordova-plugin-ionic-webview](https://github.com/ionic-team/cordova-plugin-ionic-webview) to replace UIWebView with WKWebView. 
+DearScholar has been built using [Apache Cordova](https://cordova.apache.org/), and [Framework7 (v5.4.1)](https://framework7.io) by Vladimir Kharlampidi and his team. Next to several general Cordova plugins, DearScholar uses [fingerprint-aio](https://github.com/NiklasMerz/cordova-plugin-fingerprint-aio) to enable Touch ID and Face Recognition, [sqlite-evcore-extbuild-free](https://www.npmjs.com/package/cordova-sqlite-evcore-extbuild-free) to use sqlite tables, and [cordova-plugin-ionic-webview](https://github.com/ionic-team/cordova-plugin-ionic-webview) to replace UIWebView with WKWebView. 
 
 ## Options and settings
 In MySQL, a series of tables need to be set up. The tables include variables (in the columns) and specific settings or respondents' data (in the rows). Below, these tables with all options, if applicable, are described.
@@ -226,51 +225,14 @@ Lastly, a series of empty response tables in which respondents' data is stored, 
 
 *Note*. Why store both "timestamp" and "surveydate"? Respondents are asked to fill out a question module on a specific measurement occasion or survey date (date and time) as displayed in the measurement schedule. To check if respondents indeed completed the question module on the required survey date (or somewhat later), the timestamp is useful.
 
-## Guide for potential contributors and academic researchers
-Scholars who want to use DearScholar in their research project are advised to contact the main author directly. Contributors and academic researchers who want to build their own version of DearScholar can follow the following four steps.
-
-### Step 1: Install DearScholar on your own computer
-To get the developer version of DearScholar running on your computer, the following steps can be taken:
-1) download the app files and folders as specified below (under file structure);
-2) set up an Apache webserver (for testing you may want to use a local Docker container); 
-3) on this webserver, set up all the required tables and files.
-3) install the desktop app of [Phonegap](https://phonegap.com/getstarted/);
-4) compile DearScholar on your computer using Phonegap ... the app appears in your web browser; and 
-5) you can start improving DearScholar.
-
-#### Sample study
-To set up a sample study, the following scripts can be used:
-1) [samplestudy-part1.php](https://github.com/pmkruyen/dearscholar/blob/master/samplestudy-part1.php) to populate the authentication table;
[…]
