+++ readme.md
@@ -226,7 +226,7 @@ Lastly, a series of empty response tables in which respondents' data is stored, 
 *Note*. Why store both "timestamp" and "surveydate"? Respondents are asked to fill out a question module on a specific measurement occasion or survey date (date and time) as displayed in the measurement schedule. To check if respondents indeed completed the question module on the required survey date (or somewhat later), the timestamp is useful.
 
 ## Installation
-Researchers and contributors who want to test, contribute, or want to build their own version of DearScholar are advised to consecutively: install the server (step 1); set up the MySQL database and PhPMyAdmin panel on the server (step 2-4); set up the communication channel between the server and the app (step 5); run DearScholar in an browser (step 6); test the app in an emulated Android and iOS environment (step 7-8); and finally, build, and test the app on a real device (step 9). Obviously, step 10 would be to submit a clone of DearScholar to the App Store and Google Play Store, but that step is beyond this manual.
+Researchers and contributors who want to test, contribute, or want to build their own version of DearScholar are advised to consecutively: install the server (Step 1); set up the MySQL database and PhPMyAdmin panel on the server (Step 2-4); set up the communication channel between the server and the app (Step 5); run DearScholar in a browser (Step 6); test the app in an emulated Android and iOS environment (Step 7-8); and finally, build, and test the app on a real device (Step 9). Obviously, Step 10 would be to submit a clone of DearScholar to the App Store and Google Play Store, but that step is beyond this manual. However, in Step 10, I share some experiences with the App Store to consider when submitting a research app to the App Store.
 
 *Note* It is my ambition to automate step 1-4. Meanwhile, performing the steps manually gives a good first impression about how DearScholar functions.
 
@@ -286,26 +286,27 @@ To let DearScholar communicate with the MySQL tables, copy the file ```dearschol
 ### Step 6: Run DearScholar in a browser 
 Set ```var locationPhP = "http://localhost/dearscholar.php"``` in ```App/www/js/dearscholar.js```.
 
-Download and install Cordova with ```sudo npm install -g cordova```. Check you have installed version ```9.0.0``` by running ```cordova -v```.
+Download and install Cordova with ```sudo npm install -g cordova```. Check you have installed version ```9.0.0``` by running ```cordova -v```. While you do not necessarily need Cordova for Step 6, you will need it for the following steps.
 
 #### Mac and Window users
-Mac and Windows users can use the [Phonegap desktop app](https://phonegap.com/getstarted/) for testing. Import the ```App folder```  in the Phonegap desktop app. In your browser, navigate to the PhoneGap port shown at the bottom of the PhoneGap screen (e.g., ```http://localhost:3000/```). You now see the login page of DearScholar. Enable the ```inspect``` option in your browser to render the app screen in 'normal' properties for mobile devices. 
+Mac and Windows users can use the [Phonegap desktop app](https://phonegap.com/getstarted/) for testing. Import the ```App folder```  in the Phonegap desktop app. When PhoneGap is done with rendering, navigate to the PhoneGap port in your browser shown at the bottom of the PhoneGap screen (e.g., ```http://localhost:3000/```). You now see the login page of DearScholar. Enable the ```inspect``` option in your browser to render the app screen in 'normal' properties for mobile devices. 
 
 Fill out the respondent credentials inserted in Step 3 of these instructions, click on 'Log in met PIN', and start testing. 
 
 #### Linux users and everybody else who prefers the command line
 Navigate to the ```App folder``` and run the following commands in a terminal window:
 
-```cordova platform add browser```
-```cordova emulate```
+    ```cordova platform add browser```
 
-A browser window pops up serving the app from ```http://localhost:8000/index.html```. That looks nice, but DearScholar does not work. Change the address in ```http://localhost:8000/``` and you are good to go.
+    ```cordova emulate```
+
[…]
