+++ readme.md
@@ -226,7 +226,7 @@ Lastly, a series of empty response tables in which respondents' data is stored, 
 *Note*. Why store both "timestamp" and "surveydate"? Respondents are asked to fill out a question module on a specific measurement occasion or survey date (date and time) as displayed in the measurement schedule. To check if respondents indeed completed the question module on the required survey date (or somewhat later), the timestamp is useful.
 
 ## Installation
-Researchers and contributors who want to test, contribute, or want to build their own version of DearScholar are advised to consecutively: install the server (Step 1); set up the MySQL database and PhPMyAdmin panel on the server (Step 2-4); set up the communication channel between the server and the app (Step 5); run DearScholar in a browser (Step 6); test the app in an emulated Android and iOS environment (Step 7-8); and finally, build, and test the app on a real device (Step 9). Obviously, Step 10 would be to submit a clone of DearScholar to the App Store and Google Play Store, but that step is beyond this manual. However, in Step 10, I share some experiences with the App Store to consider when submitting a research app to the App Store.
+Researchers and contributors who want to test, contribute, or want to build their own version of DearScholar are advised to consecutively: install the server (Step 1); set up the MySQL database and PhPMyAdmin panel on the server (Step 2-4); set up the communication channel between the server and the app (Step 5); run DearScholar in a browser (Step 6); test the app in an emulated Android and iOS environment (Step 7-8); and finally, build, and test the app on a real device (Step 9). Obviously, Step 10 would be to submit a clone of DearScholar to the App Store and Google Play Store, but that step is beyond this manual. However, in Step 10 in these installation instructions, I share some experiences with the App Store to consider when submitting a research app to the App Store.
 
 *Note* It is my ambition to automate step 1-4. Meanwhile, performing the steps manually gives a good first impression about how DearScholar functions.
 
