+++ readme.md
@@ -228,7 +228,7 @@ Lastly, a series of empty response tables in which respondents' data is stored, 
 ## Installation
 Researchers and contributors who want to test, contribute, or want to build their own version of DearScholar are advised to consecutively: install the server (Step 1); set up the MySQL database and PhPMyAdmin panel on the server (Step 2-4); set up the communication channel between the server and the app (Step 5); run DearScholar in a browser (Step 6); test the app in an emulated Android and iOS environment (Step 7-8); and finally, build, and test the app on a real device (Step 9). Obviously, Step 10 would be to submit a clone of DearScholar to the App Store and Google Play Store, but that step is beyond this manual. However, in Step 10 in these installation instructions, I share some experiences with the App Store to consider when submitting a research app to the App Store.
 
-*Note.* It is my ambition to automate step 1-4. Meanwhile, performing the steps manually gives a good first impression about how DearScholar functions.
+*Note*. It is my ambition to automate step 1-4. Meanwhile, performing the steps manually gives a good first impression about how DearScholar functions.
 
 ### Step 0: Clone the DearScholar repository from GitHub
 
@@ -294,20 +294,14 @@ Mac and Windows users can use the [Phonegap desktop app](https://phonegap.com/ge
 Fill out the respondent credentials inserted in Step 3 of these instructions, click on 'Log in met PIN', and start testing. 
 
 #### Linux users and everybody else who prefers the command line
-Navigate to the ```App folder``` and run the following commands in a terminal window:
+Run the following commands in the ```App folder```: ```cordova platform add browser``` followed by ```cordova emulate```.
 
-     ```cordova platform add browser```
-
-     ```cordova emulate```
-
-A browser window pops up serving the app at ```http://localhost:8000/index.html```. That looks nice, but DearScholar does not work. Change the address in ```http://localhost:8000/``` and you are good to go.
-
+A browser window pops up serving the app at ```http://localhost:8000/index.html```. That looks nice, but DearScholar does not work. Change the address to ```http://localhost:8000/``` and you are good to go.
 
 #### Notes
-*Note 1.* When Cordova builds the browser platform you sometimes receive a message about missing modules. Install these modules. Delete the ```platform folder``` which was generated by Cordova in the ```app folder``` by ```cordova platform rm browser``` and rerun ```cordova platform add browser``` and ```cordova emulate```.
+*Note 1*. When Cordova builds the browser platform you sometimes receive a message about missing modules. Install these modules. Delete the ```platform folder``` which was generated by Cordova in the ```app folder``` by ```cordova platform rm browser``` and rerun ```cordova platform add browser``` and ```cordova emulate```.
 
-*Note 2.* In your browser, the login screen should show the login button with the label 'Log in met PIN'. When you see a button with the label 'Error: Restart DearScholar', the cordova-plugin-fingerprint-aio has not loaded correctly. Likely, the debugging console displays the error "Fingerprint is not defined". In this case, you cannot log in to DearScholar. To solve this issue, when using Phonegap desktop app, wait a few minutes and 'Force Reload' the browser window. When using the command line or to speed things up you, delete  the ```platform folder``` which was generated by Cordova in the ```app folder``` by ```cordova platform rm browser``` and rerun ```cordova platform add browser``` and ```cordova emulate```.
-
[…]
