### 2020-08-01 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

**Browser**

Can you verify the new installation instructions for the Browser [Step 6] (https://github.com/pmkruyen/dearscholar#step-6-run-dearscholar-in-a-browser)?

I deleted the old **Guide for contributors** section from the Readme because of the overlap with the new (and better) installation instructions.

I cannot verify / do not completely understand how you used the command `google-chrome --test-type --disable-web-security --user-data-dir=/tmp/dearscholar ` and if/how it should be added. Suggestions are welcome :)

Best,
Peter

PS. The Android instructions will follow.
