### 2020-08-02 pmkruyen/dearscholar#18 @kinow
>Can you verify the new installation instructions for the Browser [Step 6] (https://github.com/pmkruyen/dearscholar#step-6-run-dearscholar-in-a-browser)?

Looks good!

>I deleted the old Guide for contributors section from the Readme because of the overlap with the new (and better) installation instructions.

Good idea. I wonder if the installation instructions would deserve its own page (wiki, INSTALL.md, site page, etc) someday. Just because it seems to be getting quite lengthy (not that it bothers me, on the contrary :)

>I cannot verify / do not completely understand how you used the command google-chrome --test-type --disable-web-security --user-data-dir=/tmp/dearscholar and if/how it should be added. Suggestions are welcome :)

:+1: for now I think the application should work with any browser. If I run into problems using the current version I will try it again and raise a PR.
