### 2020-08-02 pmkruyen/dearscholar#18 @kinow
@pmkruyen

The `readme.md` file has this line for running Docker `docker run -i -t -p "80:80" -v ${PWD}/app:/app -v ${PWD}/mysql:/var/lib/mysql mattrayner/lamp:latest-1804`. I think it can be broken down into:

- `docker`, command
- `run`, subcommand
- `i`, interactive
- `t`, simulate tty/terminal
- `p "80:80"`, to map the container port `80` to the host computer's port `80`
- `-v ${PWD}/app:/app`, to create a volume at `/app` in the container that mounts the contents of the current directory `$PWD` `/app` (the `App` in this project); i.e. `/app` in the container will have the content of `./App`
- `-v ${PWD}/mysql:/var/lib/mysql`, to create a volume at `/var/lib/mysql` with the contents of the file `mysql` from this repository (**!** this is the problem I'm having, `/var/lib/mysql` is normally a directory)
- `mattrayner/lamp:latest-1804` container identifier, with the maintainer's ID, name of container, and the tag

The **!** marks where I think we have a problem right now. If we navigate to the container's documentation at https://github.com/mattrayner/docker-lamp/blob/c3f1244fec229deaed61e94d1229ba831b1da465/README.md, it says:

>The below command will run the docker image mattrayner/lamp:latest, creating a mysql/ folder within your project. This folder will be linked to /var/lib/mysql where all of the MySQL files from container lives. You will now be able to stop/start the container and keep your database changes.

See the part **creating a mysql/ folder within your project**, which is not what we have here.

I think this is the root cause of my current issue. We probably need to find out how we can load the contents of the file `mysql` into a MySQL database in that container. Then I think my browser-platform app will be able to communicate with the backend :crossed_fingers:
