### 2020-08-02 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

> Good idea. I wonder if the installation instructions would deserve its own page (wiki, INSTALL.md, site page, etc) someday. Just because it seems to be getting quite lengthy (not that it bothers me, on the contrary :)

Had the same thought last night. Will do after the whole installation manual does deliver :)

**Browser**
 :white_check_mark:

**Backend**
Thanks for your notes and suggestions about the docker. After I have wrote the instructions for the Android emulator, will start working on solving the issues with setting up the docker.

Best,
Peter

PS. Thank you for your PR!
