### 2020-08-03 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

Great that Browser works! Thanks for your help! That is the most convenient way of testing the app.

**Android**
Almost there... but I cannot connect to the backend (receiving 404 errors), drives me crazy. I will give it another try tonight; or else will upload it to the installation instructions as 'work in progress'.

**Backend**
Meanwhile, let's get the backend running :)

There are three issues--as in, points of departure for getting the backend running 😄 )--I noticed in your comments.

_Issue 1_

I ran

`docker run -i -t -p "80:80" -v ${PWD}/app:/app -v ${PWD}/mysql:/var/lib/mysql mattrayner/lamp:latest-1804`

... and the terminal shows....

`Updating for PHP 7.4
Replacing CLI php.ini values
Editing APACHE_RUN_GROUP environment variable
Editing phpmyadmin config
Setting up MySQL directories
Allowing Apache/PHP to write to the app
Allowing Apache/PHP to write to MySQL
Editing MySQL config`

.... and all works.

Inspecting your code, I see that you changed the ports and you use a capital letter "A" in app. Does that cause the problem?

What do you see when you navigate to `localhost/myphpadmin` in your browser? What happens when you wipe the image completely and re-install / re-initialize it?

_Issue 2_
To be sure: The `app folder` in `docker run` does not refer to the `app folder` in the repo, but to the persistent location of the created webserver by `mattrayner/lamp`. Here is where you need to put the `dearscholar.php` file (see the instructions, step 5, this is a bit confusing I now see). It would be great indeed to include DearScholar in the docker too, but that is something on my bucket list.

_Issue 3_
To be sure: Along similar lines, you cannot directly load the content of my `mysql file` into the MySQL database, but (at the moment) should use the mysql command line in the docker image (see step 1 and 2 of the manual). Can you open the mysql docker in your terminal (see step 1 of the manual).

In short, can you let me know what happens when you re-install the Docker image, preferably also delete the app folder and mysql folder created the image (both folders are at located at `Users/peterkruyen/`) and run `docker run -i -t -p "80:80" -v ${PWD}/app:/app -v ${PWD}/mysql:/var/lib/mysql mattrayner/lamp:latest-1804`.

Thanks,
Peter
