### 2020-08-03 pmkruyen/dearscholar#18 @kinow
> Almost there... but I cannot connect to the backend (receiving 404 errors), drives me crazy. I will give it another try tonight; or else will upload it to the installation instructions as 'work in progress'.

Sounds like a good plan, if we have at least the browser platform working, that should be enough to run the app locally. Later we can work on the Android or iOS platforms, or tell users that they can download the apk or use the store version for the time being.

>Inspecting your code, I see that you changed the ports and you use a capital letter "A" in app. Does that cause the problem?

I don't think so. I tried both ways yesterday. But will try again today to confirm.

Just synced the repo. Then executed:

```
docker system prune -a -f
docker network prune
docker volume prune
```

Then running `docker run -i -t -p "80:80" -v ${PWD}/app:/app -v ${PWD}/mysql:/var/lib/mysql mattrayner/lamp:latest-1804`:

![image](https://user-images.githubusercontent.com/304786/89164863-c21d5100-d5cb-11ea-9b2f-80d4d15f1771.png)

>What do you see when you navigate to localhost/myphpadmin in your browser? What happens when you wipe the image completely and re-install / re-initialize it?

With the container in that state:

![image](https://user-images.githubusercontent.com/304786/89164972-f133c280-d5cb-11ea-8af5-1e367dd55ed9.png)

and

![image](https://user-images.githubusercontent.com/304786/89165100-30621380-d5cc-11ea-86f4-60e6fe041f93.png)

(slight change in address)

>To be sure: The app folder in docker run does not refer to the app folder in the repo, but to the persistent location of the created webserver by mattrayner/lamp. Here is where you need to put the dearscholar.php file (see the instructions, step 5, this is a bit confusing I now see). It would be great indeed to include DearScholar in the docker too, but that is something on my bucket list.

Ah, I see. I thought we were mapping the `App` folder within the container. I am stuck at step 1 at the moment, trying to get pass "Step 1: Server setup". I can skip it, but looks like mysql is not running.

I believe my container has no services listening on any ports at the moment. Its initialization seems to have failed.

![image](https://user-images.githubusercontent.com/304786/89165432-a8c8d480-d5cc-11ea-9f7e-e497fb95d3c9.png)

>To be sure: Along similar lines, you cannot directly load the content of my mysql file into the MySQL database, but (at the moment) should use the mysql command line in the docker image (see step 1 and 2 of the manual). Can you open the mysql docker in your terminal (see step 1 of the manual).

:+1: looks like I was thinking too much ahead. Your instructions would have showed me how to use the `app` and `mysql` folders. Once we manage to fix the initial set up, everything should work :crossed_fingers:

>In short, can you let me know what happens when you re-install the Docker image, preferably also delete the app folder and mysql folder created the image (both folders are at located at Users/peterkruyen/) and run docker run -i -t -p "80:80" -v ${PWD}/app:/app -v ${PWD}/mysql:/var/lib/mysql mattrayner/lamp:latest-1804.

Sure. Removed the `app` directory (looks like it had been created), and re-executed the exact command `docker run -i -t -p "80:80" -v ${PWD}/app:/app -v ${PWD}/mysql:/var/lib/mysql mattrayner/lamp:latest-1804`.

![image](https://user-images.githubusercontent.com/304786/89165715-0f4df280-d5cd-11ea-9301-469161ad35fd.png)
