### 2020-08-03 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

Thanks for your feedback and suggestions.

I am googling your error but have not figured out what’s goung wrong.

An idea: have you tries to run the docker command with sudo rights?

Peter

> Op 3 aug. 2020 om 11:06 heeft Bruno P. Kinoshita <notifications@github.com> het volgende geschreven:
>
> ﻿
> Almost there... but I cannot connect to the backend (receiving 404 errors), drives me crazy. I will give it another try tonight; or else will upload it to the installation instructions as 'work in progress'.
>
> Sounds like a good plan, if we have at least the browser platform working, that should be enough to run the app locally. Later we can work on the Android or iOS platforms, or tell users that they can download the apk or use the store version for the time being.
>
> Inspecting your code, I see that you changed the ports and you use a capital letter "A" in app. Does that cause the problem?
>
> I don't think so. I tried both ways yesterday. But will try again today to confirm.
>
> Just synced the repo. Then executed:
>
> docker system prune -a -f
> docker network prune
> docker volume prune
> Then running docker run -i -t -p "80:80" -v ${PWD}/app:/app -v ${PWD}/mysql:/var/lib/mysql mattrayner/lamp:latest-1804:
>
>
>
> What do you see when you navigate to localhost/myphpadmin in your browser? What happens when you wipe the image completely and re-install / re-initialize it?
>
> With the container in that state:
>
>
>
> and
>
>
>
> (slight change in address)
>
> To be sure: The app folder in docker run does not refer to the app folder in the repo, but to the persistent location of the created webserver by mattrayner/lamp. Here is where you need to put the dearscholar.php file (see the instructions, step 5, this is a bit confusing I now see). It would be great indeed to include DearScholar in the docker too, but that is something on my bucket list.
>
> Ah, I see. I thought we were mapping the App folder within the container. I am stuck at step 1 at the moment, trying to get pass "Step 1: Server setup". I can skip it, but looks like mysql is not running.
>
> I believe my container has no services listening on any ports at the moment. Its initialization seems to have failed.
>
>
>
> To be sure: Along similar lines, you cannot directly load the content of my mysql file into the MySQL database, but (at the moment) should use the mysql command line in the docker image (see step 1 and 2 of the manual). Can you open the mysql docker in your terminal (see step 1 of the manual).
>
> 👍 looks like I was thinking too much ahead. Your instructions would have showed me how to use the app and mysql folders. Once we manage to fix the initial set up, everything should work 🤞
>
> In short, can you let me know what happens when you re-install the Docker image, preferably also delete the app folder and mysql folder created the image (both folders are at located at Users/peterkruyen/) and run docker run -i -t -p "80:80" -v ${PWD}/app:/app -v ${PWD}/mysql:/var/lib/mysql mattrayner/lamp:latest-1804.
>
> Sure. Removed the app directory (looks like it had been created), and re-executed the exact command docker run -i -t -p "80:80" -v ${PWD}/app:/app -v ${PWD}/mysql:/var/lib/mysql mattrayner/lamp:latest-1804.
>
>
>
