### 2020-08-03 pmkruyen/dearscholar#18 @pmkruyen
And if your run the docker command, I think you should also rename the `App folder`. 😄
