### 2020-08-03 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

**Android**
I can register the device with some adjustments in the index.html file and config.xml (that is I can access the php-file to communicate with the server), but cannot load the tables in the emulator yet. That should be the final step!

**Backend**
Hope my suggestion above does the trick 🤞

Peter
