### 2020-08-03 pmkruyen/dearscholar#18 @kinow
Problem solved @pmkruyen !

>Wait a sec... I also see that you are trying to run docker run -i -t -p "80:80" -v ${PWD}/app:/app -v ${PWD}/mysql:/var/lib/mysql mattrayner/lamp:latest-1804 from the project folder.
>
>When trying this too, I got the same error because of the presence of mysql in that folder (and also anApp folder in this directory). Could you try to run the command in another directory or renaming the mysql file? (I also see that the Docker makes the mysql executable when ran in that directory).

Sorry! You were correct here. Just `cd /tmp/ && docker....` fixed the issue :grimacing:

Will continue with the other instructions today :+1:

Thanks Peter!
Bruno
