### 2020-08-03 pmkruyen/dearscholar#18 @kinow
Further progress today @pmkruyen

Right now I think I have the backend working. I can see the POST request in the browser going to the Docker container, and being returned as success.

However, the log in still fails with

>Er is iets mis gegaan met het opslaan van uw PIN code. Controleer uw gebruikersnaam en wachtwoord en probeer vervolgens opnieuw of neem contact op met Peter Kruyen.

![image](https://user-images.githubusercontent.com/304786/89235675-c042a480-d642-11ea-8ee7-0fccfcbe20bb.png)

I logged in with user `test1`, password `test1`. Then typed `1234` twice as my PIN. It looks like the PIN was actually stored in the database (that `pintable` table was empty before I tried).

![image](https://user-images.githubusercontent.com/304786/89235710-d486a180-d642-11ea-89fc-3170457f5aa4.png)

Any idea what I could be doing wrong?
