### 2020-08-04 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

Happy the Docker works. Will improve the instructions when you/we passed all setup stages.

Ok, let’s start working on the PIN error. Check: how is the password stored on the server (mysql database)?

Has it been hashed too (see the instructions, and the example.php script about how to achieve this)?

Will look at it further tonight!

(Also i will check if I did not make a mistake in setting the correct database rights).

Best,
Peter

> Op 4 aug. 2020 om 01:09 heeft Bruno P. Kinoshita <notifications@github.com> het volgende geschreven:
>
> ﻿
> Further progress today @pmkruyen
>
> Right now I think I have the backend working. I can see the POST request in the browser going to the Docker container, and being returned as success.
>
> However, the log in still fails with
>
> Er is iets mis gegaan met het opslaan van uw PIN code. Controleer uw gebruikersnaam en wachtwoord en probeer vervolgens opnieuw of neem contact op met Peter Kruyen.
>
>
>
> I logged in with user test1, password test1. Then typed 1234 twice as my PIN. It looks like the PIN was actually stored in the database (that pintable table was empty before I tried).
>
>
>
> Any idea what I could be doing wrong?
>
