### 2020-08-04 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

Can you also check what happens when you try to access the other tables in `phpmyadmin`. I got a #1146 error now e.g., `#1146 - Table 'project.questiontable' doesn't exist` This is also causing me not being able to start the app properly in the Android emulator :)

Maybe caused by moving around with the Docker image?

I see online several solutions, such as stopping and restarting `mysql`. When I found the solution tonight, I let you know.

Peter

PS. Do you also receive this error when trying to click on the tables in the left panel of phpmyadmin?
