### 2020-08-04 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

I removed the `mysql folder`, restarted the Docker image, and populated the database following the first steps of the instructions.

All works, indeed the Android emulator works too :) (albeit the rendering is not 100% correct). Will start updating the instructions tonight.

Hope we get it working on your computer very soon too 😄 .

Peter
