### 2020-08-05 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

Following your comments, I have moved all the technical details on the Wiki page (and restructured them a bit).

**Backend**
I will update / finalize these installation instructions (Step II and III) when DearScholar is working on your laptop.

**Android**
I will have to upload some additional testing files tonight. I drop a small note when I have done so, and you can then verify the installation instructions for the Android emulator when you have time. 😄

Thanks for all your patience and help.

Peter
