### 2020-08-05 pmkruyen/dearscholar#18 @kinow
Thanks @pmkruyen !

I had an evening meeting with the rest of my team ($work) yesterday, and today was organizing everything to start a 2 days leave (thu/fri NZ).

Planning to use some of this spare time to have another go at reviewing the build instructions from scratch, after cleaning the whole environment.

I also believe we should be able to run some command like `docker-compose up`, or `docker build . -t dearscholar`, or even `make docker`, etc., and get 1 or more containers with everything necessary for running DearScholar.

Without having to perform any extra steps, such as running SQL or PHP commands, or editing files. Just would need some time to try - this is not necessary for the JOSS review, IMO. We can use the build instructions in the Wiki if they work. But I want to practice some Docker fu, so will try to automate the steps later :)

Will try to reproduce the build tomorrow I think :+1:

Thanks a lot!!!
Bruno
