### 2020-08-05 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

I just uploaded the files (in the newly created `installation folder`). Good luck with testing, hope that it works🤞

>
> I also believe we should be able to run some command like docker-compose up, or docker build . -t dearscholar, or even make docker, etc., and get 1 or more containers with everything necessary for running DearScholar.
>
> Without having to perform any extra steps, such as running SQL or PHP commands, or editing files. Just would need some time to try - this is not necessary for the JOSS review, IMO. We can use the build instructions in the Wiki if they work. But I want to practice some Docker fu, so will try to automate the steps later :)
>

Yes, this would be really welcome when the JOSS review is over and the paper has been accepted for publication (especially because all the testing on emulated Android devices requires quite some steps, from which I learned a lot by the way). Thank you Bruno.

Have a nice leave / long weekend and thanks for the effort. Hope that a fresh restart solves the login issues and you get the Android emulator working.

Best,
Peter

PS. I will be on vacation next week (my responses will somewhat be delayed).

PPS. When you loaded the tables in DearScholar and you change settings (e.g., put in different questions), it is important to push the reset button in the app otherwise the changes will not be shown (see the 'settings and options' sections too) 😄
