### 2020-08-07 pmkruyen/dearscholar#18 @kinow
Probably restarting the review tomorrow @pmkruyen :+1: if not the whole process, at least i) cleaning local environment (remove platforms, and the docker containers), and ii) sync repo, and iii) re-read the instructions.

Will try to follow the instructions again with care either tomorrow or Sunday.
