### 2020-08-07 pmkruyen/dearscholar#18 @pmkruyen
Hi Bruno,

Thank you so much! Hope we got everything working 😄 .

Have a good weekend,
Peter
