### 2020-08-07 pmkruyen/dearscholar#20 @pmkruyen
Hi Bruno,

**Automatic backend testing**
Today I have written a simple php script to perform some basic checks to test whether a survey is correctly programmed by the researcher (following the required settings, see also the new Wiki page).

The script definitely needs some improvements (e.g. checking other elements of the backend + rendering in html). I will work on this after my vacation. Your feedback is very much appreciated.

(A better approach would be to write a small program that helps researchers to set up a survey correctly, but that is something for next year 😄.)

**Automatic app testing**
Will have a closer look at this after my vacation. I will see how far I get.

**Manual steps for testing**
Regardless of the degree to which I succeed in implementing some automatic tests, I will write down some manual steps for testing in line with the JOSS guideliness. Will start working on this when you succeed in installing the backend + app (hopefully after my vacation too 😄).

Thanks!

Peter
