### 2020-08-07 pmkruyen/dearscholar#20 @kinow
Hi Peter,

If you have a PHP script, maybe that could be converted to a PHPUnit test?

Either way, if the container can be set up so that running the script works inside or outside the container, it would be easy to automate that with GitHub actions or Travis CI, to run with every commit, preventing regressions.

Enjoy your vacation! We can discuss more about that later :tropical_drink:
Bruno
