### 2020-08-08 pmkruyen/dearscholar#18 @kinow
Followed new installation instructions from the beginning to end, line by line. Successfully built and ran DearScholar locally :+1:

![image](https://user-images.githubusercontent.com/304786/89703401-60c7fa00-d99e-11ea-9d83-0a90127ec6ad.png)

Closing now! :tada:

Thanks @pmkruyen !
