### 2020-08-08 #2506 @kinow
Followed new build instructions written by @pmkruyen , successfully built and executed DearScholar locally. Other users should be able to reproduce it, given they follow the instructions and there are no packages removed from NPM/DockerHub. :+1: thanks @pmkruyen !
