### 2020-08-08 pmkruyen/dearscholar#20 @kinow
@pmkruyen build done. I've played with DearScholar, and it reminds me of an application we have at NIWA for Citizen Science. In our application, researchers are able to create custom forms, like in SurveyMonkey or Lime Survey, but with more functionalities (e.g. some GIS information to use maps, questions referencing questions, etc).

The functionality appears to be OK.

For my JOSS review, I think the only part missing now is this issue. Once we have tests, I can announce in the review thread that my review is done with no issues :-)
