### 2020-08-08 3cb1159
Add Dockerfile, entrypoint, and ignore files
