### 2020-08-08 0b7afdb
Move apt-get to top to avoid re-creating layer, and remove copy of dearscholar.php (done in entrypoint now)
