+++ Dockerfile
@@ -3,17 +3,15 @@
 
 FROM mattrayner/lamp:latest-1804
 
+# Install nc, to check for MySQL socket status in the docker-entrypoint.sh
+RUN apt-get install -y netcat
+
 COPY Installation/docker-entrypoint.sh /
 RUN chmod 755 /docker-entrypoint.sh
 
 # Copy installation files over to the container
 RUN mkdir /installation
 COPY Installation/* /installation/
-
-RUN cp -T /installation/dearscholar.php /var/www/html/dearscholar.php
-
-# Install nc, to check for MySQL socket status in the docker-entrypoint.sh
-RUN apt-get install -y netcat
 
 ENTRYPOINT ["/docker-entrypoint.sh"]
 CMD ["/run.sh"]
