### 2020-08-08 pmkruyen/dearscholar#20 @pmkruyen
Hi Bruno,

> it reminds me of an application we have at NIWA for Citizen Science.

Is it somewhere in the public domain? Is there a link? I will refer to it in the paper.

> Either way, if the container can be set up so that running the script works inside or outside the container, it would be easy to automate that with GitHub actions or Travis CI, to run with every commit, preventing regressions.

Good idea. Will start working on this after my vacation (including the phpunit test idea).

> Enjoy your vacation! We can discuss more about that later 🍹

🍹🍹🍹

Thank you
