+++ .travis.yml
@@ -1,0 +1,20 @@
+language: objective-c
+osx_image: xcode11.2
+  
+before_install:
+- nvm install v12.16.2 --reinstall-packages-from=node
+- nvm use v12.16.2
+- echo rvm_autoupdate_flag=2 >> ~/.rvmrc
+- wget http://dl.google.com/android/android-sdk_r24.4.1-macosx.zip
+- unzip android-sdk_r24.4.1-macosx.zip
+- export ANDROID_HOME=$PWD/android-sdk-macosx
+- export ANDROID_SDK_ROOT=$PWD/android-sdk-macosx
+- export PATH=${PATH}:${ANDROID_HOME}/tools:${ANDROID_HOME}/platform-tools
+- echo yes | android update sdk --filter platform-tools --no-ui --force > /dev/null
+- echo yes | android update sdk --filter android-28 --no-ui --force > /dev/null
+- echo yes | android update sdk --filter build-tools-28.0.3 --all --no-ui --force > /dev/null
+install:
+- npm install -g cordova
+script:
+- cordova platform add ios
+- cordova platform add android
