+++ .travis.yml
@@ -16,6 +16,7 @@ before_install:
 - cd app
 install:
 - npm install -g cordova
+- npm i glob
 script:
 - cordova platform add ios
 - cordova platform add android
