+++ .travis.yml
@@ -16,7 +16,6 @@ before_install:
 - cd app
 install:
 - npm install -g cordova
-- npm i -g glob
 script:
-- cordova platform add ios
-- cordova platform add android
+- cordova platform add ios@5.1.1
+#- cordova platform add android
