+++ .travis.yml
@@ -29,5 +29,5 @@ addons:
   browserstack:
     username: "peterkruyen1"
     access_key: 
-      secure: "QflrXxjk8HswLeCLKeXj1ShjpYpQanYxvvsMepNE57EfUgIj83gyyjPk9Z5yemf3Ugzgr95oz6VxZLhJeqoI+66RUgR+BqDKZ+D2atpvCQkI4u1jP9jynYTMS80kLcopBU3dOUwiCnKc6oWard/Jq1k6k1+qIucZoYGKOnulqUoUNOaKgUkH2yl443cpEqxZQPkqza4N9/K5Cu6U6kOBLjTJA6oVJ4swpmqjK/OK7bUoas8jSEp89Eh/0PPCWQivVxfItW6NGNaZ2i0XFJbJ4hIwHFMIvFWh9zTic56KZPJxpQpjGrggVLy93bsqycKYlMCoDG0HQ2MIvo77jj0iPlasFqgYyE2ThWT40PdOXuVS6JOv2++S7jT8vOiSNR479yALmAj7BLanPh+L8IW6z8oc60cQfVY/0RZ8ClY8h+ihtFQ28298pEdpRl7JN93FgFdZv+Ddy/jpejcc0XPwuVfK+ekWTorLXADPFdoEHVH1x7po3WNCsE29traFhNnzQPob6VOXMLtHr4ZABheR3cY3yIla3hrIzZFrChg+dh4roZjfq6vHg2Y8K/1wfnGmWGxmmvkh3IbfE5O5V2qyGbQyZ1v5O1Tw7Zp07cQQ9X51PBhp/jdFif0dzWgdFGcmk4f8C1e5sHJZZFsTLs4l+S3navx0cfUI7me24cQRejQ="
-    app_path: "browserstack_android_sample/build/outputs/apk/debug/browserstack_android_sample-debug.apk"
+      secure: "f4H6eoDaJo0X9RcVZOhcvi+H5iVRlW62ioWBPf/I47Cwg7xk8ZaRXd7+zoIc9pSHO69KYvOV1aYeIyDLsbX6x7yuqGRjLt7u42RHek71k2iexHVpn2loEfH8oQIFfu7XbFuWf+CJXXsXHvmhvAjWkLObMEgmhMGzRyBJOMkjH9Jv72j2jrfFjD5pBimmffqH/pPawrzWjJdM85tnTbDe/DJ2ebDM0nJ0LK72W88sn2bboHB9/W3oNi9Ekw0JwQTgb+KvCDURXzVDlhdBN0/EKQTtqXzpC9TudzT+z+H3X3hHVLPkbAcVuSX56qSav+OGi8tKprvctdOpI3j7BUe95Df0Iosb/3jkP3oAwjrmStH0xsifxIayEqBKExAgTFe1TTuGP4dEq3TD641MDa5pUfDbiLZMgvX6dXbj/0q3+4R21wwPKmGSVYzN+H460mnr1HBbnESOI6jtw95BHuQvocOwbTzswj9JAathvuiDd1yz7UIG1cn+U4fMzvzq/Ca8NZ/ruV2R9wKGi8h9dVloHU2lqsi6fk+i8cbTJ9SBQ5n2rHttaOAE1HmyGQG9pN31hLCA7Hs+GYZxI/bwzMIk5rZvTAiVEK1wZEeM+Mli+PLwA5XWWxo6/KxHrrCL1OgtCwagIILrY1YRziFtsQ/bfossExDRd0xQT9JnNmuj7c8="  
+      app_path: "dearscholar.apk"
