+++ .travis.yml
@@ -29,5 +29,5 @@ addons:
   browserstack:
     username: "peterkruyen1"
     access_key: 
-      secure: "f4H6eoDaJo0X9RcVZOhcvi+H5iVRlW62ioWBPf/I47Cwg7xk8ZaRXd7+zoIc9pSHO69KYvOV1aYeIyDLsbX6x7yuqGRjLt7u42RHek71k2iexHVpn2loEfH8oQIFfu7XbFuWf+CJXXsXHvmhvAjWkLObMEgmhMGzRyBJOMkjH9Jv72j2jrfFjD5pBimmffqH/pPawrzWjJdM85tnTbDe/DJ2ebDM0nJ0LK72W88sn2bboHB9/W3oNi9Ekw0JwQTgb+KvCDURXzVDlhdBN0/EKQTtqXzpC9TudzT+z+H3X3hHVLPkbAcVuSX56qSav+OGi8tKprvctdOpI3j7BUe95Df0Iosb/3jkP3oAwjrmStH0xsifxIayEqBKExAgTFe1TTuGP4dEq3TD641MDa5pUfDbiLZMgvX6dXbj/0q3+4R21wwPKmGSVYzN+H460mnr1HBbnESOI6jtw95BHuQvocOwbTzswj9JAathvuiDd1yz7UIG1cn+U4fMzvzq/Ca8NZ/ruV2R9wKGi8h9dVloHU2lqsi6fk+i8cbTJ9SBQ5n2rHttaOAE1HmyGQG9pN31hLCA7Hs+GYZxI/bwzMIk5rZvTAiVEK1wZEeM+Mli+PLwA5XWWxo6/KxHrrCL1OgtCwagIILrY1YRziFtsQ/bfossExDRd0xQT9JnNmuj7c8="  
-      app_path: "dearscholar.apk"
+      secure: "HG0OIrQvscABfgB10uWrbhGYfGI8bitpBwWbYq3+Jzcr3NMoBL+9cnVaXLdJEyumuN9/YXewIJGOyn5ShF7y+w2Gll59ydZvvzHusc87tUUt9nNZGwaT8mde5FAyh53JvxEtRWKa6x8+fiBlOSuD0PENxA68Y8Ialn/B70RjO7V9dlS/1pomADLORZzd/p3cfhaXhIGFMlTRMcLLsKlveTvuUKNJGhbmIb+ZMuy+HDcM4HLAt0IX3wv9xKMxsh/fCVI1Q77CeS0aUOe6roTWK1Dtbfis0NlPKq1nh5+gOoLvTPFoFWAzuQolAlVWpVjrzTFDT3xYOLjnACgCJbhWyyHejl7YTaIp/chkr1XohRkowjg1pD/EnM+801dzO83bnKggIZmz06ATUEJh+JVxxad/VGiltUS8Hboqs/lGKeqQRTr+gN3y5YKAv4L+jll4ZjvTwFQ+iN9XMax7Y6gN/B4hyt8E6r/ZWmYrGq+OvTJEaNYkim+kCPEBRGoqx6zEX9Mk9HWgkHkcxrqIBoAfhL0ZEWPQ8PAGAR+N89TNW9FLU2ivvDjwYDOJHROHVLCD8PQ5ycyUBKc/+1JypBonvz+cbClVtFAo39l5QQRoblSxzQ9kHahZOcQ/w00LOwUwtcHQhKF+X6njQFBMW5OwrM5CK9Htp7rHko93zjfUmSI="
+    app_path: "dearscholar.apk"
