+++ .travis.yml
@@ -29,5 +29,5 @@ addons:
   browserstack:
     username: "peterkruyen1"
     access_key: 
-      secure: "HG0OIrQvscABfgB10uWrbhGYfGI8bitpBwWbYq3+Jzcr3NMoBL+9cnVaXLdJEyumuN9/YXewIJGOyn5ShF7y+w2Gll59ydZvvzHusc87tUUt9nNZGwaT8mde5FAyh53JvxEtRWKa6x8+fiBlOSuD0PENxA68Y8Ialn/B70RjO7V9dlS/1pomADLORZzd/p3cfhaXhIGFMlTRMcLLsKlveTvuUKNJGhbmIb+ZMuy+HDcM4HLAt0IX3wv9xKMxsh/fCVI1Q77CeS0aUOe6roTWK1Dtbfis0NlPKq1nh5+gOoLvTPFoFWAzuQolAlVWpVjrzTFDT3xYOLjnACgCJbhWyyHejl7YTaIp/chkr1XohRkowjg1pD/EnM+801dzO83bnKggIZmz06ATUEJh+JVxxad/VGiltUS8Hboqs/lGKeqQRTr+gN3y5YKAv4L+jll4ZjvTwFQ+iN9XMax7Y6gN/B4hyt8E6r/ZWmYrGq+OvTJEaNYkim+kCPEBRGoqx6zEX9Mk9HWgkHkcxrqIBoAfhL0ZEWPQ8PAGAR+N89TNW9FLU2ivvDjwYDOJHROHVLCD8PQ5ycyUBKc/+1JypBonvz+cbClVtFAo39l5QQRoblSxzQ9kHahZOcQ/w00LOwUwtcHQhKF+X6njQFBMW5OwrM5CK9Htp7rHko93zjfUmSI="
+      secure: "mWFqS9xtQSgL74ZWjP6GAE+9/zrkqlSx7SMQ383VurqsihVKWnMR4QvF4JabgLU8GxTYbAQOHoPfhuQwEw3lDHd/cqIazz1HhiMN6WODDNKdfXsxwFtfTIyulVyZZ3dYrOaGu7QKRkz2ZScbbfseF0ap0/GCPHg439ffh1HTncViNbwlF7nkU0vk7XoxE8Gs6Sn3jn2jrbG9IVjhao0lWAzqR90PHb4MqizZYZWh1DSWRcvIhxM9j43sR4fjkHb4AfnFRbyNgrwyp718t2ERnDxAlnJYY72SLmDv0F8ZtKnl6ioIpqCw+bQ8h26tD9woTG9IPnrQ9T5RoiMBQqxYmTcKz4RkJs/VFxEtW6oW3xuOGQC3lqMgJKggPyaQ0espE2382dTEurC8nU6Aj23n6YHTKxVLNUJlwmwTWiGhfAW/HMqxWPgg+uwqdXs8fvRPkiIGVi3UfBcDXDU/+4jCnKZsw4aidkeJDKqRtIwa+1jBjBneSYTw+ctXzGPVEBQu+HCy31acs1VTiui8FlZw4RD+PyCI2iSKNQkQrPjzjl1kWbX6XCm/aI5Ijdu4PGvHh+8m1gPJZI406yfc099ob5uXcexu9yaBmyQOpu3EuWwBDPtmYfyDh7kNmm8lz2/znAwtboYfW7HA1NfJkwHY/sGNICdqBOkpxqPG+SBj7sk="
     app_path: "dearscholar.apk"
