+++ .travis.yml
@@ -29,5 +29,5 @@ addons:
   browserstack:
     username: "peterkruyen1"
     access_key: 
-      secure: "mWFqS9xtQSgL74ZWjP6GAE+9/zrkqlSx7SMQ383VurqsihVKWnMR4QvF4JabgLU8GxTYbAQOHoPfhuQwEw3lDHd/cqIazz1HhiMN6WODDNKdfXsxwFtfTIyulVyZZ3dYrOaGu7QKRkz2ZScbbfseF0ap0/GCPHg439ffh1HTncViNbwlF7nkU0vk7XoxE8Gs6Sn3jn2jrbG9IVjhao0lWAzqR90PHb4MqizZYZWh1DSWRcvIhxM9j43sR4fjkHb4AfnFRbyNgrwyp718t2ERnDxAlnJYY72SLmDv0F8ZtKnl6ioIpqCw+bQ8h26tD9woTG9IPnrQ9T5RoiMBQqxYmTcKz4RkJs/VFxEtW6oW3xuOGQC3lqMgJKggPyaQ0espE2382dTEurC8nU6Aj23n6YHTKxVLNUJlwmwTWiGhfAW/HMqxWPgg+uwqdXs8fvRPkiIGVi3UfBcDXDU/+4jCnKZsw4aidkeJDKqRtIwa+1jBjBneSYTw+ctXzGPVEBQu+HCy31acs1VTiui8FlZw4RD+PyCI2iSKNQkQrPjzjl1kWbX6XCm/aI5Ijdu4PGvHh+8m1gPJZI406yfc099ob5uXcexu9yaBmyQOpu3EuWwBDPtmYfyDh7kNmm8lz2/znAwtboYfW7HA1NfJkwHY/sGNICdqBOkpxqPG+SBj7sk="
+      secure: "$BROWSERSTACK_ACCESS_KEY"
     app_path: "dearscholar.apk"
