### 2020-08-10 pmkruyen/dearscholar#20 @kinow
Hi @pmkruyen

> Hi Bruno,
>
> > it reminds me of an application we have at NIWA for Citizen Science.
>
> Is it somewhere in the public domain? Is there a link? I will refer to it in the paper.

The app is in the Android & iOS application stores. The best link to use in the references would be https://citizenscience.niwa.co.nz/. The source code is not open at the moment.

> > Either way, if the container can be set up so that running the script works inside or outside the container, it would be easy to automate that with GitHub actions or Travis CI, to run with every commit, preventing regressions.
>
> Good idea. Will start working on this after my vacation (including the phpunit test idea).

Great! That should be easy to review.

> > Enjoy your vacation! We can discuss more about that later 🍹
>
> 🍹🍹🍹
>
> Thank you

Thanks!
Bruno
