+++ .travis.yml
@@ -29,5 +29,5 @@ addons:
   browserstack:
     username: "peterkruyen1"
     access_key: 
-      secure: ""NdISIA2fhBCGOOvNsZyhXQ1DOsH94s2T52m5ADWy/+y5rmkm4g4we3r8UnpKObwswgnj9lV+tgJLATv+NZZiNoQYC6TPrcoJqIG8znM8Ie1eVItgLFVUUHp+7s13JBO0fLqoTtmBrBBwYp7OQbzocl8FKkrqFMDBGyfQj8DDPtAaKNQ+rkID+iEd+gVPk/U4iaXTPQbYCQ7eamjHUGbRorSJEpTT7Gll74VCVpnxGj/5MoOUgTTPIOp82oI4ciHDjLpIMRSOlCRkQuG4UgVveGnk1CKMRwkK/LpqhPUVhKjUu0J8OXkHYVsQgqkFQ2ZoHFHZwSeMpWk5AVkX+FxEBlEJsWq/mZK6ET67sDuGTSajoKe2q/dDK0WDQgcpIIODDmuKCiHY+VpAW7zgMvD7Z9lndQUi727UG3SB6hWbDar9/Wa5aXP0qOp9Q3e7ux6T7YC++4Cn7dJgCQgMhv/oouVZSPxz8SjmqreKrRydjjNWIsAzzuBCWP6S+OS+i96UeZhicEhWn91Mz1OZhW2lvnTN11w6eTekpdQcZJBWwD6UstjGR+uomjOgC3tmn6LuIQIcm7y1pP1srja1YiS9ApDyYPiJTSanFicsRayqTsvtgmWBhW8Bo2FxMqFxP5QRTC8aPpWjpaYsqEAQBq1EfB5+zG26vLYRpheNVZ4NZ5k="
+      secure: "rV5CNlzaTPV7gVtn76Kppk/ckTz60R7S+6gaQBwMrFFToHdzdQDkuX3h3uJrjqxTrPkdHbZTuZ8duk/TToJ051ZE/u8u5WDCx+1LJuy9dHj6kCbht3FnghLMP1qThJ1fHrCoS1f6NtjdU060oZCM9Je5PMg9PEcmqtdH0fBeSUDSfBMFx9+jCwedLoJgKIC+9MRqfFWY0C07Ks8z6Q9tBZI6NnE0fujvPc72JC3wL8qyttgiYeBw0gMfdlK4yhM04axRHDGCFNx9OOJex7nASaIrog8btXlmCRM06RmlDoSHYG4B4c6bgomnRa9ceA+hzjaKsRZ8WJBF+4xPMSFJZJ37DpxKl8zXv1e2CL4OotMYq56PNRjA1cHYAgz446WTwXLfC7y3U1B8c7kHFZNTLRoI3By3dcb7jcWDYUvOHaeKB8k+V3FdoFTECDbrf2kqbsOuP7F+YZjt0RpZizqm9lUD2rvNjRvubK6Qrc2Mw793Bf6/dZmZvBfFrYyX/BzPDIzrlrLjBwHwFpqqEb5a/HF5x4CZLerChSt0+5tdReeCqMgZNgKRmfTrJkbKTolgSaQ/Q9lG1E4NPMnlFzW+hl9xCKCcl92oibaNlU7cUlNX5njfBFFjYhDGFfyfpynUv2P41pTHm0G2e+kWavU5bioqQieFz3fXvMpnW2tGnQU="
     app_path: "dearscholar.apk"
