+++ .travis.yml
@@ -27,7 +27,7 @@ script:
 
 addons:
   browserstack:
-    username: "peterkruyen1"
+    username: peterkruyen1
     access_key: 
-      secure: "rV5CNlzaTPV7gVtn76Kppk/ckTz60R7S+6gaQBwMrFFToHdzdQDkuX3h3uJrjqxTrPkdHbZTuZ8duk/TToJ051ZE/u8u5WDCx+1LJuy9dHj6kCbht3FnghLMP1qThJ1fHrCoS1f6NtjdU060oZCM9Je5PMg9PEcmqtdH0fBeSUDSfBMFx9+jCwedLoJgKIC+9MRqfFWY0C07Ks8z6Q9tBZI6NnE0fujvPc72JC3wL8qyttgiYeBw0gMfdlK4yhM04axRHDGCFNx9OOJex7nASaIrog8btXlmCRM06RmlDoSHYG4B4c6bgomnRa9ceA+hzjaKsRZ8WJBF+4xPMSFJZJ37DpxKl8zXv1e2CL4OotMYq56PNRjA1cHYAgz446WTwXLfC7y3U1B8c7kHFZNTLRoI3By3dcb7jcWDYUvOHaeKB8k+V3FdoFTECDbrf2kqbsOuP7F+YZjt0RpZizqm9lUD2rvNjRvubK6Qrc2Mw793Bf6/dZmZvBfFrYyX/BzPDIzrlrLjBwHwFpqqEb5a/HF5x4CZLerChSt0+5tdReeCqMgZNgKRmfTrJkbKTolgSaQ/Q9lG1E4NPMnlFzW+hl9xCKCcl92oibaNlU7cUlNX5njfBFFjYhDGFfyfpynUv2P41pTHm0G2e+kWavU5bioqQieFz3fXvMpnW2tGnQU="
+      secure: rV5CNlzaTPV7gVtn76Kppk/ckTz60R7S+6gaQBwMrFFToHdzdQDkuX3h3uJrjqxTrPkdHbZTuZ8duk/TToJ051ZE/u8u5WDCx+1LJuy9dHj6kCbht3FnghLMP1qThJ1fHrCoS1f6NtjdU060oZCM9Je5PMg9PEcmqtdH0fBeSUDSfBMFx9+jCwedLoJgKIC+9MRqfFWY0C07Ks8z6Q9tBZI6NnE0fujvPc72JC3wL8qyttgiYeBw0gMfdlK4yhM04axRHDGCFNx9OOJex7nASaIrog8btXlmCRM06RmlDoSHYG4B4c6bgomnRa9ceA+hzjaKsRZ8WJBF+4xPMSFJZJ37DpxKl8zXv1e2CL4OotMYq56PNRjA1cHYAgz446WTwXLfC7y3U1B8c7kHFZNTLRoI3By3dcb7jcWDYUvOHaeKB8k+V3FdoFTECDbrf2kqbsOuP7F+YZjt0RpZizqm9lUD2rvNjRvubK6Qrc2Mw793Bf6/dZmZvBfFrYyX/BzPDIzrlrLjBwHwFpqqEb5a/HF5x4CZLerChSt0+5tdReeCqMgZNgKRmfTrJkbKTolgSaQ/Q9lG1E4NPMnlFzW+hl9xCKCcl92oibaNlU7cUlNX5njfBFFjYhDGFfyfpynUv2P41pTHm0G2e+kWavU5bioqQieFz3fXvMpnW2tGnQU=
     app_path: "dearscholar.apk"
