+++ .travis.yml
@@ -27,6 +27,6 @@ script:
 
 addons:
   browserstack:
-    username: "petertest7Bhuj7"
+    username: "petertest7Ch2zR"
     access_key: 
-      secure: "VXtOpNhf4QwocZRtoGczdeZwJx/j98pSyNRxTXKRSAIfyU5W4dNHsLK+fmJu+CyIRvNhxqen/r1xtGtKLGJrzplqsbA21cuqMQlMtz9PnYKsM2SVnRRxpgQ34Xn+afwYgC+Dbri45uYrsR5Pf/ee5wX3losJUN4uopVjWeoliIFB31mRCbmUfPv+0HVuWT+P8tLOQFsexvC4OixJyEbJ6y8JfCO3eqDjDTurMR4OijEh1Ytkl91DTrduRhM/3vQnupTV3490oWECaLZbyHby7Gj3li7EUtAX6GRWFjYmxg/7HtTfXzR79oKvzjbpWbXSoJ5olmeIXQN7LFdq1KLfp17XHxwp5K4/AdZVtE/79QzYt59vgijDTKDf4b4bhhcfq1oaqO+pYTIzP1ajrPMDIzvPBo3wmwMz2b+l2oquooIkD+LwEozRpa7Q6A0teYH0QerDjwnlbZ9hWUtAynpZf6XtqxR7j+K62lL6cyyByPlkMnIf7IP7TolJQPMZ/ELK4vIzJ2e7eKsQM1GMxfh9u+r3xWke2IRD6pPIQ/jXK/yu8R+Yud4m/qD/tPfkcwGQVVveBuffzjSn11fqBu4OpeSjtH/JRIV+ViG/TZ/yDilAfdMbdyUP7E1t5whwDgJnXuXWbEwvpxkjoZ13YMMZSOGaPUb5/H8yfXMxemqIgzs="
+      secure: "GLxxcC94qpNu3c6hUkRaKY+Mzl2XqNzT8R75xrqmjce1ooYLSWP9UufAZ47mSkxUf6cxgequW5uPxc5N5dYLYkmsrYLMIcS+J9aVrrKCLicz/dv3Zle4+wtS19MRMx/RrWiYGvXp7IZFMUaV4OJFIWDv/oljL+AaDoaiwagMOrFepgg0IKbYduQd5sQqsojHXeENNj8nHTUuT6LHZm4TPpI23TYZc7tNWzM5vUht5N2N0LPe2ji6SSp9IFiTETnT3/TEA9PMNUDs89N0WeTS94U2yt6q2yTv0Ebe9ARCVaDOMXeHuowEIV7lmj4qbcTDCT4MlbedhcJ0SbEWNmKuhitI0bf7HP+TFMMcrBJZ+ip+tb03OCAidHp+IkLVxaIRbzlNP5BfYebbmdLCBpKBINjcYdD0tNLlug+mMlns96UfYn8c3e+alnRKHrcfGEoDjJwPgQIBPZK2rJrED05+1gTK5p35mpHgGFgdqaqS0ujU8rNTeYpTILcfAx2iUcBZj5tnU/h3BpWuXPt4sMNkke32i9VZwEZlzaeKJ+WXA5j4JOJNHYidrYFTBnMHj3TTSLbKFTcGelUsZoI4Lha3fYAzNBiLbeorMyw/9vcNiA+bHsUFIVhs1nYi7joJTUx5wyKXtzc5IKnm7qclst4MH2cSIrw2Y9BhnkngcFCVbBI="
