### 2020-08-16 pmkruyen/dearscholar#20 @pmkruyen
Hi Bruno,

After a couple of frustrating evenings (I could not let it go during my vacation), I managed to get the basics of phpunit (and object oriented programming) to test changes in `dearscholar.php`. I have to admit, this new testing (and programming) knowledge is really helpful.

You can find the files needed for testing in the [phpunit folder](https://github.com/pmkruyen/dearscholar/tree/master/testing). Actually, I decided to write 22 integration tests instead of unit tests (as I learned the vocabulary when working on it) because I had difficulties with mocking mysql databases when I started with this "sub-project".

To run the tests on your machine:
1. start the `docker` (+ populate the database + move `dearscholar.php` in the` app` folder created by the docker.
2. in `phpmyadmin` change in the `PINtable` the `code` column for user `test3` to `REVOKE`.
3. navigate to the `phpunit` folder, run `php composer.phar update` and `./vendor/bin/phpunit` ... and the tests should be executed.

Hope it works 😃 . All feedback is very welcome!

**What's next?**
What I would like to do now, is to automate these php tests on uploading new commits to GitHub. As you can see on the `README` page, I have started to use Travis CI, which now only tests whether the app is successfully build in Android and iOS.

From your comments at #18, I understood that you compiled a new docker that includes the `dearscholar.php` file and a prepopulated database. Is this correct? Could you please share with me the code you used to build this docker? I need to tweak the Docker a bit to update the `dearscholar.php` file (I changed some parts of the code based on the testing sessions <-- as I said really helpful new knowledge :) ) and change in the `PINtable` the `code` column for user `test3` to `REVOKE`. Sorry that I have not found the time to explore your docker image yet during my vacation! In case you did not build the docker, I am sorry that I misunderstood you and try to build one myself.

**And further**
* When I have managed to automate the php unit tests, I will start with writing a script to test the javascript files and try to automate the app testing in browserstack (and then I'm done :)).
* With regard to the later, Travis CI is (still) unable to push the app to BrowserStack automatically due to issues with the encrypted key, the BrowserStack people are looking at it.

Best,
Peter

PS. Thanks for the reference to your app, looks cool! Will delve further in it when I start working on the last version of the paper.
