### 2020-08-16 pmkruyen/dearscholar#20 @pmkruyen
Hi Bruno,

> From your comments at #18, I understood that you compiled a new docker that includes the dearscholar.php file and a prepopulated database. Is this correct? Could you please share with me the code you used to build this docker? I need to tweak the Docker a bit to update the dearscholar.php file (I changed some parts of the code based on the testing sessions <-- as I said really helpful new knowledge :) ) and change in the PINtable the code column for user test3 to REVOKE. Sorry that I have not found the time to explore your docker image yet during my vacation! In case you did not build the docker, I am sorry that I misunderstood you and try to build one myself.

I found and added your docker composer. Thanks! Will start with my attempt to automate the unitphp testing in travis ci (and also update the installation instructions).

Thanks again, I learn a lot from your help!

Peter
