### 2020-08-17 pmkruyen/dearscholar#20 @kinow
Hi @pmkruyen !

Welcome back!

I struggle to disconnect from projects when I have interesting problems. I hope it didn't distract you too much from your vacation, but that at the same time you had some fun with phpunit and Docker :-)

Tests working for me already! I found one minor issue with the `Dockerfile`, as it is using `installation` instead of `Installation` when referencing the local directory. I'm sending a PR for that later.

But once I fixed that, started the container, and then ran the tests, everything worked with no issues! :rocket:

```bash
kinow@ranma:~/Development/javascript/workspace/dearscholar/testing/phpunit$ ./vendor/phpunit/phpunit/phpunit
PHPUnit 9.3.7 by Sebastian Bergmann and contributors.

......................                                            22 / 22 (100%)

Time: 00:01.610, Memory: 6.00 MB

OK (22 tests, 31 assertions)
```

Looks like you will be a bit more busy with tests, but for the JOSS paper review, we already have what we needed IMO. I'll update the review ticket now, and keep lurking here.

I learned a lot working with you too @pmkruyen ! And hope to keep having fun working with you in the future, either with this or other JOSS related projects, or in different projects :nerd_face:

Cheers
Bruno
