### 2020-08-17 #2506 @kinow
We now have documentation on how to build DearScholar, scripts in-place that help others to build in their own environments, and also simple unit tests and some functional tests in BrowserStack (@pmkruyen is writing even more tests now :tada: ).

I am done with the review, the DearScholar app passes all the items in the checklist. I've left performance unchecked as there are no claims about it in the paper, but I will say I didn't notice any performance issues using it locally :+1:

@cMadan let me know if I have any other steps in the review process that I need to complete. Otherwise, I think I have finished reviewing my first JOSS paper? :grin:

And thanks Peter for the help and patience with the review. And also for the fun and learning experience while working on the issues together.

Cheers
Bruno
