### 2020-08-17 pmkruyen/dearscholar#20 @pmkruyen
Hi Bruno,

Haha, no worries!

Thanks for all help. Super cool. And yes, I hope we stay connected!

Obviously, I keep you updated... certainly let you know when I managed to put everything in Travis.

Peter
