### 2020-08-17 #2506 @pmkruyen
Hi @cMadan,

I really appreciated @kinow's constructive and pleasant input, questions, and help (he even constructed a docker file for the project to automate the installation). He helped me to push the project to another level! I learned a lot from him and hope we stay connected and and he stays connected to the project in some way. 👍

Peter

PS. And thanks for sending a reminder to @gcdeshpande, hope to hear from him soon.
