### 2020-08-17 pmkruyen/dearscholar#20 @kinow
Or GitHub actions if you prefer. So far the only difference I have found between Travis and GitHub is that GitHub doesn't support some platforms like ARM. But if you only need a place to run Docker, PHP, JS, etc... either should work fine :+1:
