+++ .travis.yml
@@ -1,43 +1,45 @@
 
-
-
-matrix:
-  include:
-    # Step 1: test dearscholar.php 
-    -language: php
-    php: 7.3.3
-
-    services:
+ services:
       - docker
 
     before_install:
       - docker build -t dearscholar:dev -f Dockerfile .
       - docker run -ti --rm --name dearscholar -p "80:80" dearscholar:dev  
 
-    # Step 3: test the compilation process
-    -language: objective-c
-    osx_image: xcode11.2
 
-    before_install:
-      - nvm install v12.16.2 --reinstall-packages-from=node
-      - nvm use v12.16.2
-      - echo rvm_autoupdate_flag=2 >> ~/.rvmrc
-      - wget http://dl.google.com/android/android-sdk_r24.4.1-macosx.zip
-      - unzip android-sdk_r24.4.1-macosx.zip
-      - export ANDROID_HOME=$PWD/android-sdk-macosx
-      - export ANDROID_SDK_ROOT=$PWD/android-sdk-macosx
-      - export PATH=${PATH}:${ANDROID_HOME}/tools:${ANDROID_HOME}/platform-tools
-      - echo yes | android update sdk --filter platform-tools --no-ui --force > /dev/null
-      - echo yes | android update sdk --filter android-28 --no-ui --force > /dev/null
-      - echo yes | android update sdk --filter build-tools-28.0.3 --all --no-ui --force > /dev/null
-      - cd app
+jobs:
+  include:
     
-    install:
-    - npm install -g cordova
+    # Step 1: test dearscholar.php 
+    language: php
+    php: 7.3.3
 
-    script:
-    - cordova platform add ios@5.1.1
-    - cordova platform add android
+   
+
+    # Step 3: test the compilation process
+    #language: objective-c
+    #osx_image: xcode11.2
+
+    #before_install:
+    #  - nvm install v12.16.2 --reinstall-packages-from=node
+    #  - nvm use v12.16.2
+    #  - echo rvm_autoupdate_flag=2 >> ~/.rvmrc
+    #  - wget http://dl.google.com/android/android-sdk_r24.4.1-macosx.zip
+    #  - unzip android-sdk_r24.4.1-macosx.zip
+    #  - export ANDROID_HOME=$PWD/android-sdk-macosx
+    #  - export ANDROID_SDK_ROOT=$PWD/android-sdk-macosx
+    #  - export PATH=${PATH}:${ANDROID_HOME}/tools:${ANDROID_HOME}/platform-tools
+    #  - echo yes | android update sdk --filter platform-tools --no-ui --force > /dev/null
+    #  - echo yes | android update sdk --filter android-28 --no-ui --force > /dev/null
+    #  - echo yes | android update sdk --filter build-tools-28.0.3 --all --no-ui --force > /dev/null
+    #  - cd app
+    
+    #install:
+    #- npm install -g cordova
+
+    #script:
+    #- cordova platform add ios@5.1.1
+    #- cordova platform add android
 
 
 #addons:
