+++ .travis.yml
@@ -1,5 +1,8 @@
 
- services:
+    language: php
+    php: 7.3.3
+
+    services:
       - docker
 
     before_install:
@@ -7,13 +10,11 @@
       - docker run -ti --rm --name dearscholar -p "80:80" dearscholar:dev  
 
 
-jobs:
-  include:
-    
-    # Step 1: test dearscholar.php 
-    language: php
-    php: 7.3.3
+#jobs:
+#  include:
 
+    # Step 1: test dearscholar.php 
+ 
    
 
     # Step 3: test the compilation process
