+++ .travis.yml
@@ -1,6 +1,6 @@
 
     language: php
-    php: 7.3.3
+    php: 7.3
 
     services:
       - docker
