+++ .travis.yml
@@ -1,5 +1,7 @@
 
     language: php
+      php:
+        - '7.4'
     
     services:
       - docker
