+++ .travis.yml
@@ -1,7 +1,7 @@
 
     language: php
       php:
-        - '7.4'
+        - 7.4
     
     services:
       - docker
