+++ .travis.yml
@@ -1,7 +1,8 @@
 
     language: php
-      php:
-        - 7.4
+    
+    php:
+      - '7.4'
     
     services:
       - docker
