+++ .travis.yml
@@ -9,7 +9,7 @@
 
     before_install:
       - docker build -t dearscholar:dev -f Dockerfile .
-      - docker run -ti --rm --name dearscholar -p "80:80" -p "3306":"3306" dearscholar:dev
+      - docker run -d -ti --rm --name dearscholar -p "80:80" -p "3306":"3306" dearscholar:dev
       - cd testing/phpunit
       - php composer.phar update
     
