+++ .travis.yml
@@ -10,6 +10,7 @@
     before_install:
       - docker build -t dearscholar:dev -f Dockerfile .
       - docker run -d -ti --rm --name dearscholar -p "80:80" -p "3306:3306" dearscholar:dev
+      - sleep 15
       - docker container ps -a
       - cd testing/phpunit
       - php composer.phar update
