+++ .travisPHP.yml
@@ -1,59 +1,0 @@
-
-    language: php
-    
-    php:
-      - '7.4'
-    
-    services:
-      - docker
-
-    before_install:
-      - docker build -t dearscholar:dev -f Dockerfile .
-      - docker run -d -ti --rm --name dearscholar -p "80:80" -p "3306:3306" dearscholar:dev
-      - sleep 15
-      - docker container ps -a
-      - cd testing/phpunit
-      - php composer.phar update
-    
-    script:
-      - ./vendor/bin/phpunit  
-
-
-#jobs:
-#  include:
-
-    # Step 1: test dearscholar.php 
- 
-   
-
-    # Step 3: test the compilation process
-    #language: objective-c
-    #osx_image: xcode11.2
-
-    #before_install:
-    #  - nvm install v12.16.2 --reinstall-packages-from=node
-    #  - nvm use v12.16.2
-    #  - echo rvm_autoupdate_flag=2 >> ~/.rvmrc
-    #  - wget http://dl.google.com/android/android-sdk_r24.4.1-macosx.zip
-    #  - unzip android-sdk_r24.4.1-macosx.zip
-    #  - export ANDROID_HOME=$PWD/android-sdk-macosx
-    #  - export ANDROID_SDK_ROOT=$PWD/android-sdk-macosx
-    #  - export PATH=${PATH}:${ANDROID_HOME}/tools:${ANDROID_HOME}/platform-tools
-    #  - echo yes | android update sdk --filter platform-tools --no-ui --force > /dev/null
-    #  - echo yes | android update sdk --filter android-28 --no-ui --force > /dev/null
-    #  - echo yes | android update sdk --filter build-tools-28.0.3 --all --no-ui --force > /dev/null
-    #  - cd app
-    
-    #install:
-    #- npm install -g cordova
-
-    #script:
-    #- cordova platform add ios@5.1.1
-    #- cordova platform add android
-
-
-#addons:
-#  browserstack:
-#    username: "petertest7Ch2zR"
-#    access_key: 
-#      secure: "GLxxcC94qpNu3c6hUkRaKY+Mzl2XqNzT8R75xrqmjce1ooYLSWP9UufAZ47mSkxUf6cxgequW5uPxc5N5dYLYkmsrYLMIcS+J9aVrrKCLicz/dv3Zle4+wtS19MRMx/RrWiYGvXp7IZFMUaV4OJFIWDv/oljL+AaDoaiwagMOrFepgg0IKbYduQd5sQqsojHXeENNj8nHTUuT6LHZm4TPpI23TYZc7tNWzM5vUht5N2N0LPe2ji6SSp9IFiTETnT3/TEA9PMNUDs89N0WeTS94U2yt6q2yTv0Ebe9ARCVaDOMXeHuowEIV7lmj4qbcTDCT4MlbedhcJ0SbEWNmKuhitI0bf7HP+TFMMcrBJZ+ip+tb03OCAidHp+IkLVxaIRbzlNP5BfYebbmdLCBpKBINjcYdD0tNLlug+mMlns96UfYn8c3e+alnRKHrcfGEoDjJwPgQIBPZK2rJrED05+1gTK5p35mpHgGFgdqaqS0ujU8rNTeYpTILcfAx2iUcBZj5tnU/h3BpWuXPt4sMNkke32i9VZwEZlzaeKJ+WXA5j4JOJNHYidrYFTBnMHj3TTSLbKFTcGelUsZoI4Lha3fYAzNBiLbeorMyw/9vcNiA+bHsUFIVhs1nYi7joJTUx5wyKXtzc5IKnm7qclst4MH2cSIrw2Y9BhnkngcFCVbBI="
