### 2020-08-20 pmkruyen/dearscholar#20 @pmkruyen
Hi Bruno,

I updated the Wiki page (installation and testing). As you can see, I managed to do automatic testing of the php code in Travis (thanks to your Docker). Now, moving on to automatic testing of the javascript functions :)

Cheers,
Peter
