### 2020-08-20 pmkruyen/dearscholar#20 @kinow
Hi Peter,

You even included the badge already. Great job! Another useful resource is code coverage, showing how much of the code was tested. You should be able to use phpunit to get the value (may need another package installed).

But probably an enhancement for later :-)

Cheers
Bruni
