+++ Android26/www/js/dearscholar.js
@@ -261,8 +261,8 @@ var mainView = app.views.create('.view-main')
 
 ///////////////////////////////////////////////////////////////////////////////
 // General variables
-//var locationPhP = "http://localhost/dearscholar.php";
-var locationPhP = "https://peterkruyen.net/diary/dearscholar.php";
+var locationPhP = "http://localhost/dearscholar.php";
+//var locationPhP = "https://peterkruyen.net/diary/dearscholar.php";
 
 ///////////////////////////////////////////////////////////////////////////////
 // Style the login screen
@@ -679,7 +679,7 @@ $$('#acceptConsent').on('click', function (e) {
                         tx.executeSql('CREATE TABLE IF NOT EXISTS surveyInformation (surveydate)');
                         tx.executeSql('CREATE TABLE IF NOT EXISTS surveyStructure (consent, manual)');
                         tx.executeSql('CREATE TABLE IF NOT EXISTS moduleStructure (module0, name, page, mandatory, image, imageleft, imagetop)');
-                        tx.executeSql('CREATE TABLE IF NOT EXISTS pageStructure (module1, classp, idp, header, backbuttontid, backbuttontext, backbuttononclick, backbuttonhref, nextbuttonid, nextbuttontext, nextbuttononclick,nextbuttonhref)');
+                        tx.executeSql('CREATE TABLE IF NOT EXISTS pageStructure (module1, classp, idp, header, backbuttonid, backbuttontext, backbuttononclick, backbuttonhref, nextbuttonid, nextbuttontext, nextbuttononclick,nextbuttonhref)');
                         tx.executeSql('CREATE TABLE IF NOT EXISTS questiontable (module2, tab, idq, type, question, categories, footer)');
                         tx.executeSql('CREATE TABLE IF NOT EXISTS responseTable (id, surveydate, timestamp)');
                         
@@ -709,7 +709,7 @@ $$('#acceptConsent').on('click', function (e) {
                                                  pageStructure[i].data.classp,
                                                  pageStructure[i].data.idp,
                                                  pageStructure[i].data.header,
-                                                 pageStructure[i].data.backbuttontid,
+                                                 pageStructure[i].data.backbuttonid,
                                                  pageStructure[i].data.backbuttontext,
                                                  pageStructure[i].data.backbuttononclick,
                                                  pageStructure[i].data.backbuttonhref,
@@ -820,13 +820,13 @@ $$('#acceptConsent').on('click', function (e) {
     }
 
     // function to populate the pageStructure database
-    function addItemPageStructure(DiaryDatabase, module1, classp, idp, header, backbuttontid, backbuttontext, backbuttononclick , backbuttonhref, nextbuttonid, nextbuttontext, nextbuttononclick,nextbuttonhref) {
+    function addItemPageStructure(DiaryDatabase, module1, classp, idp, header, backbuttonid, backbuttontext, backbuttononclick , backbuttonhref, nextbuttonid, nextbuttontext, nextbuttononclick,nextbuttonhref) {
 
         DiaryDatabase.transaction(function (tx) {
 
-        var query = "INSERT INTO pageStructure (module1, classp, idp, header, backbuttontid, backbuttontext, backbuttononclick , backbuttonhref, nextbuttonid, nextbuttontext, nextbuttononclick,nextbuttonhref) VALUES (?,?,?,?,?,?,?,?,?,?,?,?)";
+        var query = "INSERT INTO pageStructure (module1, classp, idp, header, backbuttonid, backbuttontext, backbuttononclick , backbuttonhref, nextbuttonid, nextbuttontext, nextbuttononclick,nextbuttonhref) VALUES (?,?,?,?,?,?,?,?,?,?,?,?)";
 
[…]
