+++ Android26/www/lib/framework7/components/app/app-class.js
@@ -1,308 +1,0 @@
-import $ from 'dom7';
-import Template7 from 'template7';
-import { window, document } from 'ssr-window';
-import Utils from '../../utils/utils';
-import Device from '../../utils/device';
-import Framework7Class from '../../utils/class';
-import EventsClass from '../../utils/events-class';
-import ConstructorMethods from '../../utils/constructor-methods';
-import ModalMethods from '../../utils/modal-methods';
-import loadModule from './load-module';
-
-class Framework7 extends Framework7Class {
-  constructor(params) {
-    super(params);
-    if (Framework7.instance) {
-      throw new Error('Framework7 is already initialized and can\'t be initialized more than once');
-    }
-
-    const passedParams = Utils.extend({}, params);
-
-    // App Instance
-    const app = this;
-
-    Framework7.instance = app;
-
-    // Default
-    const defaults = {
-      version: '1.0.0',
-      id: 'io.framework7.testapp',
-      root: 'body',
-      theme: 'auto',
-      language: window.navigator.language,
-      routes: [],
-      name: 'Framework7',
-      lazyModulesPath: null,
-      initOnDeviceReady: true,
-      init: true,
-      autoDarkTheme: false,
-      iosTranslucentBars: true,
-      iosTranslucentModals: true,
-      component: undefined,
-      componentUrl: undefined,
-    };
-
-    // Extend defaults with modules params
-    app.useModulesParams(defaults);
-
-    // Extend defaults with passed params
-    app.params = Utils.extend(defaults, params);
-
-    const $rootEl = $(app.params.root);
-
-    Utils.extend(app, {
-      // App Id
-      id: app.params.id,
-      // App Name
-      name: app.params.name,
-      // App version
-      version: app.params.version,
-      // Routes
-      routes: app.params.routes,
-      // Lang
-      language: app.params.language,
-      // Root
-      root: $rootEl,
-      // RTL
-      rtl: $rootEl.css('direction') === 'rtl',
-      // Theme
-      theme: (function getTheme() {
-        if (app.params.theme === 'auto') {
-          if (Device.ios) return 'ios';
-          if (Device.desktop && Device.electron) return 'aurora';
-          return 'md';
-        }
-        return app.params.theme;
-      }()),
-      // Initially passed parameters
-      passedParams,
-      online: window.navigator.onLine,
-    });
-
-    // Save Root
-    if (app.root && app.root[0]) {
-      app.root[0].f7 = app;
-    }
-
-    // Install Modules
-    app.useModules();
-
-    // Init Data & Methods
-    app.initData();
-
-    // Auto Dark Theme
-    const DARK = '(prefers-color-scheme: dark)';
-    const LIGHT = '(prefers-color-scheme: light)';
-    app.mq = {};
-    if (window.matchMedia) {
-      app.mq.dark = window.matchMedia(DARK);
-      app.mq.light = window.matchMedia(LIGHT);
-    }
-    app.colorSchemeListener = function colorSchemeListener({ matches, media }) {
-      if (!matches) {
-        return;
-      }
-      const html = document.querySelector('html');
-      if (media === DARK) {
-        html.classList.add('theme-dark');
-      } else if (media === LIGHT) {
-        html.classList.remove('theme-dark');
-      }
-    };
-
-    // Init
-    if (app.params.init) {
-      if (Device.cordova && app.params.initOnDeviceReady) {
-        $(document).on('deviceready', () => {
-          app.init();
-        });
-      } else {
-        app.init();
-      }
-    }
-
-    // Return app instance
-    return app;
-  }
-
[…]
