+++ Android26/www/lib/framework7/components/color-picker.js
@@ -1,1 +1,0 @@
-(function framework7ComponentLoader(e,r){void 0===r&&(r=!0);document,window;var l=e.$,o=(e.Template7,e.utils),n=(e.device,e.support,e.Class),a=(e.Modal,e.ConstructorMethods),t=(e.ModalMethods,{render:function(e){var r=e.params,l=r.sliderLabel,o=r.sliderValue,n=r.sliderValueEditable,a=r.alphaLabelText;return'\n      <div class="color-picker-module color-picker-module-alpha-slider">\n        <div class="color-picker-slider-wrap">\n          '+(l?'\n            <div class="color-picker-slider-label">'+a+"</div>\n          ":"")+'\n          <div class="range-slider color-picker-slider color-picker-slider-alpha"></div>\n          '+(o?'\n            <div class="color-picker-slider-value">\n              '+(n?'\n                <input type="number" step="0.01" min="0" max="1" class="color-picker-value-alpha">\n              ':'\n                <span class="color-picker-value-alpha"></span>\n              ')+"\n            </div>\n          ":"")+"\n        </div>\n      </div>\n    "},init:function(e){function r(r){var l=e.value.alpha,o=parseFloat(r.target.value);Number.isNaN(o)?r.target.value=l:(o=Math.max(0,Math.min(1,o)),e.setValue({alpha:o}))}e.alphaRangeSlider=e.app.range.create({el:e.$el.find(".color-picker-slider-alpha"),min:0,max:1,step:.01,value:1,on:{change:function(r,l){var o=Math.floor(100*l)/100;e.setValue({alpha:o})}}}),e.$el.on("change",".color-picker-module-alpha-slider input",r),e.destroyAlphaSliderEvents=function(){e.$el.off("change",".color-picker-module-alpha-slider input",r)}},update:function(e){var r=e.value,l=e.params,o=l.sliderValue,n=l.sliderValueEditable,a=r.alpha;e.alphaRangeSlider.value=a,e.alphaRangeSlider.layout(),o&&n?e.$el.find("input.color-picker-value-alpha").val(a):e.$el.find("span.color-picker-value-alpha").text(a)},destroy:function(e){e.alphaRangeSlider&&e.alphaRangeSlider.destroy&&e.alphaRangeSlider.destroy(),delete e.alphaRangeSlider,e.destroyAlphaSliderEvents&&e.destroyAlphaSliderEvents(),delete e.destroyAlphaSliderEvents}}),i={render:function(){return'\n      <div class="color-picker-module color-picker-module-current-color">\n        <div class="color-picker-current-color"></div>\n      </div>\n    '},update:function(e){e.$el.find(".color-picker-module-current-color .color-picker-current-color").css("background-color",e.value.hex)}},s={render:function(e){var r=e.params,l=r.hexLabel,o=r.hexLabelText;return'\n      <div class="color-picker-module color-picker-module-hex">\n        <div class="color-picker-hex-wrap">\n          '+(l?'\n            <div class="color-picker-hex-label">'+o+"</div>\n          ":"")+'\n          <div class="color-picker-hex-value">\n            '+(r.hexValueEditable?'\n              <input type="text" class="color-picker-value-hex">\n            ':'\n              <span class="color-picker-value-hex"></span>\n            ')+"\n          </div>\n        </div>\n      </div>\n    "},init:function(e){function r(r){var l=e.value.hex,o=r.target.value.replace(/#/g,"");if(Number.isNaN(o)||!o||3!==o.length&&6!==o.length)r.target.value=l;else{var n=parseInt(o,16);n>parseInt("ffffff",16)&&(o="ff
[…]
