### 2020-08-31 pmkruyen/dearscholar#20 @pmkruyen
Hi Bruno,

Sorry for the delayed response, I'm back again. Updated Cordova to the latest release (there is no need to have a seperate Android26 app anymore, the only issue I have now is a a black status bar on Android, haven't figured out how to fix this yet, but anyway).

Yes to have code coverage is a good idea, will implement it after I have implemented the javascript testing.

Best,
Peter
