### 2020-08-31 pmkruyen/dearscholar#20 @kinow
Hi Peter

>Updated Cordova to the latest release (there is no need to have a seperate Android26 app anymore, the only issue I have now is a a black status bar on Android, haven't figured out how to fix this yet, but anyway).

Great! Now the building of DearScholar is even simpler!

>Yes to have code coverage is a good idea, will implement it after I have implemented the javascript testing.

:+1: I've just finished adding coverage for a Python project with `python-coverage` + `pytest` and Codacy to store the reports in another projects. Let me know if you need any help with testing, review, etc.
