### 2020-09-06 #2506 @pmkruyen
Hi @cMadan and @kinow,

FYI: the published version on the App Store + Google Play is now completely in sync with the GitHub version (`1.0.49`).

Cheers,
Peter
