### 2020-09-15 #2506 @pmkruyen
Hi @gcdeshpande,

Thank you for reviewing my paper & app. I just received a note from our IT department. My new Mac has arrived 😄, but I need to hand in my current working horse in return 😢 . Will wait till you finished your review. Can you let me know if you are able to start the review within the next 7 days or so as stated above by @cMadan?

Looking forward to your feedback.

Thanks,
Peter
