### 2020-09-28 #2506 @gcdeshpande
Hi @pmkruyen and @cMadan,
I am extremely sorry the delay in reviewing.
I just finished reading the article.
I found the work interesting. Exploring similar apps.. will provide feedback on or before Thursday.
