### 2020-10-05 #2506 @pmkruyen
Hi @gcdeshpande, thanks for starting the review. Can you give an indication when you are able to perform the remaining checks? When you need help with installing the app or experience any issues, please let me know. Best, Peter
