### 2020-10-25 #2506 @gcdeshpande
@pmkruyen  is the user interface of an application available in English? Any plans of making the app multi-lingual.
