### 2020-10-25 #2506 @pmkruyen
Hi @gcdeshpande, thanks for continuing the review.

At the moment, the only active survey is in Dutch (hence, the credentials I gave you 'direct' you to the Dutch survey). In principle, however, the app is multilingual. If you install the app on your own machine (e.g., laptop), the app is populated with an English sample survey [(see the installation instructions on the Wiki](https://github.com/pmkruyen/dearscholar/wiki/Installation) and specifically the [settings and options page](https://github.com/pmkruyen/dearscholar/wiki/Settings-and-options).

Best,
Peter
