### 2020-11-01 #2506 @gcdeshpande
On Unbuntu 20 when I run the below command

sudo docker run -ti --rm --name dearscholar -p "80:80" -p "3306:3306" dearscholar:dev

I get the error

docker: Error response from daemon: driver failed programming external connectivity on endpoint dearscholar (f317ffc10697b7d5089ef2251b212a5a7663435dd92cb6a9b9f034a05d2d8495): Error starting userland proxy: listen tcp 0.0.0.0:3306: bind: address already in use.
