+++ paper.bib
@@ -10,7 +10,7 @@
 }
 
 @article{Bolger,
-  author = {N. Bolger, A. Davis and E. Rafaeli},
+  author = {N. Bolger and A. Davis and E. Rafaeli},
   title = {Diary methods: Capturing life as it is lived},
   journal = {Annual Review of Psychology},
   volume = {54},
