+++ paper.bib
@@ -32,7 +32,7 @@
 }
 
 @article{Elliott,
-  author = {R. Elliott, A. Shankar and A. Patterson},
+  author = {R. Elliott and A. Shankar and A. Patterson},
   title = {Processes, relationships, settings, products and consumers: The case for qualitative diary research},
   journal = {Qualitative Market Research},
   volume = {8},
@@ -43,7 +43,7 @@
 }
 
 @article{Grimmelikhuijsen,
-  author = {S. Grimmelikhuijsen, L. Tummers and S. K. Pandey},
+  author = {S. Grimmelikhuijsen and L. Tummers and S. K. Pandey},
   title = {Promoting state-of-the-art methods in public management research},
   journal = {International Public Management Journal},
   volume = {20},
@@ -54,7 +54,7 @@
 }
 
 @incollection{Iida,
-  author       = {M. Iida, P. E. Shrout, J.P. Laurenceau and N. Bolger}, 
+  author       = {M. Iida and P. E. Shrout and J.P. Laurenceau and N. Bolger}, 
   title        = {Using diary methods in psychological research},
   booktitle    = {APA handbook of research methods in psychology},
   publisher    = {American Psychological Association},
@@ -99,7 +99,7 @@
 }
 
 @article{Ohly ,
-  author = {S. Ohly, S. Sonnentag, C. Niessen and D. Zapf},
+  author = {S. Ohly and S. Sonnentag and C. Niessen and D. Zapf},
   title = {Diary studies in organizational research},
   journal = {Journal of Personnel Psychology},
   volume = {9},
