+++ paper.bib
@@ -78,7 +78,7 @@
 
 @article{Jonge,
   author = {J. de Jonge},
-  title = {Editorial: A closer look at JOOP},
+  title = {Editorial: A closer look at {JOOP}},
   journal = {Journal of Occupational and Organizational Psychology},
   volume = {84},
   year = {2011},
@@ -110,7 +110,7 @@
 
 @article{Rico ,
   author = {R. Rico},
-  title = {Editorial letter: Publishing at EJWOP},
+  title = {Editorial letter: Publishing at {EJWOP}},
   journal = {European Journal of Work and Organizational Psychology},
   volume = {22},
   year = {2013},
