+++ paper.bib
@@ -22,7 +22,7 @@
 
 @article{Bono,
   author = {J. E. Bono and G. McNamara},
-  title = {From the editors: Publishing in AMJ—Part 2: Research design},
+  title = {From the editors: Publishing in {AMJ—Part 2}: Research design},
   journal = {Academy of Management Journal},
   volume = {54},
   year = {2011},
