### 2020-11-04 #2506 @pmkruyen
Hi @cMadan, thanks for the hint. I forgot this issue and to ask about it. I solved the 'and' problem; also I corrected the capital letter issue (e.g. EJOP). Is there any other issue with the bib file?
