+++ paper.bib
@@ -59,7 +59,7 @@
   booktitle    = {APA handbook of research methods in psychology},
   publisher    = {American Psychological Association},
   year         = {2012},
-  editor       = {H. Cooper, P. M. Camic, D. L. Long, A. T. Panter, D. Rindskopf and K. J. Sher},
+  editor       = {H. Cooper and P. M. Camic and D. L. Long and A. T. Panter and D. Rindskopf and K. J. Sher},
   volume       = 1,
   pages        = {277–305},
   doi = {10.1037/13619-016}
