### 2020-11-05 #2506 @cMadan
"de Jonge, J." would be correct . You can use `{` and `}` around the last name to group together the two words.
