+++ paper.bib
@@ -77,7 +77,7 @@
 }
 
 @article{Jonge,
-  author = {J. de Jonge},
+  author = {J. {De Jonge}},
   title = {Editorial: A closer look at {JOOP}},
   journal = {Journal of Occupational and Organizational Psychology},
   volume = {84},
