### 2020-11-05 #2506 @pmkruyen
@cMadan, thanks! Yes, that's the APA style. Let me know if there are any other issues. Peter
