### 2020-11-06 #2506 @cMadan
@pmkruyen, the references are good now. I did notice that (1) the capitalisation of iPhone and iPad are inconsistent, and (2) the footnotes are not formatted in the PDF. See here (https://pandoc.org/MANUAL.html) for syntax to try. Please do also read over the text carefully in case there's anything else I missed.
