+++ paper.md
@@ -18,7 +18,7 @@ bibliography: paper.bib
 # Background 
 Increasingly, scholars in the social and behavioral sciences prefer longitudinal data over cross-sectional data to explore new research phenomena, test effects, and build theories. Indeed, top-tier journals in, among others, management, psychology, and organization science nowadays tend to reject studies in which no longitudinal data has been used to investigate causal relations [e.g., @Bono; @Jonge; @Rico], in particular when testing mediation effects [cf. @Kline]. 
 
-Diary studies are one particular class of research methods in which “self-report instruments [are] used repeatedly …  to investigate social, psychological, and physiological processes, within everyday situations…” [@Bolger,578]. While there is a rich tradition of diary studies [cf. @Iida; @Ohly], these methods are used relatively infrequently to collect data compared to other methods such as experiments, panel studies, and archival research. Given the administrative burden of diary studies for both respondents (research participants) and researchers, their relative unfamiliarity (or unpopularity) is understandable.
+Diary studies are one particular class of research methods in which “self-report instruments [are] used repeatedly …  to investigate social, psychological, and physiological processes, within everyday situations…” [@Bolger,578]. While there is a rich tradition of diary studies [cf. @Iida; @Ohly], these methods are used relatively infrequently to collect data compared to other methods such as experiments, panel studies, and archival research. Given the administrative burden of diary studies for both respondents (i.e., research participants) and researchers, their relative unfamiliarity (or unpopularity) is understandable.
 
 However, given their potential to obtain a better understanding of both *between* and, especially *within* individual differences over time, scholars across disciplines call for more research in which diary methods are applied, such as in 
 in public administration [e.g.,@Bakker; @Grimmelikhuijsen], marketing [e.g., @Elliott], and health research [e.g., @Jones].
