+++ paper.md
@@ -29,7 +29,7 @@ DearScholar is a hybrid, open-source smartphone application (app) that can be us
 A limited number of alternative (commercial) diary research apps have been developed, including [Indeemo](https://www.indeemo.com), [LifeData](https://www.lifedatacorp.com/), [PIEL Survey](https://pielsurvey.org/), [Open Data Kit](https://www.opendatakit.org), [RedCap](https://www.projectredcap.org), and [Teamscope](https://www.teamscopeapp). Acknowledging their value, some of these apps target the researcher as primary respondent instead of research participants; some are rather expensive for (large-scale) projects and make it difficult to change diary tasks during the study period; some are especially designed to collect qualitative data or quantitative data for one particular type of platform only (often Android); some store data outside the European Union, which is problematic for European researchers; and—last but not least—most alternative apps are closed-source projects.
 
 # Use cases
-Currently, DearScholar is applied in a study by Glenn Houtgraaf MSc, dr. Peter M. Kruyen and prof. dr. Sandra van Thiel to investigate work-related creativity in government organizations. The app is used to follow about 100 participants over a period of six months, asking them closed- and open-ended questions at bi-weekly measurement occassions to investigate creative processes. In 2021, Liesbeth Faas MSc, dr. Peter M. Kruyen, and prof. dr. Sandra van Thiel will replicate this study in local care teams.
+Currently, DearScholar is applied in a study by Glenn Houtgraaf MSc, dr. Peter M. Kruyen and prof. dr. Sandra van Thiel to investigate work-related creativity in government organizations. The app is used to follow about 100 participants over a period of six months, asking them closed- and open-ended questions at bi-weekly measurement occasions to investigate creative processes. In 2021, Liesbeth Faas MSc, dr. Peter M. Kruyen, and prof. dr. Sandra van Thiel will replicate this study in local care teams.
 
 # Acknowledgement
 DearScholar is developed within the context of the research programme "The creative public servant: Observations, explanations and consequences" with project number 406.18.R8.028, financed by the Dutch Research Council (NWO). The author wants to express his graditute to Glenn Houtgraaf MSc, Liesbeth Faas MSc, and the ICT Services (Radboud University, Nijmegen, the Netherlands) for their advice, testing, and feedback; prof. dr. Sandra van Thiel for her encouragements; and last, but not least, all (pilot) respondents for their effort and feedback during the developmental process.
