+++ paper.md
@@ -20,11 +20,10 @@ Increasingly, scholars in the social and behavioral sciences prefer longitudinal
 
 Diary studies are one particular class of research methods in which “self-report instruments [are] used repeatedly …  to investigate social, psychological, and physiological processes, within everyday situations…” [@Bolger,578]. While there is a rich tradition of diary studies [cf. @Iida; @Ohly], these methods are used relatively infrequently to collect data compared to other methods such as experiments, panel studies, and archival research. Given the administrative burden of diary studies for both respondents (i.e., research participants) and researchers, their relative unfamiliarity (or unpopularity) is understandable.
 
-However, given their potential to obtain a better understanding of both *between* and, especially *within* individual differences over time, scholars across disciplines call for more research in which diary methods are applied, such as in 
-in public administration [e.g.,@Bakker; @Grimmelikhuijsen], marketing [e.g., @Elliott], and health research [e.g., @Jones].
+However, given their potential to obtain a better understanding of both *between* and, especially *within* individual differences over time, scholars across disciplines call for more research in which diary methods are applied, such as in public administration [e.g.,@Bakker; @Grimmelikhuijsen], marketing [e.g., @Elliott], and health research [e.g., @Jones].
 
 # DearScholar and other diary research apps
-DearScholar is a hybrid, open-source smartphone application (app) that can be used on iOS devices (iPhones and iPads) and Android devices (basically all other smartphones and tablets) to conduct diary studies (and, obviously, can also be used for other types of longitudinal research such as repeated-survey designs and log studies). DearScholar's aim is to facilitate the research process for both respondents and researchers. Researchers can specify the number of measurement occasions (measurement schedule), survey layout, and question format. Respondents only have to download the app from the App Store[^1] or Google Play[^2], fill out their assigned credentials, and start participating.
+DearScholar is a hybrid, open-source smartphone application (app) that can be used on iOS devices (iPhones and iPads) and Android devices (basically all other smartphones and tablets) to conduct diary studies and-obviously-can also be used for other types of longitudinal research such as repeated-survey designs and log studies. DearScholar's aim is to facilitate the research process for both respondents and researchers. Researchers can specify the number of measurement occasions (i.e., measurement schedule), survey layout, and question format. Respondents only have to download the app from the App Store[^1] or Google Play[^2], fill out their assigned credentials, and start participating.
 
 A limited number of alternative (commercial) diary research apps have been developed, including [Indeemo](https://www.indeemo.com), [LifeData](https://www.lifedatacorp.com/), [PIEL Survey](https://pielsurvey.org/), [Open Data Kit](https://www.opendatakit.org), [RedCap](https://www.projectredcap.org), and [Teamscope](https://www.teamscopeapp). Acknowledging their value, some of these apps target the researcher as primary respondent instead of research participants; some are rather expensive for (large-scale) projects and make it difficult to change diary tasks during the study period; some are especially designed to collect qualitative data or quantitative data for one particular type of platform only (often Android); some store data outside the European Union, which is problematic for European researchers; and—last but not least—most alternative apps are closed-source projects.
 
