+++ paper.md
@@ -16,14 +16,14 @@ bibliography: paper.bib
 ---
 
 # Background 
-Increasingly, scholars in the social and behavioral sciences prefer longitudinal data over cross-sectional data to explore new research phenomena, test effects, and build theories. Indeed, top-tier journals in, among others, management, psychology, and organization science nowadays tend to reject studies in which no longitudinal data has been used to investigate causal relations [e.g., @Bono; @Jonge; @Rico], in particular when testing mediation effects [cf. @Kline]. 
+Increasingly, scholars in the social and behavioral sciences prefer longitudinal data over cross-sectional data to explore new research phenomena, test effects, and build theories. Indeed, top-tier journals in, among others, management, psychology, and organization science nowadays tend to reject studies in which no longitudinal data has been used to investigate causal relations [e.g., @Bono; @Jonge; @Rico] in particular when testing mediation effects [cf. @Kline]. 
 
-Diary studies are one particular class of research methods in which “self-report instruments [are] used repeatedly …  to investigate social, psychological, and physiological processes, within everyday situations…” [@Bolger,578]. While there is a rich tradition of diary studies [cf. @Iida; @Ohly], these methods are used relatively infrequently to collect data compared to other methods such as experiments, panel studies, and archival research. Given the administrative burden of diary studies for both respondents (i.e., research participants) and researchers, their relative unfamiliarity (or unpopularity) is understandable.
+Diary studies are one particular class of research methods in which “self-report instruments [are] used repeatedly …  to investigate social, psychological, and physiological processes, within everyday situations…” [@Bolger,578]. While there is a rich tradition of diary studies [cf. @Iida; @Ohly], these methods are used relatively infrequently to collect data compared to other methods to collect longitudinal data such as experiments, panel studies, and archival research. Given the administrative burden of diary studies for both respondents (i.e., research participants) and researchers, their relative unfamiliarity (or unpopularity) is understandable.
 
-However, given their potential to obtain a better understanding of both *between* and, especially *within* individual differences over time, scholars across disciplines call for more research in which diary methods are applied, such as in public administration [e.g.,@Bakker; @Grimmelikhuijsen], marketing [e.g., @Elliott], and health research [e.g., @Jones].
+However, because of their potential to obtain a better understanding of both *between* and, especially *within* individual differences over time, scholars across disciplines call for more research in which diary methods are applied, such as in public administration [e.g.,@Bakker; @Grimmelikhuijsen], marketing [e.g., @Elliott], and health research [e.g., @Jones].
 
 # DearScholar and other diary research apps
-DearScholar is a hybrid, open-source smartphone application (app) that can be used on iOS devices (iPhones and iPads) and Android devices (basically all other smartphones and tablets) to conduct diary studies and-obviously-for other types of longitudinal research such as repeated-survey designs and log studies. DearScholar's aim is to facilitate the research process for both respondents and researchers. Researchers can specify the number of measurement occasions (i.e., measurement schedule), survey layout, and question format. Respondents only have to download the app from the App Store[^1] or Google Play[^2], fill out their assigned credentials, and start participating.
+DearScholar is a hybrid, open-source smartphone application (app) that can be used on iOS devices (iPhones and iPads) and Android devices (basically all other smartphones and tablets) to conduct diary studies and, obviously, for other types of longitudinal research such as repeated-survey designs and log studies. DearScholar's aim is to facilitate the research process for both respondents and researchers. Researchers can specify the number of measurement occasions (i.e., measurement schedule), survey layout, and question format. Respondents only have to download the app from the App Store[^1] or Google Play[^2], fill out their assigned credentials, and start participating.
 
[…]
