### 2020-11-06 #2506 @pmkruyen
Hi @cMadan, thanks for the manual (it was unclear for me that I could do this, at the initial submission I had included a work around for the footnotes. I corrected this now. Also I corrected three typos and rephrased several sentences. Hope the text is ok now. Best, Peter
