### 2020-11-11 #2506 @cMadan
@pmkruyen, everything looks good to me now!

To move forward with accepting your submission, there are a few last things to take care of:

You may find this helpful: https://guides.github.com/activities/citable-code/
