+++ paper.bib
@@ -65,6 +65,12 @@
   doi = {10.1037/13619-016}
 }
 
+@misc{Indeemo,
+  author = {Indeemo},
+  year = {2020},
+  url = {https://www.indeemo.com}
+}
+
 @article{Jones,
   author = {R. K. Jones},
   title = {The unsolicited diary as a qualitative research tool for advancing research capacity in the field of health and illness},
@@ -98,6 +104,12 @@
   doi = {10.1080/01973533.2015.1049349}
 }
 
+@misc{LifeData,
+  author = {LifeData},
+  year = {2020},
+  url = {https://www.lifedatacorp.com}
+}
+
 @article{Ohly ,
   author = {S. Ohly and S. Sonnentag and C. Niessen and D. Zapf},
   title = {Diary studies in organizational research},
@@ -106,6 +118,24 @@
   year = {2010},
   pages = {79-93},
   doi = {10.1027/1866-5888/a000009}
+}
+
+@misc{Open Data Kit,
+  author = {Open Data Kit},
+  year = {2020},
+  url = {https://www.opendatakit.org}
+}
+
+@misc{PIEL Survey,
+  author = {PIEL Survey},
+  year = {2020},
+  url = {https://pielsurvey.org}
+}
+
+@misc{RedCap,
+  author = {RedCap},
+  year = {2020},
+  url = {https://www.projectredcap.org}
 }
 
 @article{Rico ,
@@ -117,4 +147,10 @@
   number = {1},
   pages = {1-3},
   doi = {10.1080/1359432X.2013.752247}
+}
+
+@misc{Teamscope,
+  author = {Teamscope},
+  year = {2020},
+  url = {https://www.teamscopeapp}
 }
