+++ paper.bib
@@ -110,7 +110,7 @@
   url = {https://www.lifedatacorp.com}
 }
 
-@article{Ohly ,
+@article{Ohly,
   author = {S. Ohly and S. Sonnentag and C. Niessen and D. Zapf},
   title = {Diary studies in organizational research},
   journal = {Journal of Personnel Psychology},
@@ -120,13 +120,13 @@
   doi = {10.1027/1866-5888/a000009}
 }
 
-@misc{Open Data Kit,
+@misc{Open,
   author = {Open Data Kit},
   year = {2020},
   url = {https://www.opendatakit.org}
 }
 
-@misc{PIEL Survey,
+@misc{PIEL,
   author = {PIEL Survey},
   year = {2020},
   url = {https://pielsurvey.org}
@@ -138,7 +138,7 @@
   url = {https://www.projectredcap.org}
 }
 
-@article{Rico ,
+@article{Rico,
   author = {R. Rico},
   title = {Editorial letter: Publishing at {EJWOP}},
   journal = {European Journal of Work and Organizational Psychology},
