+++ paper.bib
@@ -121,13 +121,13 @@
 }
 
 @misc{Open,
-  author = {Open Data Kit},
+  author = {{Open Data Kit}},
   year = {2020},
   url = {https://www.opendatakit.org}
 }
 
 @misc{PIEL,
-  author = {PIEL Survey},
+  author = {{PIEL Survey}},
   year = {2020},
   url = {https://pielsurvey.org}
 }
