+++ paper.bib
@@ -152,5 +152,5 @@
 @misc{Teamscope,
   author = {Teamscope},
   year = {2020},
-  url = {https://www.teamscopeapp}
+  url = {https://www.teamscopeapp.com}
 }
