+++ paper.bib
@@ -132,8 +132,9 @@
   url = {https://pielsurvey.org}
 }
 
-@misc{RedCap,
-  author = {RedCap},
+@misc{REDCap,
+  author = {the REDCap consortium},
+  title = {REDCap},
   year = {2020},
   url = {https://www.projectredcap.org}
 }
