+++ paper.bib
@@ -132,9 +132,8 @@
   url = {https://pielsurvey.org}
 }
 
-@misc{REDCap,
-  author = {the REDCap consortium},
-  title = {REDCap},
+@misc{RedCap,
+  author = {REDCap},
   year = {2020},
   url = {https://www.projectredcap.org}
 }
