### 2020-11-11 #2506 @pmkruyen
Hi @cMadan,

Thanks for the checks.

- I made a tagged release: v1.0.49 https://github.com/pmkruyen/dearscholar/releases/tag/v1.0.49
- I archieved the reviewed software in Zenodo
- I checked the metadata (added @kinow under 'other' contributors; hope this is okay with him:  https://zenodo.org/record/4267938#.X6vacJNKhb8
- the Zenodo DOI is: 10.5281/zenodo.426793

And then, I decided to change the formatting of the URL addresses in the bib file, because I became unsure about the proper formating here. Better to have asked this first to you :( Could you have a last look at the way I styled/formatted the urls in the paper please?

Also, because of these small changes, do I need to create a new tagged release?  I could not re-create the current release (this is a beginner's fault).

Thanks,
Peter
