### 2020-11-13 #2506 @cMadan
@pmkruyen, this url formatting looks good. The previous was fine too, but this does look better.
