### 2020-11-13 #2506 @kthyng
@cMadan Why is the performance checkbox unchecked for both reviewers?
