### 2020-11-13 #2506 @kthyng
Looks like the zenodo doi wasn't copied correctly before, so I'll fix it
