### 2020-11-13 #2506 @kthyng
@pmkruyen Is it convention to have "prof dr" and "dr" not capitalized?
