### 2020-11-13 #2506 @kthyng
Note to myself that the version is up to date, Zenodo archive looks good, and paper is good other than the question above. Waiting to hear back about the capitalization and the unchecked boxes.
