### 2020-11-13 #2506 @kinow
> > @cMadan Why is the performance checkbox unchecked for both reviewers?
>
> Ops, sorry @kthyng ! Checked it now. There are no claims about performance in the paper. Also, I tested it locally (and also briefly against the server with automated tests) and found no performance issues +1
