### 2020-11-13 #2506 @pmkruyen
Hi @kthyng,

> Is it convention to have "prof dr" and "dr" not capitalized?

Good point. In the Netherlands, this is the [convention](https://taaladvies.net/taal/advies/vraag/218/dr_ir_mr_dr_ir_mr/#:~:text=We%20spreken%20de%20afkortingen%20uit,in%20het%20afgekorte%20woord%20voorkomt.) I now see that in the English language, it is always Prof. Dr. I changed the text accordingly.

> the unchecked boxes
As @kinow noted I have no specific performance claims.

Best,
Peter
