### 2020-11-13 #2506 @kthyng
> Good point. In the Netherlands, this is the [convention](https://taaladvies.net/taal/advies/vraag/218/dr_ir_mr_dr_ir_mr/#:~:text=We%20spreken%20de%20afkortingen%20uit,in%20het%20afgekorte%20woord%20voorkomt.) I now see that in the English language, it is always Prof. Dr. I changed the text accordingly.

At least in the U.S. we don't use "Dr Prof", just "Dr". But, you can specify however is appropriate for you, I think.

Thanks @kinow!

@gcdeshpande Can you verify that you are satisfied with this JOSS submission and check off the performance box in your review?
