### 2020-11-15 #2506 @cMadan
@kthyng, I figured it's better to not 'enforce' US norms on titles, so had left that as-is. As for the performance checkbox, I was fine with it unchecked as no performance claims were made, though functionality was obviously assessed. @gcdeshpande signed off in comment https://github.com/openjournals/joss-reviews/issues/2506#issuecomment-720124281. (Part of why I'm responding like this is that @gcdeshpande doesn't always check Github regularly and I often had to follow up via email to prompt responses here.)
