### 2020-11-16 #2506 @danielskatz
👋 @gcdeshpande - If there are no performance claims, the box should then be checked, as it says "If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)"
