### 2020-11-17 #2506 @kthyng
> @kthyng, I figured it's better to not 'enforce' US norms on titles, so had left that as-is. As for the performance checkbox, I was fine with it unchecked as no performance claims were made, though functionality was obviously assessed. @gcdeshpande signed off in comment [#2506 (comment)](https://github.com/openjournals/joss-reviews/issues/2506#issuecomment-720124281). (Part of why I'm responding like this is that @gcdeshpande doesn't always check Github regularly and I often had to follow up via email to prompt responses here.)

I agree either way is fine now that I know that was following some convention anyway. I'll assume the author doesn't care too much either way and finish this up now!
