### 2020-11-17 #2506 @pmkruyen
Thanks @cMadan, @kinow, @gcdeshpande, and @kthyng for your time, help and feedback. This review process was a super transparent and nice experience, I learned a lot too (though I missed some nights of sleep solving @kinow's issues 😂👍) . Hope now that other researchers will start using DearScholar in their own projects 😁. All the best! Peter
